### 2020-08-24 #2592 @meg-simula
@santisoler @nschloe would you be available to review this submission for JOSS? Please take a look at our reviewer guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html, and let us know. Your help would be much appreciated.
