### 2020-08-26 f9598b8
Changing license of SeismicMesh to GPLv3
