### 2020-08-26 dd046aa
adding note regarding segyio installation and scikit-build dep
