+++ setup.py
@@ -81,14 +81,19 @@ class CMakeBuild(build_ext):
         )
 
 
+with open("README.rst", "r") as fh:
+    long_description = fh.read()
+
 setup(
     name="SeismicMesh",
-    version="1.0.0",
+    version="1.2.3",
     author="Keith Roberts",
-    author_email="keithrbt0gmail.com",
+    author_email="keithrbt0@gmail.com",
     description="2D/3D serial and parallel triangular mesh generation and mesh improvement for seismology",
-    long_description="",
+    long_description=long_description,
+    long_description_content_type="text/x-rst",
     setup_requires=["pybind11"],
+    license_files="LICENSE",
     install_requires=[
         "numpy",
         "segyio",
@@ -113,4 +118,14 @@ setup(
     ],
     cmdclass=dict(build_ext=CMakeBuild),
     zip_safe=False,
+    classifiers=[
+        "Development Status :: 4 - Beta",
+        "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
+        "Operating System :: OS Independent",
+        "Programming Language :: Python :: 3",
+        "Topic :: Scientific/Engineering",
+        "Topic :: Scientific/Engineering :: Mathematics",
+        "Topic :: Scientific/Engineering :: Physics",
+        "Topic :: Scientific/Engineering :: Visualization"],
+    python_requires='>=3.0',
 )
