+++ setup.py
@@ -86,7 +86,8 @@ with open("README.rst", "r") as fh:
 
 setup(
     name="SeismicMesh",
-    version="1.2.3",
+    url="https://github.com/krober10nd/SeismicMesh",
+    version="1.2.4",
     author="Keith Roberts",
     author_email="keithrbt0@gmail.com",
     description="2D/3D serial and parallel triangular mesh generation and mesh improvement for seismology",
