### 2020-08-26 0b3e42a
adding MANFEST.in so source dist builds correctly
