### 2020-08-26 9f8efd1
adding pybind dependency on README.rst
