### 2020-08-27 08c742f
Corrections to install.rst
