### 2020-08-27 8d61ad9
updating to codestyle consistent with blackv20b
