+++ tests/test_3dmesher_par.py
@@ -43,9 +43,18 @@ def test_3dpar_mesher():
 
     mshgen = SeismicMesh.MeshGenerator(ef)
 
-    points, cells = mshgen.build(max_iter=50, nscreen=1, seed=0, axis=0,)
+    points, cells = mshgen.build(
+        max_iter=50,
+        nscreen=1,
+        seed=0,
+        axis=0,
+    )
 
-    points, cells = mshgen.build(points=points, max_iter=15, mesh_improvement=True,)
+    points, cells = mshgen.build(
+        points=points,
+        max_iter=15,
+        mesh_improvement=True,
+    )
 
     if rank == 0:
         # import meshio
