### 2020-08-27 6f2ab3f
applying black to setup.py
