### 2020-08-27 ba52a7d
Update setup.py
Include LICENSE.txt and license comment in setup.py per #19
