### 2020-08-27 df24ec3
adding pypi badges to README.rst
