### 2020-08-27 3a8e01f
Using syntax highlighting in README.rst
