### 2020-08-27 2fa7b96
syntax fix in README.rst
