### 2020-08-27 839b703
Cleaning up README.rst example
