### 2020-08-27 b304030
switching to setup.cfg
