### 2020-08-28 78bf6f7
adding 2D example on README.rst
