+++ README.rst
@@ -1,9 +1,9 @@
 .. image:: https://circleci.com/gh/krober10nd/SeismicMesh/tree/par3d.svg?style=shield
-        :target: https://circleci.com/gh/krober10nd/SeismicMesh/tree/par3d 
+        :target: https://circleci.com/gh/krober10nd/SeismicMesh/tree/par3d
 
 .. image:: https://codecov.io/gh/krober10nd/SeismicMesh/branch/par3d/graph/badge.svg
   	:target: https://codecov.io/gh/krober10nd/SeismicMesh
-    
+
 .. image:: https://img.shields.io/badge/code%20style-black-000000.svg
         :target: https://github.com/ambv/black
 
@@ -12,17 +12,17 @@
 
 .. image:: https://readthedocs.org/projects/seismicmesh/badge/?version=par3d
         :target: https://seismicmesh.readthedocs.io/en/par3d/?badge=par3d
-	
+
 .. image:: https://img.shields.io/badge/License-GPLv3-blue.svg
 	:target: https://www.gnu.org/licenses/gpl-3.0
-	
+
 .. image:: https://zenodo.org/badge/216707188.svg
    :target: https://zenodo.org/badge/latestdoi/216707188
 
-.. image:: https://img.shields.io/pypi/pyversions/SeismicMesh.svg?style=flat-square 
+.. image:: https://img.shields.io/pypi/pyversions/SeismicMesh.svg?style=flat-square
    :target: https://pypi.org/pypi/SeismicMesh
 
-.. image:: https://img.shields.io/pypi/v/SeismicMesh.svg?style=flat-square 
+.. image:: https://img.shields.io/pypi/v/SeismicMesh.svg?style=flat-square
    :target: https://pypi.org/project/SeismicMesh
 
 .. image:: https://img.shields.io/pypi/dm/SeismicMesh.svg?style=flat-square
@@ -42,7 +42,7 @@ SeismicMesh_: Mesh generation for Seismology in Python
 Installation
 =====================
 
-For installation, SeismicMesh needs `CGAL <https://www.cgal.org/>`_ and `pybind11 <https://github.com/pybind/pybind11>`_:: 
+For installation, SeismicMesh needs `CGAL <https://www.cgal.org/>`_ and `pybind11 <https://github.com/pybind/pybind11>`_::
 
     sudo apt install libcgal-dev python-pybind11
 
@@ -51,18 +51,81 @@ Index (`pypi <https://pypi.org/project/SeismicMesh/>`_), so with::
 
     pip install -U SeismicMesh
 
-For more detailed information about installation and requirements see: 
+For more detailed information about installation and requirements see:
 
 `Install <https://seismicmesh.readthedocs.io/en/par3d/install.html>`_
-- How to install SeismicMesh. 
+- How to install SeismicMesh.
 
 
-Example 
+Example
 ===========
 
-The user can quickly build quality 2D/3D meshes from seismic velocity models in serial/parallel. 
+The user can quickly build quality 2D/3D meshes from seismic velocity models in serial/parallel.
 
-**WARNING: To run the code snippet below you must download the 3D EAGE seismic velocity model from (WARNING: File is > 500 MB)** `here <https://s3.amazonaws.com/open.source.geoscience/open_data/seg_eage_models_cd/Salt_Model_3D.tar.gz>`_ 
+
+**WARNING: To run the code snippet below you must download the 2D BP2004 seismic velocity model from and then you must uncompress it** `here <>`_
+.. figure::
+    :width: 30 %
+
+ Above shows the mesh from running the code below. Note, the seismic velocity data has been interpolated onto the vertices of the mesh.
+
+.. code-block:: python
+
[…]
