+++ README.rst
@@ -63,7 +63,7 @@ Example
 The user can quickly build quality 2D/3D meshes from seismic velocity models in serial/parallel.
 
 
-**WARNING: To run the code snippet below you must download the 2D BP2004 seismic velocity model from and then you must uncompress it (e.g., gunzip)** `here <http://s3.amazonaws.com/open.source.geoscience/open_data/bpvelanal2004/vel_z6.25m_x12.5m_exact.segy.gz>`_
+**WARNING: To run the code snippet below you must download the 2D BP2004 seismic velocity model and then you must uncompress it (e.g., gunzip). This file can be downloaded from** `here <http://s3.amazonaws.com/open.source.geoscience/open_data/bpvelanal2004/vel_z6.25m_x12.5m_exact.segy.gz>`_
 
 .. figure:: https://user-images.githubusercontent.com/18619644/91577721-82722c80-e91f-11ea-82f2-519687722e7b.jpg
    :width: 30 %
@@ -124,7 +124,9 @@ The user can quickly build quality 2D/3D meshes from seismic velocity models in 
             file_format="vtk",
         )
 
-**WARNING: To run the code snippet below you must download the 3D EAGE seismic velocity model from (WARNING: File is ~500 MB and ~ 8 GB RAM)** `here <https://s3.amazonaws.com/open.source.geoscience/open_data/seg_eage_models_cd/Salt_Model_3D.tar.gz>`_
+**WARNING: To run the code snippet below you must download the 3D EAGE seismic velocity model from (WARNING: File is ~500 MB)** `here <https://s3.amazonaws.com/open.source.geoscience/open_data/seg_eage_models_cd/Salt_Model_3D.tar.gz>`_
+
+**WARNING: Computationaly demanding!! Running this example requires around 8 GB of RAM due to the 3D nature of the problem and the domain size.
 
 
 .. figure:: https://user-images.githubusercontent.com/18619644/91485472-4be5d480-e881-11ea-9abf-75ae2fb6b2b1.jpg
