+++ README.md
@@ -1,0 +1,222 @@
+![image](https://circleci.com/gh/krober10nd/SeismicMesh/tree/par3d.svg?style=shield%0A%20%20%20%20%20:target:%20https://circleci.com/gh/krober10nd/SeismicMesh/tree/par3d)
+
+![image](https://codecov.io/gh/krober10nd/SeismicMesh/branch/par3d/graph/badge.svg%0A%20:target:%20https://codecov.io/gh/krober10nd/SeismicMesh)
+
+![image](https://img.shields.io/badge/code%20style-black-000000.svg%0A%20%20%20%20%20:target:%20https://github.com/ambv/black)
+
+![image](http://www.repostatus.org/badges/latest/active.svg%0A%20:target:%20http://www.repostatus.org/#active)
+
+![image](https://readthedocs.org/projects/seismicmesh/badge/?version=par3d%0A%20%20%20%20%20:target:%20https://seismicmesh.readthedocs.io/en/par3d/?badge=par3d)
+
+![image](https://img.shields.io/badge/License-GPLv3-blue.svg%0A%20:target:%20https://www.gnu.org/licenses/gpl-3.0)
+
+[![image](https://zenodo.org/badge/216707188.svg)](https://zenodo.org/badge/latestdoi/216707188)
+
+[![image](https://img.shields.io/pypi/pyversions/SeismicMesh.svg?style=flat-square)](https://pypi.org/pypi/SeismicMesh)
+
+[![image](https://img.shields.io/pypi/v/SeismicMesh.svg?style=flat-square)](https://pypi.org/project/SeismicMesh)
+
+[![image](https://img.shields.io/pypi/dm/SeismicMesh.svg?style=flat-square)](https://pypistats.org/packages/seismicmesh)
+
+[SeismicMesh](https://github.com/krober10nd/SeismicMesh): Mesh generation for Seismology in Python
+==================================================================================================
+
+2D/3D triangular meshing for a slab of Earth based on modifications to
+the [DistMesh](http://persson.berkeley.edu/distmesh/) algorithm.
+SeismicMesh is distributed under GPL3.
+
+Installation
+============
+
+For installation, SeismicMesh needs [CGAL](https://www.cgal.org/) and
+[pybind11](https://github.com/pybind/pybind11):
+
+    sudo apt install libcgal-dev python-pybind11
+
+After that, SeismicMesh can be installed from the Python Package Index
+([pypi](https://pypi.org/project/SeismicMesh/)), so with:
+
+    pip install -U SeismicMesh
+
+For more detailed information about installation and requirements see:
+
+[Install](https://seismicmesh.readthedocs.io/en/par3d/install.html) -
+How to install SeismicMesh.
+
+Example
+=======
+
+The user can quickly build quality 2D/3D meshes from seismic velocity
+models in serial/parallel.
+
+**WARNING: To run the code snippet below you must download the 2D BP2004
+seismic velocity model and then you must uncompress it (e.g., gunzip).
+This file can be downloaded from**
+[here](http://s3.amazonaws.com/open.source.geoscience/open_data/bpvelanal2004/vel_z6.25m_x12.5m_exact.segy.gz)
+
+![Above shows the mesh from running the code below. Note, the seismic
+velocity data has been interpolated onto the vertices of the mesh using
+a different program with linear
+interpolation.](https://user-images.githubusercontent.com/18619644/91577721-82722c80-e91f-11ea-82f2-519687722e7b.jpg)
+
+``` {.sourceCode .python}
+from mpi4py import MPI
+import meshio
+
+import SeismicMesh
+
[…]
