### 2020-08-28 c508a4f
updating notes on examples
