+++ tests/test_README.py
@@ -9,6 +9,7 @@ import pytest
 this_dir = pathlib.Path(__file__).resolve().parent
 
 
+@pytest.mark.serial
 @pytest.mark.parametrize(
     "string,lineno",
     exdown.extract(this_dir.parent / "README.md", syntax_filter="python"),
