### 2020-08-28 ceda140
adding coverage badge back
