+++ tests/test_2dmesher_par.py
@@ -1,9 +1,8 @@
-import pytest
-
 import os
 
-from mpi4py import MPI
 import numpy as np
+import pytest
+from mpi4py import MPI
 
 import SeismicMesh
 
