+++ tests/test_ptin.py
@@ -1,5 +1,6 @@
-import pytest
 import numpy as np
+import pytest
+
 import SeismicMesh
 
 
