+++ SeismicMesh/generation/cpp/delaunay.cpp
@@ -1,216 +1,199 @@
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>
+#include <boost/lexical_cast.hpp>
 #include <pybind11/complex.h>
 #include <pybind11/numpy.h>
-#include <boost/lexical_cast.hpp>
+#include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
 
 #include <assert.h>
 #include <vector>
 
 #include <CGAL/Kernel/global_functions.h>
 
-#include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
-#include <CGAL/Delaunay_triangulation_3.h>
-#include <CGAL/Triangulation_vertex_base_with_info_3.h>
 #include <CGAL/Delaunay_triangulation_2.h>
+#include <CGAL/Delaunay_triangulation_3.h>
+#include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 #include <CGAL/Triangulation_vertex_base_with_info_2.h>
+#include <CGAL/Triangulation_vertex_base_with_info_3.h>
 
-
-
-typedef CGAL::Exact_predicates_inexact_constructions_kernel            Kernel;
+typedef CGAL::Exact_predicates_inexact_constructions_kernel Kernel;
 typedef CGAL::Triangulation_vertex_base_with_info_2<unsigned int, Kernel> Vb;
-typedef CGAL::Triangulation_data_structure_2<Vb>                       Tds;
-typedef CGAL::Delaunay_triangulation_2<Kernel, Tds>                    Delaunay;
+typedef CGAL::Triangulation_data_structure_2<Vb> Tds;
+typedef CGAL::Delaunay_triangulation_2<Kernel, Tds> Delaunay;
 typedef Kernel::Point_2 Point;
 
 typedef CGAL::Triangulation_vertex_base_with_info_3<unsigned int, Kernel> Vb3;
-typedef CGAL::Triangulation_data_structure_3<Vb3>                       Tds3;
-typedef CGAL::Delaunay_triangulation_3<Kernel, Tds3>                    Delaunay3;
+typedef CGAL::Triangulation_data_structure_3<Vb3> Tds3;
+typedef CGAL::Delaunay_triangulation_3<Kernel, Tds3> Delaunay3;
 typedef Kernel::Point_3 Point3;
 
-
-
-std::vector<double> c_circumballs2(std::vector<double> &vertices)
-{
-    int num_faces = vertices.size()/6;
-    std::vector<double> circumcenters;
-    for(std::size_t i=0; i < num_faces; ++i)
-    {
-        Point tmp_cc =
-                CGAL::circumcenter(
-                Point(vertices[i*6],vertices[i*6+1]),
-                Point(vertices[i*6+2],vertices[i*6+3]),
-                Point(vertices[i*6+4],vertices[i*6+5])
-                );
-        circumcenters.push_back(tmp_cc.x());
-        circumcenters.push_back(tmp_cc.y());
-        circumcenters.push_back(
-                CGAL::squared_radius(
-                    Point(vertices[i*6],vertices[i*6+1]),
-                    Point(vertices[i*6+2],vertices[i*6+3]),
-                    Point(vertices[i*6+4],vertices[i*6+5])
-                    )
-                );
-    }
-    return circumcenters;
+std::vector<double> c_circumballs2(std::vector<double> &vertices) {
+  int num_faces = vertices.size() / 6;
+  std::vector<double> circumcenters;
+  for (std::size_t i = 0; i < num_faces; ++i) {
+    Point tmp_cc =
+        CGAL::circumcenter(Point(vertices[i * 6], vertices[i * 6 + 1]),
[…]
