+++ SeismicMesh/generation/cpp/delaunay_class3.cpp
@@ -1,19 +1,19 @@
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>
 #include <pybind11/complex.h>
 #include <pybind11/numpy.h>
+#include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
 
-#include <vector>
-#include <string>
 #include <cstdarg>
-#include <cstring>
 #include <cstddef>
+#include <cstring>
 #include <iterator>
+#include <string>
+#include <vector>
 
 #include <boost/lexical_cast.hpp>
 
-#include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 #include <CGAL/Delaunay_triangulation_3.h>
+#include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 #include <CGAL/Triangulation_vertex_base_with_info_3.h>
 
 namespace py = pybind11;
@@ -27,252 +27,219 @@ using Point = K::Point_3;
 using Vertex_handle = DT::Vertex_handle;
 using Vi = DT::Finite_vertices_iterator;
 
-
-template <typename T>
-class TypedInputIterator
-{
+template <typename T> class TypedInputIterator {
 public:
-    using iterator_category = std::input_iterator_tag;
-    using difference_type = std::ptrdiff_t;
-    using value_type = T;
-    using pointer = T*;
-    using reference = T&;
+  using iterator_category = std::input_iterator_tag;
+  using difference_type = std::ptrdiff_t;
+  using value_type = T;
+  using pointer = T *;
+  using reference = T &;
 
-    explicit TypedInputIterator(py::iterator& py_iter) :
-        py_iter_(py_iter)
-    {
-    }
+  explicit TypedInputIterator(py::iterator &py_iter) : py_iter_(py_iter) {}
 
-    explicit TypedInputIterator(py::iterator&& py_iter) :
-        py_iter_(py_iter)
-    {
-    }
+  explicit TypedInputIterator(py::iterator &&py_iter) : py_iter_(py_iter) {}
 
-    value_type operator*()
-    {
-        return (*py_iter_).template cast<value_type>();
-    }
+  value_type operator*() { return (*py_iter_).template cast<value_type>(); }
 
-    TypedInputIterator operator++(int)
-    {
-        auto copy = *this;
-        ++py_iter_;
-        return copy;
-    }
+  TypedInputIterator operator++(int) {
+    auto copy = *this;
+    ++py_iter_;
+    return copy;
+  }
 
-    TypedInputIterator& operator++()
-    {
-        ++py_iter_;
-        return *this;
-    }
+  TypedInputIterator &operator++() {
+    ++py_iter_;
+    return *this;
+  }
 
-    bool operator!=(TypedInputIterator &rhs)
-    {
-        return py_iter_ != rhs.py_iter_;
-    }
+  bool operator!=(TypedInputIterator &rhs) { return py_iter_ != rhs.py_iter_; }
 
-    bool operator==(TypedInputIterator &rhs)
-    {
-        return py_iter_ == rhs.py_iter_;
-    }
+  bool operator==(TypedInputIterator &rhs) { return py_iter_ == rhs.py_iter_; }
 
 private:
-    py::iterator py_iter_;
+  py::iterator py_iter_;
 };
 
+PYBIND11_MODULE(delaunay_class3, m) {
+  py::class_<Point>(m, "Point")
+      .def(py::init<int, int, int>(), py::arg("x"), py::arg("y"), py::arg("z"))
+      .def(py::init<double, double, double>(), py::arg("x"), py::arg("y"),
+           py::arg("z"))
+      .def_property_readonly("x", &Point::x)
+      .def_property_readonly("y", &Point::y)
[…]
