### 2020-08-28 3becfe9
Code formatting (#28)
* Adding Makefile for macros

* adding clang-format

* Update Makefile
