+++ tests/test_geometry.py
@@ -1,5 +1,6 @@
-import pytest
 import numpy as np
+import pytest
+
 from SeismicMesh import geometry as geo
 
 
