+++ tests/test_3dmesher_par.py
@@ -1,7 +1,7 @@
-import pytest
 import os
 
 import numpy as np
+import pytest
 from mpi4py import MPI
 
 import SeismicMesh
