+++ SeismicMesh/generation/__init__.py
@@ -1,15 +1,7 @@
-from .mesh_generator import MeshGenerator
-
+from .cpp.c_cgal import circumballs2, circumballs3, delaunay2, delaunay3
 from .cpp.delaunay_class import DelaunayTriangulation
 from .cpp.delaunay_class3 import DelaunayTriangulation3
-
-from .cpp.c_cgal import (
-    delaunay2,
-    delaunay3,
-    circumballs2,
-    circumballs3,
-)
-
+from .mesh_generator import MeshGenerator
 
 __all__ = [
     "DelaunayTriangulation",
