+++ tests/test_2d_domain_ext.py
@@ -1,6 +1,7 @@
-import pytest
 import os
+
 import numpy as np
+import pytest
 
 import SeismicMesh
 
