+++ tests/test_2d_SDF.py
@@ -1,6 +1,5 @@
-import pytest
-
 import numpy as np
+import pytest
 
 import SeismicMesh
 
