+++ examples/example_3D_SDF2.py
@@ -1,5 +1,5 @@
-import numpy as np
 import meshio
+import numpy as np
 from mpi4py import MPI
 
 import SeismicMesh
