### 2020-08-29 da4dbac
Addressing memory and speed issue with grading
