### 2020-08-29 05e2851
applying clang
