### 2020-08-29 78e9487
Reducing memory with mesh grading (#29)
* Adding Makefile for macros

* adding clang-format

* Addressing memory and speed issue with grading

* applying clang

* Update README.md

* Fixing typo
