+++ Makefile
@@ -4,12 +4,14 @@ default:
 	@echo "\"make publish\"?"
 
 tag:
+
 	@if [ "$(shell git rev-parse --abbrev-ref HEAD)" != "par3d" ]; then exit 1; fi
 	curl -H "Authorization: token `cat $(HOME)/.github-access-token`" -d '{"tag_name": "v$(VERSION)"}' https://api.github.com/repos/krober10nd/SeismicMesh/releases
 
 upload:
 	# Make sure we're on the par3d branch
 	@if [ "$(shell git rev-parse --abbrev-ref HEAD)" != "par3d" ]; then exit 1; fi
+
 	rm -rf dist/*
 	python3 setup.py sdist bdist_wheel
 	twine upload dist/*
