### 2020-08-29 86d08e9
Update FastHJ.cpp (#30)
