+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = SeismicMesh
-version = 1.2.43
+version = 2.0.0
 url = https://github.com/krober10nd/SeismicMesh
 author = Keith Roberts
 email = keithrbt0@gmail.com
@@ -36,5 +36,5 @@ install_requires =
         mpi4py
         scikit-fmm
         pybind11 >= 2.2
-        exdown 
+        exdown
 python_requires = >=3.0
