### 2020-08-29 eba6542
setup.cfg points to markdown README instead of rst
