+++ setup.cfg
@@ -1,12 +1,12 @@
 [metadata]
 name = SeismicMesh
-version = 2.0.0
+version = 2.0.1
 url = https://github.com/krober10nd/SeismicMesh
 author = Keith Roberts
 email = keithrbt0@gmail.com
 description = 2D/3D serial and parallel triangular mesh generation and mesh improvement for seismology
-long_description = file: README.rst
-long_description_content_type = text/x-rst
+long_description = file: README.md
+long_description_content_type = text/markdown
 license = GPL-3.0-or-later
 license_files = LICENSE.txt
 classifiers =
