+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = SeismicMesh
-version = 2.0.1
+version = 2.0.2
 url = https://github.com/krober10nd/SeismicMesh
 author = Keith Roberts
 email = keithrbt0@gmail.com
@@ -31,10 +31,11 @@ install_requires =
         scipy
         meshio
         h5py
-        pybind11
         matplotlib
         mpi4py
         scikit-fmm
         pybind11 >= 2.2
+test_requires =
         exdown
+        pytest
 python_requires = >=3.0
