### 2020-08-30 fab1019
adding test deps to config
