### 2020-08-30 78109b5
removing pybind11, using tox to do testing
