+++ setup.cfg
@@ -35,7 +35,4 @@ install_requires =
         mpi4py
         scikit-fmm
         pybind11 >= 2.2
-test_requires =
-        exdown
-        pytest
 python_requires = >=3.0
