+++ tox.ini
@@ -1,0 +1,11 @@
+[tox]
+envlist = py3
+isolated_build = True
+
+[testenv]
+deps =
+    exdown
+    pytest
+    pytest-cov
+commands =
+    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
