+++ .circleci/config.yml
@@ -15,8 +15,14 @@ jobs:
       - image: circleci/python:3
     steps:
       - checkout
-      - run: sudo apt install -y libcgal-dev python-pybind11 cmake mpich libhdf5-dev
-      - run: pip install .[test] --user
+      - run: sudo apt install -y cmake
+      - run: sudo apt-get install libboost-all-dev
+      - run: sudo apt-get install libmpfr-dev
+      - run: git clone https://github.com/CGAL/cgal.git && cd cgal/ && mkdir build && cd build && cmake -DCMAKE_BUILD_TYPE=Release .. && sudo make install
+      - run: sudo apt install -y libhdf5-dev
+      - run: sudo apt install -y mpich
+      - run: sudo apt install -y python-pybind11
+      - run: pip install
       - run: pip install pytest pytest-cov pybind11
       - run:
              pip install tox
