+++ tox.ini
@@ -8,4 +8,4 @@ deps =
     pytest
     pytest-cov
 commands =
-    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
+    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "serial"
