+++ tox.ini
@@ -5,6 +5,7 @@ isolated_build = True
 [testenv]
 deps =
     exdown
+    requests
     pytest
     pytest-cov
 commands =
