### 2020-08-30 8a6e21d
Misc. fixes (#34)
* Using tox to perform testing on CI

* Removed pybind11 submodule
