+++ tox.ini
@@ -1,0 +1,12 @@
+[tox]
+envlist = py3
+isolated_build = True
+
+[testenv]
+deps =
+    exdown
+    requests
+    pytest
+    pytest-cov
+commands =
+    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "serial"
