### 2020-08-31 e18ca7f
Update codecov.yml
turn off path fixing in CI.
