+++ codecov.yml
@@ -1,6 +1,6 @@
 comment: no
 codecov:
-  disable_default_path_fixes: true
+  disable_default_path_fixes: False
 fixes:
   - ".*/dist-packages/::"
   - ".*/site-packages/::"
