+++ tox.ini
@@ -8,5 +8,6 @@ deps =
     requests
     pytest
     pytest-cov
+    mpi4py
 commands =
-    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "serial"
+    mpirun -np 2 python3 -m pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
