+++ tox.ini
@@ -10,4 +10,4 @@ deps =
     pytest-cov
     mpi4py
 commands =
-    mpirun -np 2 python3 -m pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
+    mpirun -np 2 pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
