+++ tox.ini
@@ -9,5 +9,6 @@ deps =
     pytest
     pytest-cov
     mpi4py
+whitelist_externals = mpirun
 commands =
     mpirun -np 2 pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
