+++ .travis.yml
@@ -2,6 +2,7 @@ language: python
 python:
     - 3.6
 before_install:
+    - sudo apt-get update || true
     - sudo apt-get install libboost-all-dev
     - sudo apt-get install libmpfr-dev
     - git clone https://github.com/CGAL/cgal.git && cd cgal/ && mkdir build && cd build && cmake -DCMAKE_BUILD_TYPE=Release .. && sudo make install
