+++ tox.ini
@@ -11,4 +11,4 @@ deps =
     mpi4py
 whitelist_externals = mpirun
 commands =
-    mpirun -np 2 python -m pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
+    mpirun -np 2 pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
