### 2020-08-31 159bd48
installing openmp instead of mpich
