+++ tox.ini
@@ -8,5 +8,7 @@ deps =
     requests
     pytest
     pytest-cov
+    mpi4py
+whitelist_externals = mpirun
 commands =
-    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "serial"
+    mpirun -np 2 pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
