+++ .circleci/config.yml
@@ -12,15 +12,17 @@ jobs:
   build:
     working_directory: ~/working
     docker:
-      - image: ubuntu:18.04
+      - image: circleci/python:3
     steps:
       - checkout
+      - run: sudo apt-cache search openmpi
+      - run: sudo apt-get update
+      - run: sudo apt install -y libopenmpi libopenmpi-dev openmpi-bin
       - run: sudo apt install -y cmake
       - run: sudo apt-get install libboost-all-dev
       - run: sudo apt-get install libmpfr-dev
       - run: git clone https://github.com/CGAL/cgal.git && cd cgal/ && mkdir build && cd build && cmake -DCMAKE_BUILD_TYPE=Release .. && sudo make install
       - run: sudo apt install -y libhdf5-dev
-      - run: sudo apt install -y libopenmpi libopenmpi-dev openmpi-bin
       - run: sudo apt install -y python-pybind11
       - run: pip install . --user
       - run:
