+++ .circleci/config.yml
@@ -17,7 +17,7 @@ jobs:
       - checkout
       - run: sudo apt-cache search openmpi
       - run: sudo apt-get update
-      - run: sudo apt install -y libopenmpi libopenmpi-dev openmpi-bin
+      - run: sudo apt install -y libopenmpi3 libopenmpi-dev openmpi-bin
       - run: sudo apt install -y cmake
       - run: sudo apt-get install libboost-all-dev
       - run: sudo apt-get install libmpfr-dev
