+++ tox.ini
@@ -2,6 +2,9 @@
 envlist = py3
 isolated_build = True
 
+[mpi]
+mpiexec = {env:MPIEXEC:mpiexec}
+
 [testenv]
 deps =
     exdown
@@ -11,4 +14,4 @@ deps =
     mpi4py
 whitelist_externals = mpirun
 commands =
-    mpirun -np 2 pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
+    {[mpi]mpiexec} -n 2 pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
