+++ tox.ini
@@ -12,6 +12,6 @@ deps =
     pytest
     pytest-cov
     mpi4py
-whitelist_externals = mpirun
+whitelist_externals = {[mpi]mpiexec}
 commands =
     {[mpi]mpiexec} -n 2 pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
