+++ tox.ini
@@ -14,4 +14,4 @@ deps =
     mpi4py
 whitelist_externals = {[mpi]mpiexec}
 commands =
-    {[mpi]mpiexec} -n 2 pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
+    {[mpi]mpiexec} -n 2 -mca btl ^openib -mca oob_tcp_if_include ens5 -mca btl_tcp_if_include ens5 pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term
