### 2020-08-31 9d5ce32
suppress warning and viz output
