### 2020-08-31 025c328
parallel and serial tests, removing unused plots
