+++ SeismicMesh/plots/__init__.py
@@ -1,11 +1,0 @@
-from .points import plot_points3
-from .qual import dh_angles_hist, re_ratios_hist
-from .simpplot import plot_facets, plot_tets
-
-__all__ = [
-    "plot_tets",
-    "plot_facets",
-    "plot_points3",
-    "dh_angles_hist",
-    "re_ratios_hist",
-]
