+++ tox.ini
@@ -14,4 +14,5 @@ deps =
     mpi4py
 whitelist_externals = {[mpi]mpiexec}
 commands =
-    {[mpi]mpiexec} -n 2 -mca btl ^openib pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -s -m "parallel"
+    {[mpi]mpiexec} -n 2 -mca btl ^openib pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "parallel"
+    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "serial"
