+++ .circleci/config.yml
@@ -15,8 +15,6 @@ jobs:
       - image: circleci/python:3
     steps:
       - checkout
-      - run: sudo apt-cache search openmpi
-      - run: sudo apt-get update
       - run: sudo apt install -y libopenmpi3 libopenmpi-dev openmpi-bin
       - run: sudo apt install -y cmake
       - run: sudo apt-get install libboost-all-dev
