+++ tox.ini
@@ -15,4 +15,4 @@ deps =
 whitelist_externals = {[mpi]mpiexec}
 commands =
     {[mpi]mpiexec} -n 2 -mca btl ^openib pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "parallel"
-    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "serial"
+    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term --cov-append  -m "serial"
