### 2020-08-31 b0cdef4
removing debug commands
