### 2020-08-31 bcb0b0e
Adding parallel tests to CI (#39)
* Adding 3 parallel tests (one new one) to unit test parallel capabilities on CI. 

* Updated documentation to reflect the usage of `tox` for testing. 

*
