+++ SeismicMesh/generation/mesh_generator.py
@@ -260,9 +260,6 @@ class MeshGenerator:  # noqa: C901
                 ].astype(float)
                 p = p.reshape(dim, -1).T
 
-            # np.savetxt('init.txt', p, delimiter=',')
-            # np.savetxt('fd.txt', fd(p), delimiter=',')
-            # quit()
             # 2. Remove points outside the region, apply the rejection method
             p = p[fd(p) < geps]  # Keep only d<0 points
             r0 = fh(p)
@@ -294,6 +291,7 @@ class MeshGenerator:  # noqa: C901
                 else:
                     blocks = None
                     extents = None
+                fh = migration.localize_sizing_function(fh, h0, bbox, dim, _axis, comm)
                 # send points to each subdomain
                 p, extents = migration.localize_points(blocks, extents, comm, dim)
                 extent = extents[rank]
