+++ SeismicMesh/__init__.py
@@ -6,7 +6,7 @@
 #  have received a copy of the license along with this program. If not,
 #  see <http://www.gnu.org/licenses/>.
 
-from . import decomp, geometry, migration, plots
+from . import decomp, geometry, migration
 from .generation import MeshGenerator
 from .sizing import MeshSizeFunction, ReadSegy
 
@@ -17,5 +17,4 @@ __all__ = [
     "MeshGenerator",
     "decomp",
     "migration",
-    "plots",
 ]
