+++ tox.ini
@@ -2,11 +2,17 @@
 envlist = py3
 isolated_build = True
 
+[mpi]
+mpiexec = {env:MPIEXEC:mpiexec}
+
 [testenv]
 deps =
     exdown
     requests
     pytest
     pytest-cov
+    mpi4py
+whitelist_externals = {[mpi]mpiexec}
 commands =
-    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "serial"
+    {[mpi]mpiexec} -n 2 -mca btl ^openib pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "parallel"
+    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term --cov-append  -m "serial"
