+++ .circleci/config.yml
@@ -15,12 +15,12 @@ jobs:
       - image: circleci/python:3
     steps:
       - checkout
+      - run: sudo apt install -y libopenmpi3 libopenmpi-dev openmpi-bin
       - run: sudo apt install -y cmake
       - run: sudo apt-get install libboost-all-dev
       - run: sudo apt-get install libmpfr-dev
       - run: git clone https://github.com/CGAL/cgal.git && cd cgal/ && mkdir build && cd build && cmake -DCMAKE_BUILD_TYPE=Release .. && sudo make install
       - run: sudo apt install -y libhdf5-dev
-      - run: sudo apt install -y mpich
       - run: sudo apt install -y python-pybind11
       - run: pip install . --user
       - run:
