### 2020-08-31 97968ed
testing grad sizing function
