### 2020-08-31 38f1622
adding more tests for geometry
