### 2020-08-31 dc63a74
Increasing CI coverage (#40)
* Adding more unit tests
