+++ tox.ini
@@ -6,6 +6,8 @@ isolated_build = True
 mpiexec = {env:MPIEXEC:mpiexec}
 
 [testenv]
+setenv =
+    MPLBACKEND = agg
 deps =
     exdown
     requests
