### 2020-08-31 ff72d36
Adding one more CI test (#41)
* More testing of geometry.
