+++ tests/test_geometry2.py
@@ -22,6 +22,9 @@ def test_geometry2():
     wedges = geo.get_winded_boundary_edges(cells)
     assert np.allclose(wedges, [[0, 1], [1, 3], [2, 3], [0, 2]])
 
+    intersections = geo.doAnyOverlap(points, cells, dim=2)
+    assert len(intersections) == 0
+
     # test Laplacian2
     pdata = os.path.join(os.path.dirname(__file__), "points.txt")
     cdata = os.path.join(os.path.dirname(__file__), "cells.txt")
