### 2020-09-01 31bcc12
adding pytest.ini in sdist
