### 2020-09-01 b0b9933
Update README.md
Improving clarity and purpose on README
