+++ README.md
@@ -12,9 +12,13 @@
 [SeismicMesh](https://github.com/krober10nd/SeismicMesh): Mesh generation for Seismology in Python
 ==================================================================================================
 
-2D/3D triangular meshing for a slab of Earth based on modifications to
-the [DistMesh](http://persson.berkeley.edu/distmesh/) algorithm.
-SeismicMesh is distributed under GPL3.
+*SeismicMesh* enables the generation of potentially large (O(millions) cells) 2D/3D triangular meshes that are intended for acoustic and elastic finite element wave propagators. It builds these meshes directly from seismic velocity models such as [SEG-y files](https://en.wikipedia.org/wiki/SEG-Y) directly via Python. A [mesh sizing function module](https://seismicmesh.readthedocs.io/en/par3d/api.html#seimsicmesh-meshsizefunction) is provided to map variations in seismic velocities contained in a seismic velocity model to triangular mesh sizes while ensuring that mesh size transitions vary smoothly (e.g., graded) and the Courant number can be bounded above by a constant. *SeismicMesh* is an open source tool everyone can use that perhaps just knows a little Python and aren't very familiar with other more complex and powerful mesh generation tools out there like [gmsh](https://gmsh.info/) or [cgal](https://doc.cgal.org/latest/Mesh_3/index.html).
+
+We provide capabilites to perform both 2D and 3D triangular meshing in either serial or distributed memory parallel using the message passing interface. In 3D mesh generation, we provide a simple interface to a mesh improvement method ([sliver removal](https://hal.inria.fr/inria-00430202/document)) to ensure a minimum quality bound can be ensured and the mesh will work when simulation is attempted. 
+
+The mesh generation approach is based originally on the [DistMesh](http://persson.berkeley.edu/distmesh/) algorithm and in its most basic form can mesh *any* domain that can be defined by a [signed distance function](https://en.wikipedia.org/wiki/Signed_distance_function#:~:text=In%20mathematics%20and%20its%20applications,whether%20x%20is%20in%20%CE%A9.) and a [mesh sizing function](http://persson.berkeley.edu/pub/persson05qualmesh.pdf)
+
+SeismicMesh is distributed under GPL3 and more details can be found in our [short paper](https://github.com/krober10nd/SeismicMesh/blob/par3d/paper/paper.md).
 
 Installation
 ============
