+++ README.md
@@ -19,7 +19,7 @@ A [mesh sizing function module](https://seismicmesh.readthedocs.io/en/par3d/api.
 
 Capabilites are provided to perform both 2D and 3D triangular meshing in either serial or distributed memory parallel using the message passing interface. In 3D mesh generation, we provide a simple interface to a mesh improvement method ([sliver removal](https://hal.inria.fr/inria-00430202/document)) to ensure a minimum quality bound can be ensured and the mesh will remain numerically stable when simulation is attempted with an explicit timestepping scheme.
 
-With that said, the mesh generation approach provided in this package can be operated stanadlone and is based on the [DistMesh](http://persson.berkeley.edu/distmesh/) algorithm. Thus, in its most basic operation, SeismicMesh can mesh *any* domain that can be defined by a [signed distance function](https://en.wikipedia.org/wiki/Signed_distance_function#:~:text=In%20mathematics%20and%20its%20applications,whether%20x%20is%20in%20%CE%A9.) and a [mesh sizing function](http://persson.berkeley.edu/pub/persson05qualmesh.pdf)
+With that said, the mesh generation approach provided in this package can be operated stanadlone and is based on the [DistMesh](http://persson.berkeley.edu/distmesh/) algorithm. Thus, in its most basic operation, SeismicMesh can mesh *any* domain that can be defined by a [signed distance function](https://en.wikipedia.org/wiki/Signed_distance_function#:~:text=In%20mathematics%20and%20its%20applications,whether%20x%20is%20in%20%CE%A9.) with mesh sizes that follow variations described by a [mesh sizing function](http://persson.berkeley.edu/pub/persson05qualmesh.pdf)
 
 SeismicMesh is distributed under GPL3. More details can be found in our [short paper](https://github.com/krober10nd/SeismicMesh/blob/par3d/paper/paper.md).
 
