+++ README.md
@@ -13,13 +13,13 @@
 ==================================================================================================
 
 *SeismicMesh* enables the generation of 2D/3D triangular meshes that are intended for acoustic and elastic wave propagators discretized 
-with the finite element method. Mesh sizes are built to reflect variations in seismic velocities defined are regular Cartesian grids using short Python scripts.
+with the finite element method. Mesh sizes are built to reflect variations in seismic velocities from seismic velocity models defined on regular Cartesian grids using short Python scripts.
 
-One complication to generate unstructured meshes of complex and largescale 2D/3D geophysical domains with existing popular mesh generation tools such as [gmsh](gmsh.info) or [cgal](https://doc.cgal.org/latest/Mesh_3/index.html) is the necessity to define a [mesh sizing function](http://persson.berkeley.edu/pub/persson05qualmesh.pdf). These mesh generation software generally require a sizing function *to be defined on a mesh*, which creates a circular problem of building a mesh to build a mesh. Since analytical mesh sizing functions do not exist for a typical geophysical domain, typically users must create their own trianglar interpolant. For large scale geophysical problems, the evaluation of the a triangular interpolant quickly becomes a computational and memory bottleneck. For domains with mesh size variations in the interior of the domain, typical heuristics like [boundary layer/attractor adapation](https://gmsh.info/doc/texinfo/gmsh.html) or [characteristic size calculations](https://gmsh.info/doc/texinfo/gmsh.html#t5) are not directly relevant to size elements. 
+One complication to generate unstructured meshes of complex and largescale 2D/3D geophysical domains with existing popular mesh generation tools such as [gmsh](gmsh.info) or [cgal](https://doc.cgal.org/latest/Mesh_3/index.html) is the necessity to define a [mesh sizing function](http://persson.berkeley.edu/pub/persson05qualmesh.pdf). These mesh generation software generally require a sizing function *to be defined on a mesh*, which creates a circular problem of building a mesh to build a mesh. Since analytical mesh sizing functions do not exist for a general geophysical domain, typically users must create their own mesh sizing function. For domains with mesh size variations in the interior of the domain, typical mesh sizing heuristics like [boundary layer/attractor adapation](https://gmsh.info/doc/texinfo/gmsh.html) or [characteristic size calculations](https://gmsh.info/doc/texinfo/gmsh.html#t5) are not directly relevant to size elements for geophysical problems. In a typical geophysical domain, variations in mesh size should reflect material properites such as P-wave or S-wave velocity to efficiently model waves with a sufficient number of grid-points per wavelenth. 
 
-Thus in this work we provide an alternative approach to *easily* build large scale 2D/3D [mesh sizing function module](https://seismicmesh.readthedocs.io/en/par3d/api.html#seimsicmesh-meshsizefunction) that map variations in seismic velocities from a seismic velocity models to triangular mesh sizes. Importantly, the sizing module can ensure that mesh size transitions vary smoothly (e.g., are graded) and the Courant number can be bounded above by a constant (amongst other capabilities), which are important considerations for accurate and succesful FEM. These mesh sizing functions are defined on the same grids as the original seismic velocity model bypassing the need for the user to create their own sizing function and can be queried in parallel during mesh generation for quick execution.
[…]
