### 2020-09-01 b030a00
Update README.md (#45)
* Improving clarity and purpose on README
