+++ README.md
@@ -12,9 +12,18 @@
 [SeismicMesh](https://github.com/krober10nd/SeismicMesh): Mesh generation for Seismology in Python
 ==================================================================================================
 
-2D/3D triangular meshing for a slab of Earth based on modifications to
-the [DistMesh](http://persson.berkeley.edu/distmesh/) algorithm.
-SeismicMesh is distributed under GPL3.
+*SeismicMesh* enables the generation of 2D/3D triangular meshes that are intended for acoustic and elastic wave propagators discretized 
+with the finite element method. Mesh sizes are built to reflect variations in seismic velocities from seismic velocity models defined on regular Cartesian grids using short Python scripts.
+
+One complication to generate unstructured meshes of complex and largescale 2D/3D geophysical domains with existing popular mesh generation tools such as [gmsh](gmsh.info) or [cgal](https://doc.cgal.org/latest/Mesh_3/index.html) is the necessity to define a [mesh sizing function](http://persson.berkeley.edu/pub/persson05qualmesh.pdf). These mesh generation software generally require a sizing function *to be defined on a mesh*, which creates a circular problem of building a mesh to build a mesh. Since analytical mesh sizing functions do not exist for a general geophysical domain, typically users must create their own mesh sizing function. For domains with mesh size variations in the interior of the domain, typical mesh sizing heuristics like [boundary layer/attractor adapation](https://gmsh.info/doc/texinfo/gmsh.html) or [characteristic size calculations](https://gmsh.info/doc/texinfo/gmsh.html#t5) are not directly relevant to size elements for geophysical problems. In a typical geophysical domain, variations in mesh size should reflect material properites such as P-wave or S-wave velocity to efficiently model waves with a sufficient number of grid-points per wavelenth. 
+
+Thus in this work we provide an alternative approach to build large scale 2D/3D [mesh sizing function module](https://seismicmesh.readthedocs.io/en/par3d/api.html#seimsicmesh-meshsizefunction) that map variations in seismic velocities from a seismic velocity model to triangular mesh sizes. Importantly, the sizing module can ensure that mesh size transitions vary smoothly (e.g., are graded) and the Courant number can be bounded above by a constant--amongst other capabilities--which are important considerations for accurate and succesful simulation with finite elements. These mesh sizing functions are defined on the same  regular Cartesian grids as the original seismic velocity model bypassing the need for the user to create their own sizing function. There structured nature also enables efficient performance as they can be queried in parallel during mesh generation.
+
+In addition, *SeismicMesh* supports both 2D and 3D triangular meshing in either serial or distributed memory parallel using the message passing interface. In 3D mesh generation, we provide a mesh improvement method ([sliver removal](https://hal.inria.fr/inria-00430202/document)) to ensure a minimum quality bound (e.g. minimum dihedral angle) can be enforced and the mesh thus will remain numerically stable when simulation is attempted.
+
+The mesh generation approach provided in this package can be operated stanadlone (e.g., without the sizing function class) as it is based off modifications to the [DistMesh](http://persson.berkeley.edu/distmesh/) algorithm. Thus in its most basic operation, *SeismicMesh* can mesh *any* domain that can be defined by a [signed distance function](https://en.wikipedia.org/wiki/Signed_distance_function#:~:text=In%20mathematics%20and%20its%20applications,whether%20x%20is%20in%20%CE%A9.) with mesh sizes that follow variations described by a user-defined [mesh sizing function](http://persson.berkeley.edu/pub/persson05qualmesh.pdf)
+
+*SeismicMesh* is distributed under the GPL3 licesne and more details can be found in our [short paper](https://github.com/krober10nd/SeismicMesh/blob/par3d/paper/paper.md).
 
 Installation
 ============
