+++ README.md
@@ -12,16 +12,16 @@
 [SeismicMesh](https://github.com/krober10nd/SeismicMesh): Mesh generation for Seismology in Python
 ==================================================================================================
 
-*SeismicMesh* enables the generation of 2D/3D triangular meshes that are intended for acoustic and elastic wave propagators discretized 
-with the finite element method. Using short scripts, mesh sizes are built to reflect variations in seismic velocities from seismic velocity models defined on regular Cartesian grids.
+*SeismicMesh* is a tool to generate 2D/3D triangular meshes that are used for acoustic and elastic wave propagators discretized 
+with the finite element method. With short scripts, variable resolution meshes are built that have size transitions which reflect variations in P-wave or S-wave velocities. These data are typically provided on [regular Cartesian grids for global and regional domains](https://ds.iris.edu/spud/earthmodel).
 
-One complication to generate unstructured meshes of complex and large-scale 2D/3D geophysical domains with existing popular mesh generation tools such as [gmsh](gmsh.info) or [cgal](https://doc.cgal.org/latest/Mesh_3/index.html) is the necessity to define a [mesh sizing function](http://persson.berkeley.edu/pub/persson05qualmesh.pdf). These mesh generation software generally require a sizing function *to be defined on a mesh*, which creates a circular problem of building a mesh to build a mesh. Since analytical mesh sizing functions do not exist for a general geophysical domain, typically users must create their own mesh sizing function. For domains with mesh size variations in the interior of the domain, typical mesh sizing heuristics like [boundary layer/attractor adaptation](https://gmsh.info/doc/texinfo/gmsh.html) or [characteristic size calculations](https://gmsh.info/doc/texinfo/gmsh.html) are not directly relevant to size elements for geophysical problems. In a typical geophysical domain, variations in mesh size should reflect material properties such as P-wave or S-wave velocity to efficiently model waves with a sufficient number of grid-points per wavelength. 
+A complication that arises when ones wants to generate unstructured meshes of complex and large-scale 2D/3D geophysical domains with existing popular mesh generation tools such as [gmsh](gmsh.info) or [cgal](https://doc.cgal.org/latest/Mesh_3/index.html) is the necessity to define a [mesh sizing function](http://persson.berkeley.edu/pub/persson05qualmesh.pdf). These aforementioned mesh generation software generally require a sizing function *to be defined on a mesh*, which creates a circular problem of building a mesh to build a mesh. Since analytical mesh sizing functions do not exist for a general geophysical domain, typically users must create their own mesh sizing function and it isn't always clear how to incorporate the geophysical data into it. For domains with mesh size variations in the interior of the domain, typical mesh sizing heuristics like [boundary layer/attractor adaptation](https://gmsh.info/doc/texinfo/gmsh.html) or [characteristic size calculations](https://gmsh.info/doc/texinfo/gmsh.html) are not directly relevant to size elements for seismological problems. In a typical seismologial domain, variations in mesh size generally reflect internal material properties such as P-wave or S-wave velocity to cost-effectively model waves by ensuring there are sufficient number of grid-points per wavelength. 
 
[…]
