### 2020-09-04 #2592 @meg-simula
@jorgensd would you be available to review this submission for JOSS? Please take a look at our reviewer guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html, and let us know. Your help would be much appreciated.
