### 2020-09-05 ea87eca
Update example_2D_irregular.py
