+++ examples/example_2D_irregular.py
@@ -18,7 +18,7 @@ if size > 1:
 def example_2D_Irregular():
     """
     Build a 2D mesh of a benchmark velocity  model of a Monuntainous thrust region.
-    Builds in serial or parallel. Demonstrates how to ensure the topography is respected
+    Builds in serial. Demonstrates how to ensure the topography of the free surface is respected
     in the mesh.
 
     Velocity model was downloaded from here: https://wiki.seg.org/wiki/1994_BP_migration_from_topography
