+++ tests/test_grade.py
@@ -25,7 +25,7 @@ def CheckGrade(hh, nz, nx, elen):
 @pytest.mark.serial
 def test_grade():
     fname = os.path.join(os.path.dirname(__file__), "testing.segy")
-    vp = SeismicMesh.ReadSegy(fname)
+    vp = SeismicMesh.read_segy(fname)
     wl = 5
     hmin = 10
     grade = 0.005
