+++ tests/test_2dmesher_par.py
@@ -15,7 +15,7 @@ rank = comm.Get_rank()
 def test_2dmesher_par():
 
     fname = os.path.join(os.path.dirname(__file__), "testing.segy")
-    vp = SeismicMesh.ReadSegy(fname)
+    vp = SeismicMesh.read_segy(fname)
     wl = 5
     hmin = 100
     grade = 0.005
