+++ tests/test_2dmesher_par_adapt.py
@@ -14,7 +14,7 @@ rank = comm.Get_rank()
 @pytest.mark.parallel
 def test_2dpar_mesher_adapt():
     fname = os.path.join(os.path.dirname(__file__), "testing.segy")
-    vp = SeismicMesh.ReadSegy(fname)
+    vp = SeismicMesh.read_segy(fname)
     wl = 5
     hmin = 100
     grade = 0.005
