### 2020-09-06 dd4948a
removing some old utils in generation
