+++ SeismicMesh/generation/utils.py
@@ -4,27 +4,6 @@ import numpy as np
 import scipy.sparse as spsparse
 
 
-def closest_number(n, m):
-    # Find the quotient
-    q = int(n / m)
-
-    # 1st possible closest number
-    n1 = m * q
-
-    # 2nd possible closest number
-    if (n * m) > 0:
-        n2 = m * (q + 1)
-    else:
-        n2 = m * (q - 1)
-
-    # if true, then n1 is the required closest number
-    if abs(n - n1) < abs(n - n2):
-        return n1
-
-    # else n2 is the required closest number
-    return n2
-
-
 def make_init_points(bbox, rank, size, axis, h0, dim):
     """Create a structured grid in parallel of the entire domain
     Each processor owns a part of the domain.
@@ -46,39 +25,6 @@ def make_init_points(bbox, rank, size, axis, h0, dim):
     )
     points = points.reshape(dim, -1).T
     return points
-
-
-def __setdiff_rows(A, B, return_index=False):
-    """
-    Similar to MATLAB's setdiff(A, B, 'rows'), this returns C, I
-    where C are the row of A that are not in B and I satisfies
-    C = A[I,:].
-
-    Returns I if return_index is True.
-    """
-    A = np.require(A, requirements="C")
-    B = np.require(B, requirements="C")
-
-    assert A.ndim == 2, "array must be 2-dim'l"
-    assert B.ndim == 2, "array must be 2-dim'l"
-    assert A.shape[1] == B.shape[1], "arrays must have the same number of columns"
-    assert A.dtype == B.dtype, "arrays must have the same data type"
-
-    # NumPy provides setdiff1d, which operates only on one dimensional
-    # arrays. To make the array one-dimensional, we interpret each row
-    # as being a string of characters of the appropriate length.
-    orig_dtype = A.dtype
-    ncolumns = A.shape[1]
-    dtype = np.dtype((np.character, orig_dtype.itemsize * ncolumns))
-    C = (
-        np.setdiff1d(A.view(dtype), B.view(dtype))
-        .view(A.dtype)
-        .reshape((-1, ncolumns), order="C")
-    )
-    if return_index:
-        raise NotImplementedError
-    else:
-        return C
 
 
 def unique_rows(ar):
