+++ tests/test_2d_domain_ext.py
@@ -39,7 +39,7 @@ def test_2d_domain_ext():
         bbox=(-10e3, 0, 0, 10e3),
         grade=grade,
         domain_ext=500,
-        padstyle="constant",
+        pad_style="constant",
         wl=wl,
         freq=freq,
         velocity_grid=vp,
@@ -57,7 +57,7 @@ def test_2d_domain_ext():
         bbox=(-10e3, 0, 0, 10e3),
         grade=grade,
         domain_ext=2000,
-        padstyle="linear_ramp",
+        pad_style="linear_ramp",
         wl=wl,
         freq=freq,
         velocity_grid=vp,
