+++ SeismicMesh/geometry/__init__.py
@@ -10,8 +10,8 @@ from .utils import (
     SignedDistanceFunctionGenerator,
     calc_re_ratios,
     delete_boundary_entities,
-    doAnyOverlap,
-    fixmesh,
+    do_any_overlap,
+    fix_mesh,
     get_boundary_edges,
     get_boundary_entities,
     get_boundary_facets,
@@ -23,11 +23,11 @@ from .utils import (
     laplacian2,
     linter,
     remove_external_entities,
-    simpqual,
-    simpvol,
+    simp_qual,
+    simp_vol,
     unique_rows,
     vertex_to_entities,
-    vtInEntity3,
+    vertex_in_entity3,
 )
 
 __all__ = [
@@ -41,15 +41,15 @@ __all__ = [
     "drectangle",
     "dintersect",
     "calc_dihedral_angles",
-    "doAnyOverlap",
+    "do_any_overlap",
     "linter",
     "laplacian2",
     "vertex_to_entities",
     "remove_external_entities",
     "unique_rows",
-    "fixmesh",
-    "simpvol",
-    "simpqual",
+    "fix_mesh",
+    "simp_vol",
+    "simp_qual",
     "get_centroids",
     "get_edges",
     "get_facets",
@@ -59,5 +59,5 @@ __all__ = [
     "get_boundary_edges",
     "get_boundary_facets",
     "get_winded_boundary_edges",
-    "vtInEntity3",
+    "vertex_in_entity3",
 ]
