### 2020-09-06 edc57c1
pep'8 sizing modules
