+++ SeismicMesh/sizing/mesh_size_function.py
@@ -288,8 +288,6 @@ class MeshSizeFunction:
         assert value == "edge" or value == "constant" or value == "linear_ramp"
         self.__pad_style = value
 
-    # ---PUBLIC METHODS---#
-
     def build(self, comm=None):  # noqa: C901
 
         """Builds the isotropic mesh size function according
@@ -304,7 +302,7 @@ class MeshSizeFunction:
             size = 1
 
         if rank == 0:
-            self.__ReadVelocityModel()
+            self._read_velocity_model()
             _vp = self.velocity_grid
 
             _bbox = self.bbox
@@ -331,8 +329,10 @@ class MeshSizeFunction:
 
             if _dim == 2:
                 hh_m = np.zeros(shape=(_nz, _nx), dtype=np.float32) + _hmin
-            if _dim == 3:
+            elif _dim == 3:
                 hh_m = np.zeros(shape=(_nz, _nx, _ny), dtype=np.float32) + _hmin
+            else:
+                raise ValueError("Dimension not supported")
 
             if _wl > 0:
                 if rank == 0:
@@ -364,6 +364,8 @@ class MeshSizeFunction:
                     win_sqr_mean = ndimage.uniform_filter(
                         _vp ** 2, (win_rows, win_cols, win_cols2)
                     )
+                else:
+                    raise ValueError("Dimension not supported.")
                 win_var = win_sqr_mean - win_mean ** 2
                 # normalize variance to [0,1]
                 win_var /= np.amax(win_var)
@@ -399,13 +401,13 @@ class MeshSizeFunction:
                 if rank == 0:
                     # TODO: grader needs to consider varying constant resolution by dimension
                     print("Enforcing mesh gradation in sizing function...", flush=True)
-                    hh_m = self.hj(hh_m, _width / _nx, 10000)
+                    hh_m = self._hamilton_jacobi(hh_m, _width / _nx, 10000)
 
             # Domain extension
             # edit the bbox to reflect the new domain size
             if _domain_ext > 0:
-                self = self.__CreateDomainExtension()  # edit the bbox
-                hh_m = self.__EditMeshSizeFunction(
+                self = self._create_domain_extension()  # edit the bbox
+                hh_m = self._edit_mesh_size_function(
                     hh_m, rank
                 )  # edit the sizing function
 
@@ -414,10 +416,11 @@ class MeshSizeFunction:
                 print("Building a gridded interpolant...", flush=True)
 
             if _dim == 2:
-                z_vec, x_vec = self.__CreateDomainVectors()
-
-            if _dim == 3:
-                z_vec, x_vec, y_vec = self.__CreateDomainVectors()
+                z_vec, x_vec = self._create_domain_vectors()
+            elif _dim == 3:
+                z_vec, x_vec, y_vec = self._create_domain_vectors()
+            else:
+                raise ValueError("Dimension not supported")
 
             assert np.all(hh_m > 0.0), "Mesh size_function must be strictly positive."
 
@@ -425,10 +428,13 @@ class MeshSizeFunction:
                 self.interpolant = RegularGridInterpolator(
                     (z_vec, x_vec), hh_m, bounds_error=False, fill_value=None
                 )
-            if _dim == 3:
+            elif _dim == 3:
                 self.interpolant = RegularGridInterpolator(
                     (z_vec, x_vec, y_vec), hh_m, bounds_error=False, fill_value=None
                 )
+            else:
+                raise ValueError("Dimension not supported")
+
[…]
