### 2020-09-06 98aa86f
updating examples for compliance
