### 2020-09-06 3a15f5f
pep8 compliance (#51)
* pep8 compliance
