+++ SeismicMesh/sizing/mesh_size_function.py
@@ -27,7 +27,7 @@ from ..geometry import signed_distance_functions as sdf
 from .cpp import limgrad
 
 
-def ReadSegy(fname):
+def read_segy(fname):
     """Read a velocity model from a SEG-y file"""
     with segyio.open(fname, ignore_geometry=True) as f:
         nz, nx = len(f.samples), len(f.trace)
@@ -60,7 +60,7 @@ class MeshSizeFunction:
         ny=None,
         nz=None,
         domain_ext=0.0,
-        padstyle="edge",
+        pad_style="edge",
     ):
         """Class constructor for :class:`MeshSizeFunction`
 
@@ -90,8 +90,8 @@ class MeshSizeFunction:
         :type grade: float64, optional
         :param domain_ext: width of domain extension in `-z`, `+x`, `-x`, `+y`, `-y` directions
         :type domain_ext: float64, optional
-        :param padstyle: method to pad velocity in the domain extension
-        :type padstyle: str, optional, `edge`, `linear_ramp`, `constant`
+        :param pad_style: method to pad velocity in the domain extension
+        :type pad_style: str, optional, `edge`, `linear_ramp`, `constant`
 
         :return: object populated with meta-data, :math:`f(h)`, and a :math:`f(d)`.
         :rtype: :class:`MeshSizeFunction` object
@@ -122,7 +122,7 @@ class MeshSizeFunction:
         self.ny = ny
         self.nz = nz
         self.domain_ext = domain_ext
-        self.padstyle = padstyle
+        self.pad_style = pad_style
         self.interpolant = None
 
     ### SETTERS AND GETTERS ###
@@ -280,15 +280,13 @@ class MeshSizeFunction:
         self.__domain_ext = value
 
     @property
-    def padstyle(self):
-        return self.__padstyle
+    def pad_style(self):
+        return self.__pad_style
 
-    @padstyle.setter
-    def padstyle(self, value):
+    @pad_style.setter
+    def pad_style(self, value):
         assert value == "edge" or value == "constant" or value == "linear_ramp"
-        self.__padstyle = value
-
-    # ---PUBLIC METHODS---#
+        self.__pad_style = value
 
     def build(self, comm=None):  # noqa: C901
 
@@ -304,7 +302,7 @@ class MeshSizeFunction:
             size = 1
 
         if rank == 0:
-            self.__ReadVelocityModel()
+            self._read_velocity_model()
             _vp = self.velocity_grid
 
             _bbox = self.bbox
@@ -331,8 +329,10 @@ class MeshSizeFunction:
 
             if _dim == 2:
                 hh_m = np.zeros(shape=(_nz, _nx), dtype=np.float32) + _hmin
-            if _dim == 3:
+            elif _dim == 3:
                 hh_m = np.zeros(shape=(_nz, _nx, _ny), dtype=np.float32) + _hmin
+            else:
+                raise ValueError("Dimension not supported")
 
             if _wl > 0:
                 if rank == 0:
@@ -364,6 +364,8 @@ class MeshSizeFunction:
                     win_sqr_mean = ndimage.uniform_filter(
                         _vp ** 2, (win_rows, win_cols, win_cols2)
                     )
+                else:
+                    raise ValueError("Dimension not supported.")
                 win_var = win_sqr_mean - win_mean ** 2
                 # normalize variance to [0,1]
                 win_var /= np.amax(win_var)
@@ -399,13 +401,13 @@ class MeshSizeFunction:
                 if rank == 0:
                     # TODO: grader needs to consider varying constant resolution by dimension
                     print("Enforcing mesh gradation in sizing function...", flush=True)
-                    hh_m = self.hj(hh_m, _width / _nx, 10000)
+                    hh_m = self._hamilton_jacobi(hh_m, _width / _nx, 10000)
 
[…]
