### 2020-09-07 01deb88
complete rewrite of sizing module
