+++ SeismicMesh/sizing/__init__.py
@@ -1,3 +1,13 @@
-from .mesh_size_function import MeshSizeFunction, read_segy
+from .mesh_size_function import (
+    get_sizing_function_from_segy,
+    write_velocity_model,
+    plot_sizing_function,
+)
 
-__all__ = ["MeshSizeFunction", "read_segy"]
+__all__ = [
+    "get_sizing_function_from_segy",
+    "write_velocity_model",
+    "plot_sizing_function",
+]
+
+#
