+++ SeismicMesh/sizing/mesh_size_function.py
@@ -12,7 +12,6 @@
 #
 #   You should have received a copy of the GNU General Public License
 #   along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
 import warnings
 
 import h5py
@@ -23,796 +22,393 @@ from mpi4py import MPI
 from scipy import ndimage
 from scipy.interpolate import RegularGridInterpolator
 
-from ..geometry import signed_distance_functions as sdf
 from .cpp import limgrad
 
+__all__ = [
+    "get_sizing_function_from_segy",
+    "write_velocity_model",
+    "plot_sizing_function",
+]
 
-def read_segy(fname):
+# reasonable sizing function options go here
+opts = {
+    "hmin": 150.0,
+    "hmax": 1000.0,
+    "wl": 0,
+    "freq": 2.0,
+    "grad": 0.0,
+    "grade": 0.0,
+    "space_order": 1,
+    "dt": 0.0,
+    "cr_max": 1.0,
+    "pad_style": "edge",
+    "domain_ext": 0.0,
+    "units": "m-s",
+}
+
+
+def get_sizing_function_from_segy(filename, bbox, comm=None, **kwargs):
+    """Build a mesh size function from a seismic velocity model.
+
+    :param filename: the name of a SEG-y file containing a velocity model
+    :type filename: string
+    :param bbox: bounding box containing domain extents.
+    :type bbox: tuple with size (2*dim). For example, in 2D `(zmin, zmax, xmin, xmax)`
+
+    :return: cell_size_function: a function that takes a point and gets a size
+    :rtype: a Python function object.
+
+    """
+    comm = comm or MPI.COMM_WORLD
+    if comm.rank == 0:
+        vp, nz, nx, ny = _read_velocity_model(filename)
+
+        if opts["units"] == "km-s":
+            vp *= 1000.0
+
+        # check the bbox
+        if len(bbox) == 4:
+            dim = 2
+        elif len(bbox) == 6:
+            dim = 3
+        else:
+            raise ValueError("Dimension not supported")
+
+        opts.update(kwargs)
+
+        cell_size = _initialize_sizing_function(dim, opts["hmin"], nz, nx, ny)
+
+        for key in kwargs:
+            if key in {
+                "hmin",
+                "hmax",
+                "wl",
+                "freq",
+                "cr_max",
+                "dt",
+                "space_order",
+                "grad",
+                "grade",
+                "pad_style",
+                "domain_ext",
+                "m-s",
+            }:
+                pass
+            else:
+                raise ValueError(
+                    "Option %s with parameter %s not recognized " % (key, kwargs[key])
+                )
+        cell_size = _wavelength_sizing(vp, opts["wl"], opts["freq"])
+        cell_size = np.minimum(cell_size, _gradient_sizing(vp, opts["grad"]))
+
+        cell_size = _enforce_courant_sizing(
+            vp, cell_size, opts["cr_max"], opts["dt"], opts["space_order"]
+        )
+
+        cell_size = _enforce_gradation_sizing(
+            cell_size, opts["grade"], (bbox[1] - bbox[0]) / nz
+        )
+
+        cell_size, vp, bbox = _build_domain_extension(cell_size, vp, bbox, opts)
+
+        print("Enforcing minimum element size of " + str(opts["hmin"]))
+        cell_size[cell_size < opts["hmin"]] = opts["hmin"]
+
+        print("Enforcing minimum element size of " + str(opts["hmax"]))
+        cell_size[cell_size > opts["hmax"]] = opts["hmax"]
+
+        sizing_function = _build_sizing_function(cell_size, vp, bbox)
+
[…]
