### 2020-09-07 8d47fc6
dealing with the case of no wl
