+++ SeismicMesh/sizing/mesh_size_function.py
@@ -1,416 +1,416 @@
- #   Copyright (C) 2020 Keith Roberts
- #
- #   This program is free software: you can redistribute it and/or modify
- #   it under the terms of the GNU General Public License as published by
- #   the Free Software Foundation, either version 3 of the License, or
- #   (at your option) any later version.
- #
- #   This program is distributed in the hope that it will be useful,
- #   but WITHOUT ANY WARRANTY; without even the implied warranty of
- #   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- #   GNU General Public License for more details.
- #
- #   You should have received a copy of the GNU General Public License
- #   along with this program.  If not, see <http://www.gnu.org/licenses/>.
- import warnings
- 
- import h5py
- import matplotlib.pyplot as plt
- import numpy as np
- import segyio
- from mpi4py import MPI
- from scipy import ndimage
- from scipy.interpolate import RegularGridInterpolator
- 
- from .cpp import limgrad
- 
- __all__ = [
-     "get_sizing_function_from_segy",
-     "write_velocity_model",
-     "plot_sizing_function",
- ]
- 
- # reasonable sizing function options go here
- opts = {
-     "hmin": 150.0,
-     "hmax": 1000.0,
-     "wl": 0,
-     "freq": 2.0,
-     "grad": 0.0,
-     "grade": 0.0,
-     "space_order": 1,
-     "dt": 0.0,
-     "cr_max": 1.0,
-     "pad_style": "edge",
-     "domain_ext": 0.0,
-     "units": "m-s",
- }
- 
- 
- def get_sizing_function_from_segy(filename, bbox, comm=None, **kwargs):
-     """Build a mesh size function from a seismic velocity model.
- 
-     :param filename: the name of a SEG-y file containing a velocity model
-     :type filename: string
-     :param bbox: bounding box containing domain extents.
-     :type bbox: tuple with size (2*dim). For example, in 2D `(zmin, zmax, xmin, xmax)`
- 
-     :return: cell_size_function: a function that takes a point and gets a size
-     :rtype: a Python function object.
- 
-     """
-     comm = comm or MPI.COMM_WORLD
-     if comm.rank == 0:
-         vp, nz, nx, ny = _read_velocity_model(filename)
- 
-         if opts["units"] == "km-s":
-             vp *= 1000.0
- 
-         # check the bbox
-         if len(bbox) == 4:
-             dim = 2
-         elif len(bbox) == 6:
-             dim = 3
-         else:
-             raise ValueError("Dimension not supported")
- 
-         opts.update(kwargs)
- 
-         cell_size = _initialize_sizing_function(dim, opts["hmin"], nz, nx, ny)
- 
-         for key in kwargs:
-             if key in {
-                 "hmin",
-                 "hmax",
-                 "wl",
-                 "freq",
-                 "cr_max",
-                 "dt",
-                 "space_order",
-                 "grad",
-                 "grade",
-                 "pad_style",
-                 "domain_ext",
-                 "units", 
-             }:
-                 pass
-             else:
-                 raise ValueError(
-                     "Option %s with parameter %s not recognized " % (key, kwargs[key])
-                 )
-         cell_size = np.minimum(
-             _wavelength_sizing(vp, opts["wl"], opts["freq"]),
-             _gradient_sizing(vp, opts["grad"]),
-         )
- 
-         cell_size = _enforce_courant_sizing(
[…]
