### 2020-09-07 7615567
3d binary reader
