### 2020-09-07 e53d10b
simplifiying mesh generator
