+++ SeismicMesh/generation/mesh_generator.py
@@ -15,6 +15,7 @@
 
 import math
 import time
+import collections
 
 import numpy as np
 from mpi4py import MPI
@@ -24,478 +25,341 @@ from . import utils as mutils
 from .cpp.delaunay_class import DelaunayTriangulation as DT2
 from .cpp.delaunay_class3 import DelaunayTriangulation3 as DT3
 
+__all__ = ["generate_mesh"]
 
-class MeshGenerator:  # noqa: C901
-    """Class constructor for :class:`MeshGenerator`. User can also register their callbacks to
-    the sizing function :math:`f(h)` and signed distance function `f(d)` manually.
+opts = {
+    "nscreen": 1,
+    "max_iter": 50,
+    "seed": 0,
+    "perform_checks": False,
+    "pfix": None,
+    "axis": 1,
+}
 
-    :param SizingFunction: A :class:`MeshSizeFunction` object with populated fields and callbacks to `fd` and `fh`.
-    :type SizingFunction:  A :class:`MeshSizeFunction` class object. Required if no `fd` or `fh` are passed.
-    :param fd: A function that accepts an array of points and returns the signed distance to the boundary of the domain.
-    :type fd: A function object, optional if no :class:`SizingFunction` is passed
-    :param fh: A call-back function that accepts an array of points and returns an array of desired triangular mesh sizes close by to each point in the passed array.
-    :type fh: A function object, optional if no :class:`SizingFunction` is passed.
-    :param bbox: bounding box containing domain extents.
-    :type bbox: tuple with size (2*dim). For example, in 2D `(zmin, zmax, xmin, xmax)`. Optional if no :class:`SizingFunction` is passed.
-    :param hmin: minimum triangular edgelength populating the domain in meters.
-    :type hmin: float64,optional if no :class:`SizingFunction` is passed.
-    :param pfix: points that you wish you constrain in location.
-    :type pfix: nested list [num_fixed_points x dim], optional
 
-    :return: object populated with meta-data.
-    :rtype: :class:`MeshGenerator` object
+def generate_mesh(bbox, signed_distance_function, h0, cell_size, comm=None, **kwargs):
+    """Generate a 2D/3D triangulation using callbacks to a sizing function `cell_size`
+       and signed distance function `signed_distance_function`
+
+
+    :return: p:
+    :rtype:
+    :return: t:
+    :rtype:
+
     """
 
-    def __init__(
-        self,
-        SizingFunction=None,
-        fd=None,
-        fh=None,
-        bbox=None,
-        hmin=None,
-        pfix=None,
-    ):
-        self.SizingFunction = SizingFunction
-        self.fd = fd
-        self.fh = fh
-        self.bbox = bbox
-        self.hmin = hmin
-        self.pfix = pfix
-
-    # SETTERS AND GETTERS
-    @property
-    def SizingFunction(self):
-        return self.__SizingFunction
-
-    @SizingFunction.setter
-    def SizingFunction(self, value):
-        self.__SizingFunction = value
-
-    @property
-    def fd(self):
-        return self.__fd
-
-    @fd.setter
-    def fd(self, value):
-        self.__fd = value
-
-    @property
-    def fh(self):
-        return self.__fh
-
-    @fh.setter
-    def fh(self, value):
-        self.__fh = value
-
-    @property
-    def bbox(self):
-        return self.__bbox
-
-    @bbox.setter
-    def bbox(self, value):
-        if value is None:
-            self.__bbox = value
+    # check call was correct
+    opts.update(kwargs)
+    for key in kwargs:
+        if key in {
+            "nscreen",
+            "max_iter",
+            "seed",
+            "perform_checks",
+            "pfix",
+            "axis",
+        }:
+            pass
         else:
-            assert (
[…]
