+++ SeismicMesh/generation/__init__.py
@@ -1,7 +1,7 @@
 from .cpp.c_cgal import circumballs2, circumballs3, delaunay2, delaunay3
 from .cpp.delaunay_class import DelaunayTriangulation
 from .cpp.delaunay_class3 import DelaunayTriangulation3
-from .mesh_generator import MeshGenerator
+from .mesh_generator import generate_mesh
 
 __all__ = [
     "DelaunayTriangulation",
@@ -10,5 +10,5 @@ __all__ = [
     "delaunay3",
     "circumballs2",
     "circumballs3",
-    "MeshGenerator",
+    "generate_mesh",
 ]
