+++ SeismicMesh/generation/mesh_generator.py
@@ -25,7 +25,7 @@ from . import utils as mutils
 from .cpp.delaunay_class import DelaunayTriangulation as DT2
 from .cpp.delaunay_class3 import DelaunayTriangulation3 as DT3
 
-__all__ = ["generate_mesh"]
+__all__ = ["improve_mesh", "generate_mesh"]
 
 opts = {
     "nscreen": 1,
@@ -34,7 +34,140 @@ opts = {
     "perform_checks": False,
     "pfix": None,
     "axis": 1,
+    "min_dh_bound": 10.0,
+    "max_dh_bound": 170.0,
+    "points": None,
 }
+
+
+def improve_mesh(points, bbox, signed_distance_function, h0, comm=None, **kwargs):
+    """Improve an existing 3D mesh by removing degenerate elements called
+    slivers
+
+
+    """
+    comm = comm or MPI.COMM_WORLD
+    if comm.size > 1:
+        raise NotImplementedError("Sliver removal only works in serial for now")
+
+    opts.update(kwargs)
+    _parse_kwargs(kwargs)
+
+    dim = points.ndim
+    if dim == 2:
+        raise Exception("Mesh improvement currently on works in 3D")
+
+    if not isinstance(signed_distance_function, collections.Callable):
+        raise ValueError("`signed_distance_function` is not a function!")
+    fd = signed_distance_function
+
+    if opts["max_iter"] < 0:
+        raise ValueError("`max_iter` must be > 0")
+    max_iter = opts["max_iter"]
+
+    geps = 1e-1 * h0
+    deps = np.sqrt(np.finfo(np.double).eps) * h0
+    min_dh_bound = opts["min_dh_bound"] * math.pi / 180
+    max_dh_bound = opts["max_dh_bound"] * math.pi / 180
+
+    DT = _select_cgal_dim(dim)
+
+    p = points
+
+    N = len(p)
+
+    print(
+        "Commencing sliver removal with %d vertices on rank %d." % (N, comm.rank),
+        flush=True,
+    )
+
+    count = 0
+    pold = None
+    nscreen = opts["nscreen"]
+
+    while True:
+
+        start = time.time()
+
+        # Using CGAL's incremental Delaunay triangulation capabilities.
+        if count == 0:
+            dt = DT()
+            dt.insert(p.flatten().tolist())
+        else:
+            to_move = np.where(_dist(p, pold) > 0)[0]
+            dt.move(to_move.flatten().tolist(), p[to_move].flatten().tolist())
+
+        # Get the current topology of the triangulation
+        p, t = _get_topology(dt)
+
+        pold = p.copy()
+
+        N = p.shape[0]
+
+        # Remove points outside the domain
+        t = _remove_triangles_outside(p, t, fd, geps)
+
+        # Sliver removal
+        if count != (max_iter - 1):
+            num_move = 0
+            # calculate dihedral angles in mesh
+            dh_angles = geometry.calc_dihedral_angles(p, t)
+            out_of_bounds = np.argwhere(
+                (dh_angles[:, 0] < min_dh_bound) | (dh_angles[:, 0] > max_dh_bound)
+            )
+            ele_nums = np.floor(out_of_bounds / 6).astype("int")
+            ele_nums, ix = np.unique(ele_nums, return_index=True)
+
+            if count % nscreen == 0:
+                print(
+                    "On rank: "
+                    + str(comm.rank)
+                    + " There are "
+                    + str(len(ele_nums))
+                    + " slivers...",
+                    flush=True,
+                )
+
+            move = t[ele_nums, 0]
+            num_move = move.size
+            if num_move == 0:
+                print("Termination reached...no slivers detected!", flush=True)
+                p, t, _ = geometry.fix_mesh(p, t, dim=dim, delete_unused=True)
[…]
