+++ tests/test_3dmesher.py
@@ -3,56 +3,54 @@ import os
 import numpy as np
 import pytest
 
-import SeismicMesh
+from SeismicMesh import (
+    generate_mesh,
+    geometry,
+    get_sizing_function_from_segy,
+    write_velocity_model,
+)
 
 
 @pytest.mark.serial
 def test_3dmesher():
 
     fname = os.path.join(os.path.dirname(__file__), "test3D.bin")
-    nz, nx, ny = 20, 10, 10
-
-    # Load data
-    with open(fname, "r") as file:
-        vp = np.fromfile(file, dtype=np.dtype("float32").newbyteorder("<"))
-        vp = vp.reshape(nx, ny, nz, order="F")
-        vp = np.flipud(vp.transpose((2, 0, 1)))  # z, x and then y
 
     wl = 10
     hmin = 50
     freq = 2
     grade = 0.005
-    ef = SeismicMesh.MeshSizeFunction(
-        bbox=(-2e3, 0, 0, 1e3, 0, 1e3),
+    bbox = (-2e3, 0.0, 0.0, 1e3, 0, 1e3)
+    ef, bbox = get_sizing_function_from_segy(
+        fname,
+        bbox,
         grade=grade,
         grad=hmin,
         freq=freq,
         wl=wl,
-        velocity_grid=vp,
         hmin=hmin,
+        nz=20,
+        nx=10,
+        ny=10,
     )
-    ef = ef.build()
-    ef.write_velocity_model("foo3d")
-    mshgen = SeismicMesh.MeshGenerator(ef)
-    points, cells = mshgen.build(nscreen=1, max_iter=50, seed=0)
+
+    write_velocity_model(fname, nz=20, nx=10, ny=10)
+
+    def cube(p):
+        return geometry.dblock(p, *bbox)
+
+    points, cells = generate_mesh(
+        bbox=bbox, signed_distance_function=cube, cell_size=ef, h0=hmin
+    )
     print(len(points), len(cells))
     # 16459 90552
 
     assert len(points) == 16459
     assert np.abs(len(cells) - 90552) < 600
 
-    points, cells = mshgen.build(points=points, mesh_improvement=True)
-
     import meshio
 
     meshio.write_points_cells("food3D.vtk", points, [("tetra", cells)])
-    meshio.write_points_cells(
-        "foo3d.msh",
-        points / 1000.0,
-        [("tetra", cells)],
-        file_format="gmsh22",
-        binary=False,
-    )
 
 
 if __name__ == "__main__":
