### 2020-09-08 eb4f333
change name: sliver_removal. fix test2d
