+++ tests/test_2dmesher.py
@@ -16,16 +16,22 @@ from SeismicMesh import (
 def test_2dmesher():
 
     fname = os.path.join(os.path.dirname(__file__), "testing.segy")
+    bbox = (-10e3, 0.0, 0.0, 10e3)
     wl = 5
+    freq = 5.0
     hmin = 100
+    hmax = 10e6
     grade = 0.005
+    grad = 50.0
     ef, bbox = get_sizing_function_from_segy(
         fname,
-        bbox=(-10e3, 0, 0, 10e3),
+        bbox=bbox,
         grade=grade,
-        grad=50.0,
+        grad=grad,
         wl=wl,
+        freq=freq,
         hmin=hmin,
+        hmax=hmax,
     )
     plot_sizing_function(ef, bbox)
     write_velocity_model(fname)
@@ -40,8 +46,19 @@ def test_2dmesher():
         h0=hmin,
         perform_checks=True,
     )
-    assert np.abs(len(points) - 7906) < 20
-    assert np.abs(len(cells) - 15487) < 20
+    # should have: 7690 vertices and 15045 cells
+    print(len(points), len(cells))
+    assert np.abs(len(points) - 7690) < 20
+    assert np.abs(len(cells) - 15045) < 20
+
+    # import meshio
+
+    # meshio.write_points_cells(
+    #    "blah.vtk",
+    #    points[:, [1, 0]] / 1000,
+    #    [("triangle", cells)],
+    #    file_format="vtk",
+    # )
 
 
 if __name__ == "__main__":
