### 2020-09-08 f8dcdb5
various fixes and consolidating tests
