+++ tests/test_mesher_serial.py
@@ -1,0 +1,114 @@
+import os
+
+from numpy import abs
+import pytest
+
+from SeismicMesh import (
+    generate_mesh,
+    sliver_removal,
+    geometry,
+    get_sizing_function_from_segy,
+    plot_sizing_function,
+    write_velocity_model,
+)
+
+
+@pytest.mark.serial
+def test_2dmesher():
+
+    fname = os.path.join(os.path.dirname(__file__), "testing.segy")
+    bbox = (-10e3, 0.0, 0.0, 10e3)
+    wl = 5
+    freq = 5.0
+    hmin = 100
+    hmax = 10e6
+    grade = 0.005
+    grad = 50.0
+    ef, bbox = get_sizing_function_from_segy(
+        fname,
+        bbox=bbox,
+        grade=grade,
+        grad=grad,
+        wl=wl,
+        freq=freq,
+        hmin=hmin,
+        hmax=hmax,
+    )
+    plot_sizing_function(ef, bbox)
+    write_velocity_model(fname)
+
+    def rectangle(p):
+        return geometry.drectangle(p, *bbox)
+
+    points, cells = generate_mesh(
+        bbox=bbox,
+        signed_distance_function=rectangle,
+        cell_size=ef,
+        h0=hmin,
+        perform_checks=True,
+    )
+    # should have: 7690 vertices and 15045 cells
+    print(len(points), len(cells))
+    assert abs(len(points) - 7690) < 20
+    assert abs(len(cells) - 15045) < 20
+
+    # import meshio
+
+    # meshio.write_points_cells(
+    #    "blah.vtk",
+    #    points[:, [1, 0]] / 1000,
+    #    [("triangle", cells)],
+    #    file_format="vtk",
+    # )
+
+
+@pytest.mark.serial
+def test_3dmesher():
+
+    fname = os.path.join(os.path.dirname(__file__), "test3D.bin")
+
+    wl = 10
+    freq = 2
+    hmin = 50
+    grade = 0.005
+    bbox = (-2e3, 0.0, 0.0, 1e3, 0, 1e3)
+    ef, bbox = get_sizing_function_from_segy(
+        fname,
+        bbox,
+        grade=grade,
+        grad=hmin,
+        freq=freq,
+        wl=wl,
+        hmin=hmin,
+        nz=20,
+        nx=10,
+        ny=10,
+        byte_order="little",
+    )
+
+    write_velocity_model(
+        fname, nz=20, nx=10, ny=10, byte_order="little", ofname="testing"
+    )
+
+    def cube(p):
+        return geometry.dblock(p, *bbox)
+
+    points, cells = generate_mesh(
+        bbox=bbox,
+        signed_distance_function=cube,
+        cell_size=ef,
+        h0=hmin,
+    )
+
+    points, cells = sliver_removal(
+        points=points, bbox=bbox, signed_distance_function=cube, h0=hmin
+    )
+    print(len(points), len(cells))
+    # 16459 89240
+
+    assert abs(len(points) - 16459) < 1000
+    assert abs(len(cells) - 89240) < 1000
+
+    # import meshio
+
+    # meshio.write_points_cells("food3D.vtk", points, [("tetra", cells)])
