### 2020-09-08 950f6bc
adding in 3d SDF test
