### 2020-09-08 3cc0d5e
updating README to reflect new API
