+++ tests/test_2dmesher_par.py
@@ -45,10 +45,7 @@ def test_2dmesher_par():
             file_format="vtk",
         )
         area = geometry.simp_vol(points / 1000, cells)
-        # 7658 vertices and 14965
         assert np.abs(100 - np.sum(area)) < 0.50  # km2
-        assert np.abs(7658 - len(points)) < 100
-        assert np.abs(14965 - len(cells)) < 100
 
 
 if __name__ == "__main__":
