+++ tests/test_3dmesher_par_adapt.py
@@ -4,79 +4,70 @@ import numpy as np
 import pytest
 from mpi4py import MPI
 
-import SeismicMesh
+from SeismicMesh import get_sizing_function_from_segy, generate_mesh, geometry
 
 comm = MPI.COMM_WORLD
-size = comm.Get_size()
-rank = comm.Get_rank()
 
 
 @pytest.mark.parallel
 def test_3dpar_mesher_adapt():
     fname = os.path.join(os.path.dirname(__file__), "test3D.bin")
     nz, nx, ny = 20, 10, 10
-
-    # Load data
-    with open(fname, "r") as file:
-        vp = np.fromfile(file, dtype=np.dtype("float32").newbyteorder("<"))
-        vp = vp.reshape(nx, ny, nz, order="F")
-        vp = np.flipud(vp.transpose((2, 0, 1)))  # z, x and then y
-
-    wl = 10
+    bbox = (-2e3, 0.0, 0.0, 1e3, 0.0, 1e3)
     hmin = 50
+    wl = 10
     freq = 4
     grade = 0.15
-    nz, nx, ny = 20, 10, 10
-    ef = SeismicMesh.MeshSizeFunction(
-        bbox=(-2e3, 0, 0, 1e3, 0, 1e3),
+    ef, bbox = get_sizing_function_from_segy(
+        fname,
+        bbox,
+        hmin=hmin,
+        wl=wl,
+        freq=freq,
+        grade=grade,
         nx=nx,
         ny=ny,
         nz=nz,
-        grade=grade,
-        freq=freq,
-        wl=wl,
-        velocity_grid=vp,
-        hmin=hmin,
+        byte_order="little",
     )
-    ef = ef.build()
 
-    mshgen = SeismicMesh.MeshGenerator(ef)
+    def cube(p):
+        return geometry.dblock(p, *bbox)
 
-    # generate a mesh
-    points, cells = mshgen.build(
-        max_iter=10,
-        axis=0,
+    points, cells = generate_mesh(
+        bbox=bbox, h0=hmin, cell_size=ef, signed_distance_function=cube
     )
+
     points = comm.bcast(points, 0)
 
-    # pass the points and restart with a different axis
-    points, cells = mshgen.build(
+    points, cells = generate_mesh(
         points=points,
-        max_iter=10,
+        bbox=bbox,
+        h0=hmin,
+        cell_size=ef,
+        signed_distance_function=cube,
         axis=1,
     )
 
     points = comm.bcast(points, 0)
 
-    # pass the points and restart with a different axis
-    points, cells = mshgen.build(
+    points, cells = generate_mesh(
         points=points,
-        max_iter=10,
+        bbox=bbox,
+        h0=hmin,
+        cell_size=ef,
+        signed_distance_function=cube,
         axis=2,
     )
 
-    if rank == 0:
-        # import meshio
+    if comm.rank == 0:
+        import meshio
 
-        # meshio.write_points_cells(
-        #    "foo3D_V3.vtk",
-        #    points,
-        #    [("tetra", cells)],
-        # )
+        meshio.write_points_cells(
+            "foo3D_V3.vtk",
+            points,
+            [("tetra", cells)],
+        )
 
-        vol = SeismicMesh.geometry.simp_vol(points / 1000, cells)
+        vol = geometry.simp_vol(points / 1000, cells)
         assert np.abs(2 - np.sum(vol)) < 0.10  # km2
-
-
-if __name__ == "__main__":
-    test_3dpar_mesher_adapt()
