### 2020-09-08 1767276
adding docstring for new function
