+++ docs/source/usrman/api.rst
@@ -1,7 +1,7 @@
 Modules
 ==============
 
-Here we document the public API (e.g., functions, classes and their methods)
+Here we document the public API
 
 *SeimsicMesh.Geometry*
 -------------------------------
@@ -11,21 +11,21 @@ Routines to perform geometrical/topological operations and calculate things on m
 .. automodule:: SeismicMesh.geometry
     :members:
 
-*SeimsicMesh.MeshSizeFunction*
--------------------------------
+*SeimsicMesh.get_sizing_function_from_segy*
+---------------------------------------------
 
-Convenience class to build a :math:`f(h)` mesh sizing function from a seismic velocity model.
+Function to build a :math:`f(h)` mesh sizing function from a seismic velocity model.
 Assumes the domain can be represented by a rectange (2D) or cube (3D) and thus builds a :math:`f(d)` accordingly.
-This class can be passed to :class:`MeshGenerator` to build a simplical mesh with varying mesh resolution.
+The sizing function can be passed to `generate_mesh` to build a simplical mesh with varying mesh resolution.
 
-.. autoclass:: SeismicMesh.MeshSizeFunction
+.. automodule:: SeismicMesh.get_sizing_function_from_segy
     :members:
 
-*SeimsicMesh.MeshGenerator*
+*SeimsicMesh.generate_mesh*
 -------------------------------
 
-Class to build a simplical mesh that conforms to the signed distance function :math:`f(d)`
+Function to build a simplical mesh that conforms to the signed distance function :math:`f(d)`
 and :math:`f(h)`.
 
-.. autoclass:: SeismicMesh.MeshGenerator
+.. automodule:: SeismicMesh.generate_mesh
     :members:
