### 2020-09-08 876b8ca
don't use lambdas
