### 2020-09-08 dc0d9ad
turn on output for pytest in ci
