+++ tox.ini
@@ -16,5 +16,5 @@ deps =
     mpi4py
 whitelist_externals = {[mpi]mpiexec}
 commands =
-    {[mpi]mpiexec} -n 2 -mca btl ^openib pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "parallel"
+    {[mpi]mpiexec} -n 2 -mca btl ^openib pytest -s --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "parallel"
     pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term --cov-append  -m "serial"
