+++ tox.ini
@@ -16,4 +16,5 @@ deps =
     mpi4py
 whitelist_externals = {[mpi]mpiexec}
 commands =
-    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "serial"
+    {[mpi]mpiexec} -n 2 pytest -s --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term -m "parallel"
+    pytest --cov {envsitepackagesdir}/SeismicMesh --cov-report xml --cov-report term --cov-append  -m "serial"
