### 2020-09-09 e1f0847
removing __init__.py in test directory
