+++ tests/test_3dmesher_par.py
@@ -12,6 +12,8 @@ from SeismicMesh import (
 )
 
 comm = MPI.COMM_WORLD
+size = comm.Get_size()
+rank = comm.Get_rank()
 
 
 @pytest.mark.parallel
