+++ tests/test_2dmesher_par.py
@@ -7,11 +7,9 @@ from mpi4py import MPI
 from SeismicMesh import get_sizing_function_from_segy, generate_mesh, geometry
 
 comm = MPI.COMM_WORLD
-size = comm.Get_size()
-rank = comm.Get_rank()
 
 
-@pytest.mark.parallel
+@pytest.mark.parallel2
 def test_2dmesher_par():
 
     fname = os.path.join(os.path.dirname(__file__), "testing.segy")
