+++ pytest.ini
@@ -1,4 +1,5 @@
 [pytest]
 markers =
     serial: marks serial tests (deselect with '-m "not serial"')
-    parallel: marks parallel tests (deselect with 'm' "not parallel")
+    parallel2: marks parallel2 tests (deselect with 'm' "not parallel2")
+    parallel3: marks parallel3 tests (deselect with 'm' "not parallel3")
