### 2020-09-09 c18aa4f
updating docstrings
