+++ SeismicMesh/geometry/utils.py
@@ -45,7 +45,7 @@ class SignedDistanceFunctionGenerator:
         :param method: method used to combine subsection with bounding box. Currently only the default==union.
         :type method: string, optional
         :param narrow: only calculate SDF within a region close to the isocontour default=0.0 i.e., calc'ed everywhere
-        :type narrow: float64, optional
+        :type narrow: `float`, optional
         :param comm: MPI4py communicator default==None
         :type comm: MPI4py communicator object, optional
         :param flip: Whether to flip the signed distance function or not (default=False)
@@ -161,12 +161,12 @@ def calc_re_ratios(vertices, entities, dim=2):
     """Calculate radius edge ratios--mesh quality metric
 
     :param vertices: point coordinates of the mesh vertices'
-    :type vertices: numpy.ndarray[float64 x dim]
+    :type vertices: numpy.ndarray[`float` x dim]
     :param entities: mesh connectivity table
-    :type entities: numpy.ndarray[int x (dim + 1)]
+    :type entities: numpy.ndarray[`int` x (dim + 1)]
 
     :return: ce_ratios: array of radius-to-edge ratios
-    :rtype: ce_ratios: numpy.ndarray[float64 x 1]
+    :rtype: ce_ratios: numpy.ndarray[`float` x 1]
     """
     # circumradius/shortest edge length
     if dim == 2:
@@ -203,20 +203,20 @@ def remove_external_entities(vertices, entities, extent, dim=2):
     """Remove entities with all dim+1 vertices outside block.
 
     :param vertices: point coordinates of mesh
-    :type vertices: numpy.ndarray[float64 x dim]
+    :type vertices: numpy.ndarray[`float` x dim]
     :param entities: mesh connectivity
-    :type entities: numpy.ndarray[int x (dim + 1)]
+    :type entities: numpy.ndarray[`int` x (dim + 1)]
     :param extent: coords. of the local block extents
-    :type extent: numpy.ndarray[tuple(float64 x (2*dim))]
+    :type extent: numpy.ndarray[tuple(`float` x (2*dim))]
     :param dim: dimension of mesh
-    :type dim: int, optional
+    :type dim: `int`, optional
 
     :return: vertices_new: point coordinates of mesh w/ removed entities
-    :rtype: numpy.ndarray[float64 x dim]
+    :rtype: numpy.ndarray[`float` x dim]
     :return: entities_new: mesh connectivity w/ removed entities
-    :rtype: numpy.ndarray[int x (dim +1)]
+    :rtype: numpy.ndarray[`int` x (dim +1)]
     :return: jx: mapping from old point indexing to new point indexing
-    :rtype: numpy.ndarray[int x 1]
+    :rtype: numpy.ndarray[`int` x 1]
     """
 
     if dim == 2:
@@ -247,18 +247,18 @@ def vertex_to_entities(vertices, entities, dim=2):
     """Determine which elements are connected to which vertices.
 
     :param vertices: point coordinates of mesh vertices
-    :type vertices: numpy.ndarray[float64 x dim]
+    :type vertices: numpy.ndarray[`float` x dim]
     :param entities: mesh connectivity
-    :type entities: numpy.ndarray[int x (dim + 1)]
+    :type entities: numpy.ndarray[`int` x (dim + 1)]
     :param dim: dimension of mesh
-    :type dim: int, optional
+    :type dim: `int`, optional
 
     :return: vtoe: indices of entities connected to each vertex
-    :rtype: numpy.ndarray[int x 1]
[…]
