### 2020-09-09 c27768f
attempt at SizingFunction class
