+++ SeismicMesh/sizing/__init__.py
@@ -2,10 +2,12 @@ from .mesh_size_function import (
     get_sizing_function_from_segy,
     write_velocity_model,
     plot_sizing_function,
+    SizeFunction,
 )
 
 __all__ = [
     "get_sizing_function_from_segy",
     "write_velocity_model",
+    "SizeFunction",
     "plot_sizing_function",
 ]
