+++ SeismicMesh/sizing/mesh_size_function.py
@@ -22,6 +22,8 @@ from mpi4py import MPI
 from scipy import ndimage
 from scipy.interpolate import RegularGridInterpolator
 
+from .size_function import SizeFunction
+
 from .cpp import limgrad
 
 __all__ = [
@@ -42,7 +44,7 @@ opts = {
     "dt": 0.0,
     "cr_max": 1.0,
     "pad_style": "edge",
-    "domain_ext": 0.0,
+    "domain_pad": 0.0,
     "units": "m-s",
     "nz": None,
     "nx": None,
@@ -58,8 +60,8 @@ def get_sizing_function_from_segy(filename, bbox, comm=None, **kwargs):
         The name of a SEG-y or binary file containing a seismic velocity model
     :type filename: ``string``
     :param bbox:
-        Bounding box containing domain extents.
-    :type bbox: tuple with size (2*dim). For example, in 2D `(zmin, zmax, xmin, xmax)`
+        Bounding box containing domain extents of the velocity model contained in `filename`.
+    :type bbox: `tuple` with size (2*dim). For example, in 2D `(zmin, zmax, xmin, xmax)`
     :param \**kwargs:
         See below
 
@@ -77,33 +79,32 @@ def get_sizing_function_from_segy(filename, bbox, comm=None, **kwargs):
         * *grade* (``float``) --
           Maximum allowable variation in mesh size in decimal percent (default==0.0)
         * *space_order* (``int``) --
-          The polynomial order of the basis functions (default==1)
+          Simulation will be attempted with a mesh using the polynomial order `space_order` of the basis functions (default==1)
         * *dt* (``float``) --
           Theoretical maximum stable timestep in seconds given Courant number Cr (default==0.0 s)
         * *cr_max* (``float``) --
-            `dt` is theoretically numerically stable with this Courant number.
+            The mesh simulated with this `dt` has this maximum Courant number (default==1.0)
         * *pad_style* (``string``) --
              The method (`edge`, `linear_ramp`, `constant`) to pad velocity in the domain extension region (default==None)
-        * *domain_extension* (``float``) --
-             The width of the domain extension in -z, +x, -x, +y, -y directions (default==0.0 m).
-        * *units* (``float``) --
+        * *domain_pad* (``float``) --
+             The width of the domain pad in -z, +x, -x, +y, -y directions (default==0.0 m).
+        * *units* (``string``) --
              The units of the seismic velocity model (default='m-s')
         * *nz* (``int``) --
              REQUIRED FOR BINARY VELOCITY MODEL. The number of grid points in the z-direction in the velocity model.
         * *ny* (``int``) --
-             TREQUIRED FOR BINARY VELOCITY MODEL. The number of grid points in the y-direction in the velocity model.
+             REQUIRED FOR BINARY VELOCITY MODEL. The number of grid points in the y-direction in the velocity model.
         * *nx* (``int``) --
              REQUIRED FOR BINARY VELOCITY MODEL. The number of grid points in the x-direction in the velocity model.
         * *byte_order* (``string``) --
              REQUIRED FOR BINARY VELOCITY MODEL. The order of bytes in a 3D sesimic velocity model (`little` or `big`).
 
-    :return: cell_size_function: a function that takes a point and gets a size
-    :rtype: a callable Python function object.
-    :param bbox: bounding box containing domain extents (modified if `domain_extension` is in use).
-    :rtype: ``tuple``
+    :return: an object with a `obj.bbox` and an `obj.eval` method .
+    :rtype: a :class:`SizeFunction` object
 
     """
[…]

[…]
