+++ SeismicMesh/generation/mesh_generator.py
@@ -41,16 +41,13 @@ opts = {
 }
 
 
-def sliver_removal(points, signed_distance_function, h0, comm=None, **kwargs):
+def sliver_removal(points, domain, h0, comm=None, **kwargs):
     r"""Improve an existing 3D mesh by removing degenerate elements called
     commonly referred to as `slivers`.
 
     :param points:
         The name of a SEG-y or binary file containing a seismic velocity model
     :type filename: ``string``
-    :param signed_distance_function:
-        A callable function that takes a point and returns the signed nearest distance to the domain boundary Ω.
-    :type signed_distance_function: A callable function.
     :param h0:
         The minimum element size in the domain
     :type h0: `float`
@@ -92,9 +89,9 @@ def sliver_removal(points, signed_distance_function, h0, comm=None, **kwargs):
     if dim == 2:
         raise Exception("Mesh improvement currently on works in 3D")
 
-    if not callable(signed_distance_function):
-        raise ValueError("`signed_distance_function` is not a function!")
-    fd = signed_distance_function
+    if not callable(domain):
+        raise ValueError("`domain` is not a function!")
+    fd = domain
 
     if opts["max_iter"] < 0:
         raise ValueError("`max_iter` must be > 0")
@@ -203,15 +200,17 @@ def sliver_removal(points, signed_distance_function, h0, comm=None, **kwargs):
     return p, t
 
 
-def generate_mesh(cell_size, signed_distance_function, h0, comm=None, **kwargs):
+def generate_mesh(domain, cell_size, h0, comm=None, **kwargs):
     r"""Generate a 2D/3D triangulation using callbacks to a sizing function `cell_size`
-       and signed distance function `signed_distance_function`
+       and signed distance function :class:`domain`
 
+    :param domain:
+        Either a :class:`geometry` or a function that takes a point and returns the
+        signed nearest distance to the domain boundary Ω
+    :type domain: A :class:`geometry` object (e.g., Rectangle, Cube, or Circle)
     :param cell_size:
         Either a :class:`size_function` object or a function that can evalulate a point and return a mesh size.
-    :param signed_distance_function:
-        A callable function that takes a point and returns the signed nearest distance to the domain boundary Ω.
-    :type signed_distance_function: A callable function.
+    :type cell_size:
     :param h0:
         The minimum element size in the domain
     :type h0: `float`
@@ -250,17 +249,13 @@ def generate_mesh(cell_size, signed_distance_function, h0, comm=None, **kwargs):
     opts.update(kwargs)
     _parse_kwargs(kwargs)
 
-    # unpack size function
-    if isinstance(cell_size, sizing.size_function):
-        bbox = cell_size.bbox
-        fh = cell_size.eval
-    elif callable(cell_size):
-        bbox = opts["bbox"]
-        fh = cell_size
-    else:
-        raise ValueError(
-            "`cell_size` must either be a function or a `cell_size` object"
-        )
+    # unpack domain
+    fd, bbox0 = _unpack_domain(domain)
+
+    fh, bbox1 = _unpack_sizing(cell_size)
+
+    # take maxmin of boxes
+    bbox = _minmax(bbox0, bbox1)
 
     if not isinstance(bbox, tuple):
         raise ValueError("`bbox` must be a tuple")
@@ -276,18 +271,6 @@ def generate_mesh(cell_size, signed_distance_function, h0, comm=None, **kwargs):
     if opts["max_iter"] < 0:
         raise ValueError("`max_iter` must be > 0")
     max_iter = opts["max_iter"]
-
-    # check signed_distance_function
[…]

[…]
