+++ tests/test_3dmesher_domain_extension.py
@@ -6,7 +6,7 @@ import pytest
 from SeismicMesh import (
     generate_mesh,
     sliver_removal,
-    geometry,
+    Cube,
     get_sizing_function_from_segy,
     write_velocity_model,
 )
@@ -29,7 +29,8 @@ def test_3dmesher_domain_extension(style_answer):
     hmin = 150
     grade = 0.005
     bbox = (-2e3, 0.0, 0.0, 1e3, 0, 1e3)
-    ef, bbox = get_sizing_function_from_segy(
+    cube = Cube(bbox)
+    ef = get_sizing_function_from_segy(
         fname,
         bbox,
         grade=grade,
@@ -41,7 +42,7 @@ def test_3dmesher_domain_extension(style_answer):
         nx=10,
         ny=10,
         byte_order="little",
-        domain_ext=200,
+        domain_pad=200,
         pad_style=style,
     )
 
@@ -49,20 +50,14 @@ def test_3dmesher_domain_extension(style_answer):
         fname, nz=20, nx=10, ny=10, byte_order="little", ofname="testing"
     )
 
-    def cube(p):
-        return geometry.dblock(p, *bbox)
-
     points, cells = generate_mesh(
-        bbox=bbox,
-        signed_distance_function=cube,
+        domain=cube,
         cell_size=ef,
         h0=hmin,
         perform_checks=False,
     )
 
-    points, cells = sliver_removal(
-        points=points, bbox=bbox, signed_distance_function=cube, h0=hmin
-    )
+    points, cells = sliver_removal(points=points, domain=cube, cell_size=ef, h0=hmin)
     print(len(points), len(cells))
     allclose([len(points), len(cells)], answer, atol=100)
 
