+++ SeismicMesh/geometry/signed_distance_functions.py
@@ -19,7 +19,7 @@ class Rectangle:
         self.y2 = bbox[3]
 
     def eval(self, x):
-        return _drectangle(x, self.x1, self.x2, self.y1, self.y2)
+        return drectangle(x, self.x1, self.x2, self.y1, self.y2)
 
 
 class Cube:
@@ -32,7 +32,7 @@ class Cube:
         self.z2 = bbox[5]
 
     def eval(self, x):
-        return _dblock(x, self.x1, self.x2, self.y1, self.y2, self.z1, self.z2)
+        return dblock(x, self.x1, self.x2, self.y1, self.y2, self.z1, self.z2)
 
 
 def _dcircle(p, xc, yc, r):
@@ -40,7 +40,7 @@ def _dcircle(p, xc, yc, r):
     return np.sqrt(((p - np.array([xc, yc])) ** 2).sum(-1)) - r
 
 
-def _drectangle(p, x1, x2, y1, y2):
+def drectangle(p, x1, x2, y1, y2):
     min = np.minimum
     """Signed distance function for rectangle with corners (x1,y1), (x2,y1),
     (x1,y2), (x2,y2).
@@ -49,7 +49,7 @@ def _drectangle(p, x1, x2, y1, y2):
     return -min(min(min(-y1 + p[:, 1], y2 - p[:, 1]), -x1 + p[:, 0]), x2 - p[:, 0])
 
 
-def _dblock(p, x1, x2, y1, y2, z1, z2):
+def dblock(p, x1, x2, y1, y2, z1, z2):
     min = np.minimum
     return -min(
         min(
