### 2020-09-09 4a4af77
updating docstrings
