+++ examples/__init__.py
@@ -1,3 +1,0 @@
-import example_2D
-
-__all__ = ["example_2D"]
