+++ docs/source/usrman/tutorial.rst
@@ -81,7 +81,7 @@ This seismic velocity model is passed to the *MeshSizeFunction* class along with
     from SeismicMesh import get_sizing_function_from_segy
 
     # Construct a mesh sizing function from a seismic velocity model
-    ef, bbox = get_sizing_function_from_segy(fname, bbox, other-args-go-here,...)
+    ef = get_sizing_function_from_segy(fname, bbox, other-args-go-here,...)
 
 * The user specifies the filename `fname` of the seismic velocity model (e.g., either SEG-y or binary)
 
@@ -99,28 +99,23 @@ This seismic velocity model is passed to the *MeshSizeFunction* class along with
 
 * If the user wants to define a mesh with an irregular boundary (other than a cube), then they will have to use the :class:`SignedDistanceFunctionGenerator` and pass a velocity interval range that represents the region of the domain they want meshed. See the section immediately below for instructions on how to create and use a custom signed distance function from a seismic velocity model.
 
-Creating your own signed distance functions
------------------------------------------------
+Geometry
+---------
 
-Using the function `signed_distance_function_generator`, the user can threshold their velocity model to produce a signed distance function which can then be used to mesh with. For example, in the case of an irregular free surface boundary in the 2D Foothills model, a SDF can be created by thresholding a modified velocity model. Specifically, we mesh the region of the domain with a P-wave velocity greater than 4,000 m/s::
+*SeismicMesh* can mesh any domain defined by a signed distance function. We provide three basic domain shapes: a Rectangle, a Cube, or a Circle.
 
-     # Build a signed distance function from the seismic velocity model
-     # Some pockets of velocity < 4000 exist, fill those in.
-     vp2 = ef.vp.copy()
-     vp2 = np.where(vp2 < 4000, 4001, vp2)
-     SDF = SdfGen(
-         bbox=bbox, field=vp2, min_threshold=4000.0, gridspacing=(10.0, 15.0),
-     ).SDF
+The user can build a rectangular 2D domain like so::
 
-.. note :: We modify the velocity model so that some small pockets of area with velocity less than 4,000 m/s don't appear as holes in the mesh when we use 4,000 m/s to create the signed distance function.
+    from SeismicMesh import Rectangle, Cube, Circle
 
-The output of the call to the ``SdfGen`` gives a handle to a function object, which can be passed to the mesh generator as detailed below in the section describing how to call the mesh generator.
+    rectangle = Rectangle(bbox)
+    cube = Cube(bbox)
+    circle = Circle(circle)
 
-Images shown below are the seismic P-wave velocity model for the foothills example and the resulting SDF from executing the code directly above. In this second image below, the color yellow indicates the region to-be-meshed and the greenish color indicates the region outside of the domain.
+.. note::
+    A good reference for various signed distance functions can be found [here](https://www.iquilezles.org/www/articles/distfunctions/distfunctions.htm)
 
-.. image:: Foothills.png
 
-.. image:: ExampleOfSDF.png
 
 Mesh size function
 -------------------------------------------
@@ -141,7 +136,7 @@ The highest frequency of the source wavelet :math:`f_{max}` and the smallest val
 The user is able to specify the number of vertices per wavelength :math:`\alpha_{wl}` the peak source frequency :math:`f_{max}`. This sizing heuristic also  can be used to take into account varying polynomial orders for finite elements. For instance if using quadratic P=2 elements, :math:`\alpha_{wl}` can be safely be set to 5 to avoid excessive dispersion and dissipation otherwise that would occur with P=1 elements::
 
[…]
