### 2020-09-09 6a59637
API overhaul (#52)
* Complete overhaul to API hopefully resolving the issues brought up in #50 #49 #48
