+++ SeismicMesh/geometry/signed_distance_functions.py
@@ -1,6 +1,45 @@
 import numpy as np
 
 
+class Circle:
+    def __init__(self, xc, yc, r):
+        self.xc = xc
+        self.yc = yc
+        self.r = r
+
+    def eval(self, x):
+        return _dcircle(x, self.xc, self.yc, self.r)
+
+
+class Rectangle:
+    def __init__(self, bbox):
+        self.x1 = bbox[0]
+        self.x2 = bbox[1]
+        self.y1 = bbox[2]
+        self.y2 = bbox[3]
+
+    def eval(self, x):
+        return drectangle(x, self.x1, self.x2, self.y1, self.y2)
+
+
+class Cube:
+    def __init__(self, bbox):
+        self.x1 = bbox[0]
+        self.x2 = bbox[1]
+        self.y1 = bbox[2]
+        self.y2 = bbox[3]
+        self.z1 = bbox[4]
+        self.z2 = bbox[5]
+
+    def eval(self, x):
+        return dblock(x, self.x1, self.x2, self.y1, self.y2, self.z1, self.z2)
+
+
+def _dcircle(p, xc, yc, r):
+    """Signed distance to circle centered at xc, yc with radius r."""
+    return np.sqrt(((p - np.array([xc, yc])) ** 2).sum(-1)) - r
+
+
 def drectangle(p, x1, x2, y1, y2):
     min = np.minimum
     """Signed distance function for rectangle with corners (x1,y1), (x2,y1),
