+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = SeismicMesh
-version = 2.0.4
+version = 3.0.0
 url = https://github.com/krober10nd/SeismicMesh
 author = Keith Roberts
 email = keithrbt0@gmail.com
@@ -33,6 +33,5 @@ install_requires =
         h5py
         matplotlib
         mpi4py
-        scikit-fmm
         pybind11 >= 2.2
 python_requires = >=3.0
