### 2020-09-09 36e9f30
various doc fixes
