+++ docs/source/usrman/overview.rst
@@ -2,7 +2,7 @@ Overview
 ========
 
 
-This software aims to create end-to-end workflows (e.g., from seismic velocity model to simulation ready mesh) to build quality two- and three-dimensional (2D and 3D) unstructured triangular and tetrahedral meshes for seismic domains. The main advantage of triangular meshes over cubes or rectangles is their ability to cost-effectively resolve variable material properties. These meshes are suitable for acoustic and elastic numerical wave propagators. A focus is placed on parallel unstructured mesh generation capabilities. A simple application program interfaces (API) written in Python enables the user to call parallel meshing algorithms that can be scaled up on distributed memory clusters.
+This software aims to create end-to-end workflows (e.g., from seismic velocity model to simulation ready mesh) to build quality two- and three-dimensional (2D and 3D) unstructured triangular and tetrahedral meshes for seismic domains. The main advantage of triangular meshes over cubes or rectangles is their ability to cost-effectively resolve variable material properties. These genereated meshes are suitable for acoustic and elastic numerical wave propagators and a focus is placed on parallel unstructured mesh generation capabilities. A simple application program interfaces (API) written in Python enables the user to call parallel meshing algorithms that can be scaled up on distributed memory clusters.
 
 *SeismicMesh* is currently being used to generate simplical meshes in 2D and 3D for acoustic and elastic wave propagators written using a Domain Specific Language called *Firedrake* [firedrake]_. These type of numerical simulations are used in Full Waveform Inversion, Reverse Time Migration, and Time Travel Tomography applications.
 
@@ -49,16 +49,15 @@ Inputs
 
 Seismic velocity model
 ^^^^^^^^^^^^^^^^^^^^^^^^
-.. note ::
-    The only required input file to generate a mesh is a velocity model defined as a numpy.ndarray grid containing velocity data at each grid point.
 
-* A velocity model is a regular Cartesian grid of values with potentially varying grid spacing in each dimension. *SeismicMesh* expects the
-  velocity model to be a numpy.ndarray and orientated so z is the first dimension, x is the second, (and y if in 3D is the 3rd dimension of the array).
-  Please make sure the velocity model grid is orientated in this format otherwise you will experience undeseriable results!
+A seismic velocity model is defined on an axis-aligned regular Cartesian grid in either 2D or 3D containing scalar values (typically the P-wave velocity speed at each grid point).
 
+Currently, *SeismicMesh* can read seismic velocity models from SEG-y files or in binary format.
 
 Signed distance function
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Given a point :math:`x`, the signed distance function returns the :math:`d` distance to the boundary of the domain :math:`∂ \Omega`.
 
 Let :math:`\Omega ⊂ D ∈ R^N` be the domain in :math:`N` dimensions. The boundary of the domain to-be-meshed is represented as the 0-level set of a continuous function:
 
@@ -68,17 +67,14 @@ such that:
 
 :math:`\Omega := {x ∈ D, φ(x) < 0}`
 
-where :math:`φ : D × R+ → R` is Lipschitz continuous and called the level set function. If we assume :math:`|∇φ(·)| = 0` on the set :math:`{x ∈ D, φ(x) = 0}`, then we have :math:`∂ \Omega = {x ∈ D, φ(x) = 0}` i.e., the boundary :math:`∂ \Omega` is the 0-level set of :math:`φ(·)`. The property that :math:`|∇φ(·)| = 0` is satisfied if :math:`φ(·)` is a signed distance function. Given a point :math:`x`, the signed distance function returns the :math:`d` distance to the :math:`∂ \Omega`.
-
[…]
