+++ docs/source/usrman/tutorial.rst
@@ -4,13 +4,11 @@
 
     Under construction. Contributions very welcome!
 
-Tutorial
+Basics
 ========
 
 *SeismicMesh* supports the generation of both 2D and 3D meshes in
-either serial or parallel from seismic velocity models. It also supports the generation of
-complex mesh sizing functions that are relevant to seismological applications.
-
+either serial or parallel from seismic velocity models.
 
 Here I show how to build meshes from sizing functions created with the software and explain what the options mean. The API for serial/parallel and 2D/3D is identical.
 
@@ -35,48 +33,34 @@ where `N` is the number of cores (e.g., 2,3,4 etc.)
     Oversubscribing the mesh generation problem to too many cores will surely lead to problems and slow downs. In general, keeping the minimum number of vertices per rank to between 20-50k/rank results in optimal performance.
 
 
-Data for examples
+Example data
 -------------------
 
 .. note::
     Users should create a directory called `velocity_models` and place their seismic velocity models there.
 
 
-Data for this 2D tutorial can be downloaded for the BP2004 benchmark model::
+A 2D model (BP2004)::
 
     wget http://s3.amazonaws.com/open.source.geoscience/open_data/bpvelanal2004/vel_z6.25m_x12.5m_exact.segy.gz
 
-and for the irregular free surface example here::
-
-    wget http://s3.amazonaws.com/open.source.geoscience/open_data/bpmodel94/velocity.segy.gz
-
-For more details about these two models, see the following link::
-
-    https://wiki.seg.org/wiki/2004_BP_velocity_estimation_benchmark_model
-    https://wiki.seg.org/wiki/1994_BP_migration_from_topography
-
-The original data for the 3D example can be downloaded here::
+A 3D model (EAGE Salt)::
 
     https://s3.amazonaws.com/open.source.geoscience/open_data/seg_eage_models_cd/Salt_Model_3D.tar.gz
-
-
-For more details about this 3D data, see the following link: https://wiki.seg.org/wiki/SEG/EAGE_Salt_and_Overthrust_Models
-See example_3D.py for how to read in the binary into a Numpy array.
 
 
 File I/O and visualization of meshes
 --------------------------------------
 
-Meshes are written to disk in a variety of formats using the Python package `MeshIO` (https://pypi.org/project/meshio/). Note that *SeismicMesh* makes the assumption that the first dimension is `z` and the second is `x` while the third is `y`. This is done in this way since 2D seismological simulations take place in the z-x plane and 3D in the z-x-y plane. As a result, the meshes when loaded into visualization software will appear rotated 90 degrees. For visualization, we can output in the vtk format using MeshIO (as shown in the examples) and then load the vtk file into Paraview.
+Meshes can be written to disk in a variety of formats using the Python package `meshio` (https://pypi.org/project/meshio/).
+
+.. warning::
+    Note that *SeismicMesh* makes the assumption that the first dimension is `z` and the second is `x` while the third is `y`. This is done in this way since 2D seismological simulations take place in the z-x plane and 3D in the z-x-y plane. As a result, the meshes when loaded into visualization software will appear rotated 90 degrees. For visualization, we can output in the vtk format using MeshIO (as shown in the examples) and then load the vtk file into Paraview.
 
 Some things to know
 ---------------------
 
-.. warning ::
-    In order to use these mesh sizing functions, it is assumed you have a seismic velocity model
-    defined on a structured grid as was mentioned in the overview section.
-
-This seismic velocity model is passed to the *MeshSizeFunction* class along with the domain extents ::
+This seismic velocity model is passed to the `get_sizing_function_from_segy` along with the domain extents ::
 
[…]
