### 2020-09-10 211eb8a
fixes to documentation (#55)
* various doc fixes

* updating MANIFEST to correctly include tests, README and license in source dists.
