+++ docs/source/usrman/install.rst
@@ -10,16 +10,15 @@ build *SeismicMesh*:
 * Python >= 3.0
 .. note ::
     The file ``requirements.txt`` in the main directory indicates all the Python dependencies and their respective version numbers. These packages should be installed at compile time by setuptools
-.. note ::
-    Some users have experienced  problems with the `skfmm-fmm` Python package not being found. If this occurs, try uninstalling and then reinstalling this package after attempting installation of SeismicMesh.
+
 .. note ::
     On some Linux systems, users may have to resort to `apt install python3-segyio` to installing segyio on their systems.
 
-* Pybind11 >= 2.5
+* pybind11 >= 2.5
 
 * C++ compiler (GNU or Intel) with support for std++14 or newer.
 
-* Cmake >= 3.0
+* cmake >= 3.0
 
 * CGAL >= 5.0.0 which requires:
 
@@ -33,7 +32,7 @@ build *SeismicMesh*:
     CGAL requires Boost, MPFR and GMP. These packages may already be installed on your standard Linux box.
 
 .. warning ::
-    Make sure your package manager is downloading CGAL >= 5.0 otherwise you will not be able to install *SeismicMesh
+    Make sure your package manager is downloading CGAL >= 5.0 otherwise you will not be able to install *SeismicMesh*!
 
 
 
@@ -58,7 +57,7 @@ Testing is accomplished with `tox`. The `tox` package can be installed like so::
 
     pip install tox
 
-To test the installation, serial and parallel capabilites, you can use `tox` from the top directory of the package::
+To test the installation, serial and parallel capabilities, you can use `tox` from the top directory of the package::
 
 $ tox
 
