### 2020-09-10 6cd1eef
spell checking docs (#56)
* spell checking docs with updates to API
