### 2020-09-11 1457cbe
fixsome  lgtm alerts
