### 2020-09-11 a1e0425
Update blocker.py
