### 2020-09-11 2f98797
fix some lgtm alerts (#58)
* Fixing some  lgtm alerts
