### 2020-09-15 a6b72af
Fixes (#60)
* correction for installation of pybind11

* typo in paper
