+++ README.md
@@ -37,7 +37,7 @@ Installation
 For installation, SeismicMesh needs [CGAL](https://www.cgal.org/) and
 [pybind11](https://github.com/pybind/pybind11):
 
-    sudo apt install libcgal-dev python-pybind11
+    sudo apt install libcgal-dev python3-pybind11
 
 After that, SeismicMesh can be installed from the Python Package Index
 ([pypi](https://pypi.org/project/SeismicMesh/)), so with:
