### 2020-09-16 4efa6a4
adjust SDF to reflect domain_pad
