+++ SeismicMesh/generation/mesh_generator.py
@@ -101,6 +101,10 @@ def sliver_removal(points, domain, cell_size, h0, comm=None, **kwargs):
     # take maxmin of boxes
     bbox = _minmax(bbox0, bbox1)
 
+    # rebuild the Rectangle or cube if domain padding
+    if bbox0 != bbox1:
+        fd = geometry.Cube(bbox).eval
+
     if not isinstance(bbox, tuple):
         raise ValueError("`bbox` must be a tuple")
 
@@ -271,6 +275,16 @@ def generate_mesh(domain, cell_size, h0, comm=None, **kwargs):
 
     # check bbox shape
     dim = int(len(bbox) / 2)
+
+    # rebuild the Rectangle or Cube if domain padding
+    if bbox0 != bbox1:
+        if dim == 2:
+            tmp = geometry.Rectangle(bbox)
+        elif dim == 3:
+            tmp = geometry.Cube(bbox)
+
+        fd = tmp.eval
+
     bbox = np.array(bbox).reshape(-1, 2)
 
     # check h0
