### 2020-09-16 ab35c2d
perform quick checks
