### 2020-09-16 a7eec18
some doc fixes
