### 2020-09-16 d02191b
SDF bug fix and write_velocity_model enhancement (#62)
* Bug fix: adjust SDF in `generate_mesh` and `sliver_removal` to reflect user option `domain_pad`

* Enhancement to `write_velocity_model` to consider domain padding.

* Some document formatting fixes and typos.
