### 2020-09-17 4e69fee
screen outputs for sliver removal and README example
