### 2020-09-17 22a4f68
minor improvements to dihedral angle bound   (#63)
* Screen feedback./output when running `remove_slivers`
* User ability to select dihedral angle bounds. 
*New example on README.
