### 2020-09-21 #2687 @whedon
[REVIEW]: SeismicMesh: Triangular meshing for seismology
**Submitting author:** @krober10nd (<a href="http://orcid.org/0000-0002-8470-0359">Keith Roberts</a>)
**Repository:** <a href="https://github.com/krober10nd/SeismicMesh" target ="_blank">https://github.com/krober10nd/SeismicMesh</a>
**Version:** v3.3.1
