### 2020-09-21 #2592 @meg-simula
@arfon Could you have a look at this case? I asked whedon to add a reviewer, and then start the review, but the review was started before the second reviewer was added. I tried closing that review issue, and starting a new one with the two reviewers, but now it seems that whedon is not too keen to. Any advice?
