### 2020-09-21 a2a117f
Update setup.cfg
