### 2020-09-23 #2592 @arfon
> @arfon Could you have a look at this case? I asked whedon to add a reviewer, and then start the review, but the review was started before the second reviewer was added. I tried closing that review issue, and starting a new one with the two reviewers, but now it seems that whedon is not too keen to. Any advice?

There's no automated way to do this right now sorry. The 'fix' is to add the second reviewer on the actual review issue (#2687) and add a second checklist for them over there. I'll go ahead and do this for you now...
