### 2020-09-28 33c65c5
re-location of termination logic
