+++ SeismicMesh/geometry/cpp/fast_geometry.cpp
@@ -2,8 +2,64 @@
 #include <pybind11/numpy.h>
 #include <pybind11/stl.h>
 #include <vector>
+#include <set>
+#include <algorithm>
+#include <tuple>
 
 namespace py = pybind11;
+
+using Ints = std::tuple<int,int>;
+
+template<typename T>
+std::vector<T> constUniqueVectorSort(std::vector<T> v){
+    std::sort(v.begin(),v.end());
+    auto iter = std::unique(v.begin(),v.end());
+    return std::vector<T>(v.begin(),iter);
+}
+
+template<typename T>
+std::vector<int> unpack(const std::vector<T> &v){
+    std::vector<int> out;
+    out.reserve(v.size()*2);
+    for(auto &a : v){
+        out.push_back(std::get<0>(a));
+        out.push_back(std::get<1>(a));
+    }
+    return out;
+}
+
+py::array unique_edges(
+    py::array_t<int, py::array::c_style | py::array::forcecast> edges){
+
+  std::vector<Ints> tl;
+
+  std::vector<int> cedges(edges.size());
+  std::memcpy(cedges.data(), edges.data(), edges.size() * sizeof(int));
+
+  tl.reserve(cedges.size());
+  for(size_t i=0;i<cedges.size();i+=2)
+     tl.emplace_back(cedges[i],cedges[i+1]);
+
+  auto u_edges = unpack(constUniqueVectorSort<Ints>(tl));
+
+  int num_edges = u_edges.size();
+
+  ssize_t sint = sizeof(int);
+  std::vector<ssize_t> shape = {num_edges/2, 2};
+  std::vector<ssize_t> strides = {sint * 2, sint};
+
+  // only keep unique part
+  // return 2-D NumPy array
+  return py::array(
+      py::buffer_info(u_edges.data(), /* data as contiguous array  */
+                      sizeof(int),           /* size of one scalar        */
+                      py::format_descriptor<int>::format(), /* data type */
+                      2,      /* number of dimensions      */
+                      shape,  /* shape of the matrix       */
+                      strides /* strides for each axis     */
+                      ));
+  }
+
 
 double l2_norm(const std::array<double, 3> &u) {
   double accum = 0.;
@@ -385,6 +441,7 @@ py::array calc_circumsphere_grad(
 }
 
 PYBIND11_MODULE(fast_geometry, m) {
+  m.def("unique_edges", &unique_edges);
   m.def("calc_volume_grad", &calc_volume_grad);
   m.def("calc_circumsphere_grad", &calc_circumsphere_grad);
   m.def("calc_dihedral_angles", &calc_dihedral_angles);
