### 2020-09-29 118899a
another try at a fast unique_edges
