### 2020-09-29 bc3bab0
some improvements to bottlenecks
