+++ benchmarks/README.md
@@ -1,6 +1,6 @@
 Benchmarking
 ------------
-Here we compare *SeismicMesh* against well-established existing mesh generation approaches such as [CGAL](https://doc.cgal.org/latest/Mesh_3/) and [gmsh](https://gmsh.info/doc/texinfo/gmsh.html). Specifically:
+Here we compare `SeismicMesh` against well-established existing mesh generation approaches such as [CGAL](https://doc.cgal.org/latest/Mesh_3/) and [gmsh](https://gmsh.info/doc/texinfo/gmsh.html). Specifically:
 
     * a comparison in mesh creation speed in terms of wall-clock time.
     * a comparison in cell quality.
@@ -16,31 +16,57 @@ The following set of benchmark problems are available in `benchmark.py`:
     benchmark_cuboid: # a simple box-type geometry with linearly varying mesh resolution
     benchmark_sphere: # a sphere with a ring of higher resolution in the center
 
-Run `python benchmark_cuboid.py --method METHODNAME` with `cgal` using [pygalmesh](https://github.com/nschloe/pygalmesh) to mesh and `sm` using SeismicMesh to mesh with.
+Run `python benchmark_cuboid.py --method METHODNAME` with `cgal` using [pygalmesh](https://github.com/nschloe/pygalmesh) to mesh and `sm` using `SeismicMesh` to mesh with.
 
 Expected output
 ---------------
-Using  [termplotlib](https://github.com/nschloe/termplotlib), the benchmarks produce the following output:
+
+Using [termplotlib](https://github.com/nschloe/termplotlib) and [meshplex](https://github.com/nschloe/meshplex) to calculate some mesh statistics, the benchmarks produce histograms of [dihedral angles](https://en.wikipedia.org/wiki/Dihedral_angle#:~:text=A%20dihedral%20angle%20is%20the,line%20as%20a%20common%20edge) in the cells and the cell quality. For example running `benchmark_sphere.py` produces the following output:
 
 ```
-┌───────────────────────┬──────────────────────────────────────────┬────────────────────────────────────┬──────────────────────┐
-│                       │                                          │                                    │                      │
-│  SeismicMesh          │  Mesh creation time (seconds):     8.27  │  Number of vertices:        11510  │                      │
-│                       │                                          │  Number of cells:        62017     │                      │
-│                       │                                          │                                    │                      │
-├───────────────────────┼──────────────────────────────────────────┼────────────────────────────────────┼──────────────────────┤
-│                       │                                          │                                    │                      │
-│  dihedral angles      │  min angle:      10.584                  │  mesh quality metric               │  min quality: 0.137  │
-│                █▇     │  avg angle:      42.594                  │                            ▇█▅     │  avg quality: 0.801  │
-│               ███▃    │  max angle:      56.723                  │                           ████     │  max quality: 0.999  │
-│              ▄████    │  std dev angle:   8.393                  │                          █████▇    │                      │
-│              █████    │                                          │                         ▆██████    │                      │
-│             ▇█████    │                                          │                        ▄███████    │                      │
-│            ▂██████▂   │                                          │                       ▄████████▇   │                      │
-│           ▂████████   │                                          │                      ▄██████████   │                      │
-│          ▁█████████   │                                          │                    ▃▆███████████   │                      │
-│        ▂▅██████████   │                                          │                ▁▃▄▆█████████████▃  │                      │
-│  ▁▃▄▅▆█████████████▄  │                                          │  ▁▁▂▂▃▃▃▃▄▅▅▅▆▇██████████████████  │                      │
-│                       │                                          │                                    │                      │
-└───────────────────────┴──────────────────────────────────────────┴────────────────────────────────────┴──────────────────────┘
[…]
