+++ benchmarks/README.md
@@ -69,4 +69,4 @@ Using [termplotlib](https://github.com/nschloe/termplotlib) and [meshplex](https
 Notes
 -----
 * For CGAL's mesh generator in 3D, all default quality options are assumed. A `cell_size` function is passed to create variable resolution.
-* SeismicMesh is run here in serial mode. It's important to note however that a significant speed-up can be achieved for moderate to large problems using it [parallel capabilities](https://seismicmesh.readthedocs.io/en/par3d/tutorial.html#basics).
+* SeismicMesh is run here in serial mode. It's important to note however that a significant speed-up can be achieved for moderate to large problems using its [parallel capabilities](https://seismicmesh.readthedocs.io/en/par3d/tutorial.html#basics).
