### 2020-09-30 5601a09
adding image to benchmarks
