+++ benchmarks/README.md
@@ -21,9 +21,11 @@ Run `python benchmark_cuboid.py --method METHODNAME` with `cgal` using [pygalmes
 Expected output
 ---------------
 
-Using [termplotlib](https://github.com/nschloe/termplotlib) and [meshplex](https://github.com/nschloe/meshplex) to calculate some mesh statistics, the benchmarks produce histograms of [dihedral angles](https://en.wikipedia.org/wiki/Dihedral_angle#:~:text=A%20dihedral%20angle%20is%20the,line%20as%20a%20common%20edge) in the cells and the cell quality. For example running `benchmark_sphere.py` produces the following output:
+Using [termplotlib](https://github.com/nschloe/termplotlib) and [meshplex](https://github.com/nschloe/meshplex) to calculate some mesh statistics, the benchmarks produce histograms of [dihedral angles](https://en.wikipedia.org/wiki/Dihedral_angle#:~:text=A%20dihedral%20angle%20is%20the,line%20as%20a%20common%20edge) in the cells and the cell quality.
 
-```
+
+For example running `benchmark_sphere.py` produces the following output:
+
 ┌───────────────────────────┬────────────────────────────────────────────────────┬────────────────────────────────────────────┬──────────────────────┐
 │                           │                                                    │                                            │                      │
 │  CGAL                     │  Mesh creation time (seconds):    15.84            │  Number of vertices:       108407          │                      │
@@ -64,7 +66,51 @@ Using [termplotlib](https://github.com/nschloe/termplotlib) and [meshplex](https
 │      ▁▁▁▂▂▂▃▄▅█▉████████  │                                                    │           ▁▁▁▁▁▁▁▁▂▂▂▂▃▃▄▄▅▆█████████████  │                      │
 │                           │                                                    │                                            │                      │
 └───────────────────────────┴────────────────────────────────────────────────────┴────────────────────────────────────────────┴──────────────────────┘
-```
+
+and for `benchmark_cuboid.py`....
+
+┌───────────────────────────┬────────────────────────────────────────────────────┬────────────────────────────────────────────┬──────────────────────┐
+│                           │                                                    │                                            │                      │
+│  CGAL                     │  Mesh creation time (seconds):    11.47            │  Number of vertices:        32651          │                      │
+│                           │  Mesh creation speed (vertices/seconds):  2846.85  │  Number of cells:          180996          │                      │
+│                           │                                                    │                                            │                      │
+├───────────────────────────┼────────────────────────────────────────────────────┼────────────────────────────────────────────┼──────────────────────┤
+│                           │                                                    │                                            │                      │
+│  dihedral angles          │  min angle:      12.778                            │  mesh quality metric                       │  min quality: 0.198  │
+│                    ▅█     │  avg angle:      44.652                            │                                   ▃█▅      │  avg quality: 0.813  │
+│                    ██▂    │  max angle:      56.977                            │                                  ▂███▃     │  max quality: 0.999  │
+│                   ████    │  std dev angle:   6.610                            │                                  █████     │                      │
+│                   ████    │                                                    │                                 ▇█████▂    │                      │
+│                  █████    │                                                    │                                ▄███████    │                      │
+│                 ▁█████▁   │                                                    │                               ▃████████    │                      │
+│                 ███████   │                                                    │                              ▄█████████▅   │                      │
+│                ▆███████   │                                                    │                            ▁▅███████████   │                      │
+│              ▂▆▉███████   │                                                    │                         ▁▂▅█████████████   │                      │
+│         ▁▂▃▅▇██▉███████▃  │                                                    │               ▁▁▁▂▂▃▃▄▆▆█▉███████████████  │                      │
+│                           │                                                    │                                            │                      │
+└───────────────────────────┴────────────────────────────────────────────────────┴────────────────────────────────────────────┴──────────────────────┘
[…]
