### 2020-09-30 bbb61eb
cleaning up the benchmark readme
