+++ benchmarks/README.md
@@ -1,118 +1,34 @@
 Performance
 ------------
-Here we compare `SeismicMesh` against well-established existing mesh generation approaches such as [CGAL](https://doc.cgal.org/latest/Mesh_3/) and [gmsh](https://gmsh.info/doc/texinfo/gmsh.html). Specifically:
+Here we compare `SeismicMesh` against well-established existing mesh generation tools such as [cgal](https://doc.cgal.org/latest/Mesh_3/) and [gmsh](https://gmsh.info/doc/texinfo/gmsh.html). Specifically:
 
-    * a comparison in mesh creation speed in terms of wall-clock time.
+    * a comparison in mesh creation speed in terms of wall-clock time and throughput.
     * a comparison in cell quality.
     * a comparison of dihedral angle distributions.
 
-where cell quality is defined as d * circumcircle_radius / incircle_radius (where d is 2 for triangles and 3 for tetrahedra). The value is between 0 and 1, where 1 is a perfectly symmetrical simplex.
+where cell quality is defined as d * circumcircle_radius / incircle_radius (where d is 3 for tetrahedra). The value is between 0 and 1, where 1 is a perfectly symmetrical simplex.
 
 Benchmarks
 ----------
 
-The following set of benchmark problems are available in `benchmark.py`:
+The following set of benchmark problems are available:
 
-    benchmark_cuboid: # a simple box-type geometry with linearly varying mesh resolution
-    benchmark_sphere: # a sphere with a ring of higher resolution in the center
+    benchmark_cuboid: # a simple box-type geometry with linearly varying mesh resolution along the x-axis.
+    benchmark_sphere: # a sphere with a ring of higher resolution near the center.
 
-Run `python benchmark_cuboid.py --method METHODNAME` with `cgal` using [pygalmesh](https://github.com/nschloe/pygalmesh) to mesh and `sm` using `SeismicMesh` to mesh with.
+Run `python benchmark_cuboid.py` to run all benchmarks for a particular domain (e.g., cuboid). Run `python benchmark_cuboid.py --method METHODNAME` to select either `cgal` using [pygalmesh](https://github.com/nschloe/pygalmesh) or `sm` to use `SeismicMesh`.
 
-Expected output
+Results
 ---------------
 
-Using [termplotlib](https://github.com/nschloe/termplotlib) and [meshplex](https://github.com/nschloe/meshplex) to calculate some mesh statistics, the benchmarks produce histograms of [dihedral angles](https://en.wikipedia.org/wiki/Dihedral_angle#:~:text=A%20dihedral%20angle%20is%20the,line%20as%20a%20common%20edge) in the cells and the cell quality.
+Using [termplotlib](https://github.com/nschloe/termplotlib) and [meshplex](https://github.com/nschloe/meshplex) to calculate some mesh statistics, the benchmarks produce histograms of [dihedral angles](https://en.wikipedia.org/wiki/Dihedral_angle#:~:text=A%20dihedral%20angle%20is%20the,line%20as%20a%20common%20edge) in the cells and histograms of cell quality.
 
 
-For example running `benchmark_sphere.py` produces the following output:
-
-┌───────────────────────────┬────────────────────────────────────────────────────┬────────────────────────────────────────────┬──────────────────────┐
-│                           │                                                    │                                            │                      │
-│  CGAL                     │  Mesh creation time (seconds):    15.84            │  Number of vertices:       108407          │                      │
-│                           │  Mesh creation speed (vertices/seconds):  6843.10  │  Number of cells:           18118          │                      │
-│                           │                                                    │                                            │                      │
-├───────────────────────────┼────────────────────────────────────────────────────┼────────────────────────────────────────────┼──────────────────────┤
[…]
