### 2020-09-30 f527136
adding picture to README
