+++ benchmarks/README.md
@@ -24,6 +24,8 @@ Results
 Using [termplotlib](https://github.com/nschloe/termplotlib) and [meshplex](https://github.com/nschloe/meshplex) to calculate some mesh statistics, the benchmarks produce histograms of [dihedral angles](https://en.wikipedia.org/wiki/Dihedral_angle#:~:text=A%20dihedral%20angle%20is%20the,line%20as%20a%20common%20edge) in the cells and histograms of cell quality.
 
 
+![The computer used for benchmarking is a PC running MacOS with Dual-Core Intel Core i5 clocked at 2.00 GHz with 8GB of RAM. Both mesh generation programs have been compiled with g++ v8.3.0 with the -O3 option. These benchmarks have been done using CGAL v5.0 and SeismicMesh v3.0.3](https://user-images.githubusercontent.com/18619644/94739650-afaf6180-0347-11eb-903b-aa75b59612c7.jpg)
+
 Notes
 -----
 * Mesh generation with `cgal` is accomplished via [pygalmesh](https://github.com/nschloe/pygalmesh)
