### 2020-10-01 c7d3593
adding gmsh for spheroid
