+++ benchmarks/README.md
@@ -27,9 +27,50 @@ Using [termplotlib](https://github.com/nschloe/termplotlib) and [meshplex](https
 ![The computer used for benchmarking is a PC running MacOS with Dual-Core Intel Core i5 clocked at 2.00 GHz with 8GB of RAM. Both mesh generation programs have been compiled with g++ v8.3.0 with the -O3 option. These benchmarks have been done using CGAL v5.0 and SeismicMesh v3.0.3](https://user-images.githubusercontent.com/18619644/94751279-89e18700-035e-11eb-9ddf-b42995e4a041.jpg)
 
 
+Average speed statistics can be computed via [pytest-benchmark](https://pypi.org/project/pytest-benchmark/) which runs each problem 5 times and reports some statistic to the screen.
+
+```python
+py.test --benchmark-max-time=360 benchmarks/benchmark_sphere.py
+```
+
+produces for example:
+
+```
+=========================================================================================== test session starts ============================================================================================
+platform darwin -- Python 3.7.4, pytest-6.0.2, py-1.9.0, pluggy-0.13.1
+benchmark: 3.2.3 (defaults: timer=time.perf_counter disable_gc=False min_rounds=5 min_time=0.000005 max_time=360 calibration_precision=10 warmup=False warmup_iterations=100000)
+rootdir: /Users/Keith/junk/SeismicMesh, configfile: pytest.ini
+plugins: xdist-2.1.0, cov-2.10.1, benchmark-3.2.3, forked-1.3.0
+collected 3 items
+
+benchmarks/benchmark_sphere.py ...                                                                                                                                                                   [100%]
+
+============================================================================================= warnings summary =============================================================================================
+benchmarks/benchmark_sphere.py: 20 warnings
+  /Users/Keith/junk/SeismicMesh/SeismicMesh/geometry/utils.py:154: DeprecationWarning: Converting `np.character` to a dtype is deprecated. The current result is `np.dtype(np.str_)` which is not strictly correct. Note that `np.character` is generally deprecated and 'S1' should be used.
+    dtype = np.dtype((np.character, orig_dtype.itemsize * ncolumns))
+
+-- Docs: https://docs.pytest.org/en/stable/warnings.html
+
+-------------------------------------------------------------------------------- benchmark: 3 tests --------------------------------------------------------------------------------
+Name (time in s)          Min                Max               Mean            StdDev             Median               IQR            Outliers     OPS            Rounds  Iterations
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+test_gmsh              4.1087 (1.0)       4.6007 (1.0)       4.3494 (1.0)      0.1953 (1.0)       4.2846 (1.0)      0.2935 (1.0)           2;0  0.2299 (1.0)           5           1
+test_seismic_mesh     14.3477 (3.49)     16.1393 (3.51)     15.2947 (3.52)     0.7355 (3.77)     15.3189 (3.58)     1.2345 (4.21)          2;0  0.0654 (0.28)          5           1
+test_cgal             16.2922 (3.97)     19.2086 (4.18)     17.4405 (4.01)     1.2892 (6.60)     16.9541 (3.96)     2.2134 (7.54)          1;0  0.0573 (0.25)          5           1
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+
+Legend:
+  Outliers: 1 Standard Deviation from Mean; 1.5 IQR (InterQuartile Range) from 1st Quartile and 3rd Quartile.
+  OPS: Operations Per Second, computed as 1 / Mean
[…]
