+++ benchmarks/benchmark_cuboid.py
@@ -1,169 +1,0 @@
-import time
-import argparse
-import numpy as np
-
-import meshplex
-import meshio
-
-import pygalmesh
-import pygmsh
-import SeismicMesh
-
-from helpers import print_stats_3d
-
-HMIN = 0.025
-GRADE = HMIN / 3.0
-
-
-def test_seismic_mesh(benchmark):
-    angles, quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
-        run_SeismicMesh, iterations=1, rounds=5, warmup_rounds=0
-    )
-    assert np.amin(angles / np.pi * 180) > 10.0
-
-
-def test_gmsh(benchmark):
-    angles, quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
-        run_gmsh, iterations=1, rounds=5, warmup_rounds=0
-    )
-    assert np.amin(angles / np.pi * 180) > 10.0
-
-
-def test_cgal(benchmark):
-    angles, quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
-        run_cgal, iterations=1, rounds=5, warmup_rounds=0
-    )
-    assert np.amin(angles / np.pi * 180) > 10.0
-
-
-def run_gmsh():
-    with pygmsh.geo.Geometry() as geom:
-        geom.add_box(0, 1, 0, 1, 0, 1, 1.0)
-
-        geom.set_mesh_size_callback(
-            lambda dim, tag, x, y, z: ((1 - x) * GRADE + HMIN) / 1.1
-        )
-        t1 = time.time()
-        mesh = geom.generate_mesh()
-        elapsed = time.time() - t1
-
-    mesh.write("gmsh_box.vtk")
-    points = mesh.points
-    cells = mesh.cells[2].data
-
-    num_cells = len(cells)
-    num_vertices = len(points)
-
-    plex = meshplex.MeshTetra(points, cells)
-    angles = plex.q_min_sin_dihedral_angles
-    quality = plex.q_radius_ratio
-
-    return angles, quality, elapsed, num_vertices, num_cells
-
-
-def run_cgal():
-    class Field(pygalmesh.SizingFieldBase):
-        def eval(self, x):
-            h = (1 - x[0]) * GRADE + HMIN
-            return h / 1.1
-
-    t1 = time.time()
-    mesh = pygalmesh.generate_mesh(
-        pygalmesh.Cuboid([0.0, 0.0, 0.0], [1.0, 1.0, 1.0]),
-        cell_size=Field(),
-        facet_angle=30,
-        cell_radius_edge_ratio=2.0,
-        edge_size=HMIN,
-    )
-    elapsed = time.time() - t1
-
-    mesh.write("cgal_cuboid.vtk")
-
-    plex = meshplex.MeshTetra(mesh.points, mesh.cells[1][1])
-    angles = plex.q_min_sin_dihedral_angles
-    quality = plex.q_radius_ratio
-    num_verts = len(mesh.points)
-    num_cells = len(mesh.cells[1][1])
-
-    return angles, quality, elapsed, num_verts, num_cells
-
-
-def run_SeismicMesh():
-
-    bbox = (0.0, 1.0, 0.0, 1.0, 0.0, 1.0)
-
-    cube = SeismicMesh.Cube(bbox)
-
-    def fh(x):
-        return (1 - x[:, 0]) * GRADE + HMIN
-
-    t1 = time.time()
-    points, cells = SeismicMesh.generate_mesh(
-        bbox=bbox,
-        h0=HMIN,
-        domain=cube,
-        cell_size=fh,
-        nscreen=10,
-        max_iter=25,
-        delta_t=0.3,
-    )
-    points, cells = SeismicMesh.sliver_removal(
[…]
