### 2020-10-02 d907ac2
fixing Circle, improving 2d init. points in serial, 2d benchmark
