+++ benchmarks/benchmark_circle.py
@@ -1,0 +1,161 @@
+import time
+import argparse
+import numpy
+
+import meshplex
+import meshio
+
+import pygalmesh
+import pygmsh
+import SeismicMesh
+
+from helpers import print_stats_2d
+
+
+HMIN = 0.01
+
+
+def test_seismic_mesh(benchmark):
+    quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_SeismicMesh, iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(quality) > 10.0
+
+
+def test_gmsh(benchmark):
+    quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_gmsh, iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(quality) > 10.0
+
+
+def test_cgal(benchmark):
+    quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_cgal, iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(quality) > 10.0
+
+
+def run_gmsh():
+
+    with pygmsh.geo.Geometry() as geom:
+        geom.add_circle([0.0, 0.0], 1.0, mesh_size=HMIN)
+        mesh = geom.generate_mesh()
+
+        t1 = time.time()
+        mesh = geom.generate_mesh()
+        elapsed = time.time() - t1
+
+    mesh.write("gmsh_circle.vtk")
+
+    points = mesh.points
+    cells = mesh.cells[1].data
+
+    num_cells = len(cells)
+    num_vertices = len(points)
+
+    plex = meshplex.MeshTri(points, cells)
+    quality = numpy.abs(plex.cell_quality)
+
+    return quality, elapsed, num_vertices, num_cells
+
+
+def run_cgal():
+
+    n = 30
+    points = numpy.array(
+        [
+            [numpy.cos(alpha), numpy.sin(alpha)]
+            for alpha in numpy.linspace(0.0, 2 * numpy.pi, n, endpoint=False)
+        ]
+    )
+    constraints = [[k, k + 1] for k in range(n - 1)] + [[n - 1, 0]]
+
+    t1 = time.time()
+    mesh = pygalmesh.generate_2d(
+        points, constraints, cell_size=HMIN
+    )
+    elapsed = time.time() - t1
+
+    mesh.write("cgal_circle.vtk")
+
+    points = mesh.points
+    cells = mesh.cells[0].data
+
+    num_cells = len(cells)
+    num_vertices = len(points)
+
+    plex = meshplex.MeshTri(points, cells)
+    quality = numpy.abs(plex.cell_quality)
+
+    return quality, elapsed, num_vertices, num_cells
+
+
+def run_SeismicMesh():
+
+    bbox = (-1.0, 1.0, -1.0, 1.0)
+
+    circle = SeismicMesh.geometry.Circle(0, 0, 1)
+
+    def fh(x):
+        return numpy.array([HMIN] * len(x))
+
+    t1 = time.time()
+    points, cells = SeismicMesh.generate_mesh(
+        bbox=bbox,
+        h0=HMIN,
+        domain=circle,
+        cell_size=fh,
+        nscreen=10,
+        max_iter=25,
+        delta_t=0.3,
+    )
+    elapsed = time.time() - t1
+
+    meshio.write_points_cells(
+        "SeismicMesh_circle.vtk",
+        points,
+        [("triangle", cells)],
+    )
+
+    plex = meshplex.MeshTri(points, cells)
+    quality = numpy.abs(plex.cell_quality)
+
+    num_cells = len(cells)
[…]
