### 2020-10-02 c941985
adding BP2004 benchmark
