### 2020-10-05 f10a878
fixing some tests
