### 2020-10-05 406b9ed
adding in EAGE benchmark
