+++ benchmarks/benchmark_EAGE.py
@@ -1,0 +1,164 @@
+# benchmark the generation and quality of EAGE mesh
+import time
+import zipfile
+import argparse
+import numpy
+
+import meshplex
+import meshio
+import pygmsh
+import gmsh
+
+from SeismicMesh import (
+    get_sizing_function_from_segy,
+    Cube,
+    generate_mesh,
+    sliver_removal,
+)
+
+
+from helpers import print_stats_3d
+
+# Bounding box describing domain extents (corner coordinates)
+bbox = (-4200, 0, 0, 13520, 0, 13520)
+
+# Desired minimum mesh size in domain.
+HMIN = 150.0
+
+
+def run_gmsh(ef):
+    with pygmsh.geo.Geometry() as geom:
+
+        geom.add_box(-4200.0, 0.0, 0.0, 13520.0, 0.0, 13520.0, 1.0)
+
+        geom.set_mesh_size_callback(lambda dim, tag, x, y, z: (ef.eval([x, y, z])))
+
+        gmsh.option.setNumber("Mesh.CharacteristicLengthExtendFromBoundary", 0)
+        print("generating mesh")
+        t1 = time.time()
+        mesh = geom.generate_mesh()
+        elapsed = time.time() - t1
+
+        points = mesh.points
+        cells = mesh.cells[2].data
+
+        num_cells = len(cells)
+        num_vertices = len(points)
+
+        plex = meshplex.MeshTetra(points, cells)
+        angles = plex.q_min_sin_dihedral_angles
+        quality = plex.q_radius_ratio
+
+        mesh.write("gmsh_EAGE.vtk")
+
+        return angles, quality, elapsed, num_vertices, num_cells
+
+
+def run_SeismicMesh(ef):
+
+    cube = Cube(bbox)
+
+    t1 = time.time()
+    points, cells = generate_mesh(
+        domain=cube,
+        cell_size=ef,
+        h0=HMIN,
+        max_iter=25,
+    )
+
+    points, cells = sliver_removal(
+        points=points,
+        bbox=bbox,
+        h0=HMIN,
+        domain=cube,
+        cell_size=ef,
+        min_dh_angle_bound=10,
+        max_iter=50,
+        nscreen=10,
+    )
+    elapsed = time.time() - t1
+
+    meshio.write_points_cells(
+        "sm_EAGE.vtk",
+        points,
+        [("tetra", cells)],
+        file_format="vtk",
+    )
+
+    plex = meshplex.MeshTetra(points, cells)
+    angles = plex.q_min_sin_dihedral_angles
+    quality = plex.q_radius_ratio
+
+    num_cells = len(cells)
+    num_vertices = len(points)
+
+    return angles, quality, elapsed, num_vertices, num_cells
+
+
+def _build_sizing():
+
+    # This file is in a big Endian binary format, so we must tell the program the shape of the velocity model.
+    path = "Salt_Model_3D/3-D_Salt_Model/VEL_GRIDS/"
+    # Extract binary file Saltf@@ from SALTF.ZIP
+    zipfile.ZipFile(path + "SALTF.ZIP", "r").extract("Saltf@@", path=path)
+
+    fname = path + "Saltf@@"
+
+    # Dimensions of model (number of grid points in z, x, and y)
+    nx, ny, nz = 676, 676, 210
+
+    # A graded sizing function is created from the velocity model along with a signed distance function by passing
+    # the velocity grid that we created above.
+    # More details can be found here: https://seismicmesh.readthedocs.io/en/par3d/api.html
+    ef = get_sizing_function_from_segy(
[…]
