### 2020-10-05 f53e2ed
cleaning up benchmarks
