+++ benchmarks/benchmark_EAGE.py
@@ -5,7 +5,9 @@ import argparse
 import numpy
 
 import meshplex
-import meshio
+
+# import meshio
+import pygalmesh
 import pygmsh
 import gmsh
 
@@ -24,77 +26,6 @@ bbox = (-4200, 0, 0, 13520, 0, 13520)
 
 # Desired minimum mesh size in domain.
 HMIN = 150.0
-
-
-def run_gmsh(ef):
-    with pygmsh.geo.Geometry() as geom:
-
-        geom.add_box(-4200.0, 0.0, 0.0, 13520.0, 0.0, 13520.0, 150.0)
-
-        geom.set_mesh_size_callback(
-            lambda dim, tag, x, y, z: (ef.eval([x, y, z])) / 1.1
-        )
-        gmsh.option.setNumber("Mesh.CharacteristicLengthExtendFromBoundary", 0)
-
-        print("generating mesh...")
-        t1 = time.time()
-        mesh = geom.generate_mesh()
-        elapsed = time.time() - t1
-
-        points = mesh.points
-        cells = mesh.cells[2].data
-
-        num_cells = len(cells)
-        num_vertices = len(points)
-
-        plex = meshplex.MeshTetra(points, cells)
-        angles = plex.q_min_sin_dihedral_angles
-        quality = plex.q_radius_ratio
-
-        mesh.write("gmsh_EAGE.vtk")
-
-        return angles, quality, elapsed, num_vertices, num_cells
-
-
-def run_SeismicMesh(ef):
-
-    cube = Cube(bbox)
-
-    t1 = time.time()
-    points, cells = generate_mesh(
-        domain=cube,
-        cell_size=ef,
-        h0=HMIN,
-        max_iter=25,
-    )
-
-    points, cells = sliver_removal(
-        points=points,
-        bbox=bbox,
-        h0=HMIN,
-        domain=cube,
-        cell_size=ef,
-        min_dh_angle_bound=10,
-        max_iter=50,
-        nscreen=10,
-    )
-    elapsed = time.time() - t1
-
-    meshio.write_points_cells(
-        "sm_EAGE.vtk",
-        points,
-        [("tetra", cells)],
-        file_format="vtk",
-    )
-
-    plex = meshplex.MeshTetra(points, cells)
-    angles = plex.q_min_sin_dihedral_angles
-    quality = plex.q_radius_ratio
-
-    num_cells = len(cells)
-    num_vertices = len(points)
-
-    return angles, quality, elapsed, num_vertices, num_cells
 
 
 def _build_sizing():
@@ -130,6 +61,131 @@ def _build_sizing():
     return ef
 
 
+# for pytest-benchmark
+ef = _build_sizing()
+
+
+def test_seismic_mesh(benchmark):
+    angles, quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_SeismicMesh(*ef), iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(angles / numpy.pi * 180) > 10.0
+
+
+def test_gmsh(benchmark):
+    angles, quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_gmsh(*ef), iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(angles / numpy.pi * 180) > 10.0
+
+
+def test_cgal(benchmark):
+    angles, quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_cgal(ef), iterations=1, rounds=5, warmup_rounds=0
+    )
[…]
