+++ benchmarks/benchmark_BP2004.py
@@ -21,6 +21,45 @@ from helpers import print_stats_2d
 HMIN = 75.0
 
 
+def _build_sizing():
+
+    # Name of SEG-Y file containg velocity model.
+    fname = "vel_z6.25m_x12.5m_exact.segy"
+
+    # Bounding box describing domain extents (corner coordinates)
+    bbox = (-12000, 0.0, 0.0, 67000.0)
+
+    # Construct mesh sizing object from velocity model
+    ef = get_sizing_function_from_segy(
+        fname,
+        bbox,
+        hmin=HMIN,
+        wl=10,
+        freq=2,
+        dt=0.001,
+        grade=0.15,
+    )
+    return ef
+
+
+# for pytest-benchmark
+ef = _build_sizing()
+
+
+def test_seismic_mesh(benchmark):
+    quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_SeismicMesh, args=(ef,), iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(quality) > 0.10
+
+
+def test_gmsh(benchmark):
+    quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_gmsh, args=(ef,), iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(quality) > 0.10
+
+
 def run_gmsh(ef):
     with pygmsh.geo.Geometry() as geom:
 
@@ -50,7 +89,7 @@ def run_gmsh(ef):
         plex = meshplex.MeshTri(points, cells)
         quality = numpy.abs(plex.cell_quality)
 
-        mesh.write("BP2004_gmsh.vtk")
+        # mesh.write("BP2004_gmsh.vtk")
         return quality, elapsed, num_vertices, num_cells
 
 
@@ -74,12 +113,12 @@ def run_SeismicMesh(ef):
     )
     elapsed = time.time() - t1
 
-    meshio.write_points_cells(
-        "BP2004_sm.vtk",
-        points,
-        [("triangle", cells)],
-        file_format="vtk",
-    )
+    # meshio.write_points_cells(
+    #    "BP2004_sm.vtk",
+    #    points,
+    #    [("triangle", cells)],
+    #    file_format="vtk",
+    # )
 
     plex = meshplex.MeshTri(points, cells)
     quality = numpy.abs(plex.cell_quality)
@@ -88,27 +127,6 @@ def run_SeismicMesh(ef):
     num_vertices = len(points)
 
     return quality, elapsed, num_vertices, num_cells
-
-
-def _build_sizing():
-
-    # Name of SEG-Y file containg velocity model.
-    fname = "vel_z6.25m_x12.5m_exact.segy"
-
-    # Bounding box describing domain extents (corner coordinates)
-    bbox = (-12000, 0.0, 0.0, 67000.0)
-
-    # Construct mesh sizing object from velocity model
-    ef = get_sizing_function_from_segy(
-        fname,
-        bbox,
-        hmin=HMIN,
-        wl=10,
-        freq=2,
-        dt=0.001,
-        grade=0.15,
-    )
-    return ef
 
 
 if __name__ == "__main__":
@@ -122,26 +140,20 @@ if __name__ == "__main__":
         type=str,
         default=None,
         required=False,
-        help="Run benchmark with method=('cgal','sm', 'gmsh')",
+        help="Run benchmark with method=('sm', 'gmsh')",
     )
 
     args = parser.parse_args()
 
     ef = _build_sizing()
-    if args.method == "cgal":
-        print("not done")
[…]
