### 2020-10-05 e868d8c
adding EAGE benchmark figure
