+++ benchmarks/benchmark_BP2004.py
@@ -4,7 +4,8 @@ import argparse
 import numpy
 
 import meshplex
-import meshio
+
+# import meshio
 import pygmsh
 import gmsh
 
