### 2020-10-06 b2c12e7
fixing dihedral angles and figure
