+++ paper/paper.md
@@ -33,35 +33,46 @@ Generating a high-geometric quality graded mesh for a geophysical domain represe
 
 # Statement of Need
 
-Self-contained automatic meshing workflows that incorporate geophysical data to generate unstructured meshes have been successful in several domains such as coastal ocean modeling [e.g., `OceanMesh2D` @roberts2019oceanmesh2d; `GMT` and `Terreno` @gorman2008systematic] and reservoir modeling [e.g., `MeshIT` @cacace2015meshit]. Besides generating a high-quality mesh, the distribution of elements must accurately represent physical features in the domain such as faults and non-conformities as depicted in geophysical datasets. Element distributions must be placed such that the stability requirements of the numerical solver are satisfied and feature smooth transitions in element size. These aspects motivate the development of a self-contained mesh generation program.
+Self-contained automatic meshing workflows that incorporate geophysical data to generate unstructured meshes have been successful in several domains such as coastal ocean modeling [e.g., `OceanMesh2D` @roberts2019oceanmesh2d; `GMT` and `Terreno` @gorman2008systematic] and reservoir modeling [e.g., `MeshIT` @cacace2015meshit]. Besides generating a high-quality mesh, the distribution of elements must accurately represent physical features in the domain such as faults and non-conformities. Element distributions must be placed such that the stability requirements of the numerical solver are satisfied and feature smooth transitions in element size. These aspects motivate the development of a self-contained mesh generation program.
 
-Parallel capabilities are necessary as well as tetrahedral meshes can become computationally expensive to generate. In acoustic and elastic wave propagation, a minimum number between 5 to 10 vertices per wavelength (~250-2000 m) of the source wavelet is applied to ensure the subsequent simulation is numerically accurate. For example, in a 3D mesh of a benchmark FWI model EAGE Salt [@doi:10.1190/1.1437283] this requires 507,862 cells when resolving a 2 Hz source frequency using 5 vertices per wavelength. The number of cells increases by a factor of approximately 8 (approximately 4,082,978 cells) as the source frequency is doubled.
+Parallel capabilities can be useful as well as tetrahedral meshes can become computationally expensive to generate. In acoustic and elastic wave propagation, a minimum number between 5 to 10 vertices per wavelength (~250-2000 m) of the source wavelet is applied to ensure the subsequent simulation is numerically accurate. For example, in a 3D mesh of a benchmark FWI model EAGE Salt [@doi:10.1190/1.1437283] this requires 507,862 cells when resolving a 2 Hz source frequency using 5 vertices per wavelength. The number of cells increases by a factor of approximately 8 (approximately 4,082,978 cells) as the source frequency is doubled.
 
 
 # Core functionality
 
-A schematic of `SeismicMesh` is shown in \autoref{fig:workflow}. The core functionality is as follows:
+  1. The creation of 2D/3D graded mesh size functions defined on axis-aligned regular Cartesian grids. These mesh sizing functions encode mesh resolution distributions that conform to the variations from inputted seismic velocity model data and are distributed according to several heuristics [see @SeismicMeshDocs for further details]. Mesh size function grading is accomplished using [@persson2006mesh].
 
-  1. The creation of 2D/3D graded mesh size functions defined on regular Cartesian grids. These mesh sizing functions contain mesh resolution distributions that conform to the variations from inputted seismic velocity model data and are distributed according to several heuristics [see @SeismicMeshDocs for further details]. Note that mesh size function grading is accomplished using [@persson2006mesh].
[…]
