+++ paper/paper.bib
@@ -1,3 +1,22 @@
+@incollection{cgal:r-ctm2-20b,
+  author = {Laurent Rineau},
+  title = {{2D} Conforming Triangulations and Meshes},
+  publisher = {{CGAL Editorial Board}},
+  edition = {{5.1}},
+  booktitle = {{CGAL} User and Reference Manual},
+  url = {https://doc.cgal.org/5.1/Manual/packages.html#PkgMesh2},
+  year = 2020
+}
+
+@incollection{cgal:rty-m3-20b,
+  author = {Pierre Alliez and Cl{\'e}ment Jamin and Laurent Rineau and St{\'e}phane Tayeb and Jane Tournois and Mariette Yvinec},
+  title = {{3D} Mesh Generation},
+  publisher = {{CGAL Editorial Board}},
+  edition = {{5.1}},
+  booktitle = {{CGAL} User and Reference Manual},
+  url = {https://doc.cgal.org/5.1/Manual/packages.html#PkgMesh3},
+  year = 2020
+}
 
 @misc{SeismicMeshDocs,
   title = {Tutorial: SeismicMesh},
