### 2020-10-07 50d9793
force CI to use Python 3.6.4
