+++ .circleci/config.yml
@@ -3,7 +3,7 @@ version: 2
 jobs:
   lint:
     docker:
-      - image: circleci/python:3
+      - image: circleci/python:3.6.4
     steps:
       - checkout
       - run: pip3 install -U black flake8 --user
@@ -12,7 +12,7 @@ jobs:
   build:
     working_directory: ~/working
     docker:
-      - image: circleci/python:3
+      - image: circleci/python:3.6.4
     steps:
       - checkout
       - run: python --version
