+++ .circleci/config.yml
@@ -3,16 +3,16 @@ version: 2
 jobs:
   lint:
     docker:
-      - image: circleci/python:3.6.4
+      - image: circleci/python:3.7.4
     steps:
       - checkout
-      - run: pip3 install -U black flake8 --user
+      - run: pip install -U black flake8 --user
       - run: black --check .
       - run: flake8 .
   build:
     working_directory: ~/working
     docker:
-      - image: circleci/python:3.6.4
+      - image: circleci/python:3.7.4
     steps:
       - checkout
       - run: python --version
