+++ .circleci/config.yml
@@ -18,7 +18,7 @@ jobs:
       - run: python --version
       - run: pip --version
       - run: sudo apt install -y libopenmpi3 libopenmpi-dev openmpi-bin
-      - run: sudo apt install -y cmake
+      - run: sudo apt-get update && sudo apt install -y cmake
       - run: sudo apt-get install libboost-all-dev
       - run: sudo apt-get install libmpfr-dev
       - run: git clone https://github.com/CGAL/cgal.git && cd cgal/ && mkdir build && cd build && cmake -DCMAKE_BUILD_TYPE=Release .. && sudo make install
