+++ .circleci/config.yml
@@ -26,7 +26,7 @@ jobs:
       - run: sudo apt install -y python-pybind11
       - run: pip install . --user
       - run:
-             pip install tox && tox
+             pip install tox --user && tox
       - run: bash <(curl -s https://codecov.io/bash)
 workflows:
   version: 2
