### 2020-10-07 58bc57f
Benchmarking (#65)
Minor
------
* Adding  2d/3d benchmarking against `cgal`'s mesh generation capabilities and against `gmsh`.
* Updating JOSS paper to reflect benchmark results. 
* Accelerating mesh generation by coding some bottlenecks in C++. 
* User can now specify psuedo-timestep to `generate_mesh` to control how fast points can move around in the domain.
* Some doc string typo fixes.
