+++ benchmarks/benchmark_ball.py
@@ -1,0 +1,183 @@
+from math import sqrt
+import time
+import argparse
+import numpy
+
+import meshplex
+
+# import meshio
+
+import pygalmesh
+import pygmsh
+import SeismicMesh
+
+from helpers import print_stats_3d
+
+
+HMIN = 0.025
+
+
+def test_seismic_mesh(benchmark):
+    angles, quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_SeismicMesh, iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(angles / numpy.pi * 180) > 10.0
+
+
+def test_gmsh(benchmark):
+    angles, quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_gmsh, iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(angles / numpy.pi * 180) > 10.0
+
+
+def test_cgal(benchmark):
+    angles, quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_cgal, iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(angles / numpy.pi * 180) > 10.0
+
+
+def run_gmsh():
+    with pygmsh.occ.Geometry() as geom:
+        geom.add_ball([0.0, 0.0, 0.0], 1.0)
+
+        geom.set_mesh_size_callback(
+            lambda dim, tag, x, y, z: (
+                abs(sqrt(x ** 2 + y ** 2 + z ** 2) - 0.5) / 5 + HMIN
+            )
+            / 1.1
+        )
+        t1 = time.time()
+        mesh = geom.generate_mesh()
+        elapsed = time.time() - t1
+
+    # mesh.write("gmsh_sphere.vtk")
+    points = mesh.points
+    cells = mesh.cells[2].data
+
+    num_cells = len(cells)
+    num_vertices = len(points)
+
+    plex = meshplex.MeshTetra(points, cells)
+    angles = plex.q_min_sin_dihedral_angles
+    quality = plex.q_radius_ratio
+
+    return angles, quality, elapsed, num_vertices, num_cells
+
+
+def run_cgal():
+    class Field(pygalmesh.SizingFieldBase):
+        def eval(self, x):
+            h = abs(numpy.sqrt(numpy.dot(x, x)) - 0.5) / 5 + HMIN
+            return h / 1.1
+
+    t1 = time.time()
+    mesh = pygalmesh.generate_mesh(
+        pygalmesh.Ball([0.0, 0.0, 0.0], 1.0),
+        cell_size=Field(),
+        facet_angle=30,
+        cell_radius_edge_ratio=2.0,
+        edge_size=HMIN,
+    )
+    elapsed = time.time() - t1
+
+    # mesh.write("cgal_sphere.vtk")
+
+    plex = meshplex.MeshTetra(mesh.points, mesh.cells[1][1])
+    angles = plex.q_min_sin_dihedral_angles
+    quality = plex.q_radius_ratio
+
+    num_cells = len(mesh.cells[1][1])
+    num_vertices = len(mesh.points)
+
+    return angles, quality, elapsed, num_vertices, num_cells
+
+
+def run_SeismicMesh():
+
+    bbox = (-1.0, 1.0, -1.0, 1.0, -1.0, 1.0)
+
+    def dsphere(p, xc, yc, zc, r):
+        """Signed distance function for a sphere centered at xc,yc,zc with radius  r."""
+        return (
+            numpy.sqrt((p[:, 0] - xc) ** 2 + (p[:, 1] - yc) ** 2 + (p[:, 2] - zc) ** 2)
+            - r
[…]
