+++ benchmarks/requirements.txt
@@ -1,0 +1,6 @@
+meshplex==0.13.3
+pygalmesh==0.7.2
+pygmsh==7.0.0
+termplotlib==0.3.2
+meshio==4.2.2
+termplotlib==0.3.2

[…]
