+++ paper/paper.md
@@ -61,7 +61,7 @@ A simplified version of the parallel Delaunay algorithm proposed by [ @peterka20
 
 The 2D/3D serial performance against `gmsh` and `cgal` [@cgal:rty-m3-20b, @cgal:r-ctm2-20b] in terms of cell quality and creation time where cell quality is defined as dimension (2 or 3) multiplied by the circumcircle radius divided by the incircle radius. This cell quality is between 0 and 1, where 1 is a perfectly symmetrical simplex. For the analytical geometries (e.g. disk and ball) `gmsh` is the fastest to generate a mesh and then performance is approximately similar for both `SeismicMesh` and `cgal`. `gmsh` and `cgal` produce higher minimum cell qualities overall than `SeismicMesh` but importantly all generators are capable of producing sliver-free meshes. With that said, `SeismicMesh` produces higher mean cell qualities by about 5\% as compared to `gmsh` and `cgal`. For the two seismic meshes (e.g., BP2004 and EAGE), `SeismicMesh` is faster than `cgal` but slower than `gmsh` and the mesh quality similarly follows the results observed in the analytical cases. Interpolant-based mesh sizing functions significantly slow the mesh generation time of `gmsh` increasing its mesh generation time by a factor of 3x. The `DistMesh` algorithm used in `SeismicMesh` requires far less calls to the sizing function. For example in the EAGE benchmark, `gmsh` calls the sizing function 95,756 times whereas `DistMesh` calls it just 26 times.
 
-![The mesh creation time and resulting mesh quality for four bencmark problems studied. Two analytical problems (disk and a ball) and two non-anlaytical problems with sizing functions defined via interpolants (BP2004 and EAGE). All quantities are annotated next to points in the color that matches the generator used. For the cell quality panels a straight line indicates the minimum cell quality and the dot indicates the average cell quality. \label{fig:benchmark}](Benchmarks.jpg)
+![The mesh creation time and resulting mesh quality for four benchmark problems studied over a range of problem sizes. Two analytical problems (disk and a ball) and two non-analytical problems with sizing functions defined via regular gridded interpolants (BP2004 and EAGE). See the `SeismicMesh` github repository for information regarding the benchmarks. \label{fig:benchmark}](Benchmarks.png)
 
 
 # Ongoing and future applications
