### 2020-10-08 a8fc892
typos to README
