+++ benchmarks/benchmark_disk.py
@@ -3,7 +3,8 @@ import argparse
 import numpy
 
 import meshplex
-#import meshio
+
+# import meshio
 
 import pygalmesh
 import pygmsh
