### 2020-10-08 f15113c
improving plot (#66)
* improving plot
