### 2020-10-08 831f1b6
Update MANIFEST.in
