+++ MANIFEST.in
@@ -3,6 +3,7 @@ include SeismicMesh/geometry/cpp/*.cpp
 include SeismicMesh/migration/cpp/*.cpp
 include SeismicMesh/sizing/cpp/*.cpp
 include tests/*
+include benchmarks/*
 
 include README.md LICENSE.txt 
 global-include CMakeLists.txt 
