### 2020-10-08 49b1cbf
Update setup.cfg
