### 2020-10-08 8a97747
Update setup.cfg
