+++ setup.cfg
@@ -4,7 +4,7 @@ version = 3.0.3
 url = https://github.com/krober10nd/SeismicMesh
 author = Keith Roberts
 email = keithrbt0@gmail.com
-description = 2D/3D serial and parallel triangular mesh generation and mesh improvement for seismology
+description = 2D/3D serial and parallel triangular mesh generation for seismology
 long_description = file: README.md
 long_description_content_type = text/markdown
 license = GPL-3.0-or-later
