### 2020-10-09 82c95d9
add __version__ and circle --> disk
