+++ SeismicMesh/__about__.py
@@ -1,0 +1,10 @@
+try:
+    # Python 3.8
+    from importlib import metadata
+except ImportError:
+    import importlib_metadata as metadata
+
+try:
+    __version__ = metadata.version("SeismicMesh")
+except Exception:
+    __version__ = "unknown"
