### 2020-10-09 85b19e0
add __version__ and circle --> disk (#71)
* add __version__ and circle --> disk
