### 2020-10-09 7b59198
remove uncess. function arg
