### 2020-10-09 8623923
removing nscreen from benchmarks
