### 2020-10-09 75114ad
make verbose logical (#72)
* `nscreen` -> `verbose`: enable silencing of iterative information from `generate_mesh` and `sliver_removal` information to screen.
