### 2020-10-09 590f68b
adding support for constant cell-size
