### 2020-10-09 a6b78f9
adding constant cell sizes to README
