### 2020-10-09 ddfad35
fix to README blurbs
