### 2020-10-09 b29960d
imprv initial point formation
