### 2020-10-09 2a14487
new image for the cube
