### 2020-10-09 aa6533b
enabling parallel algorithm to use staggered grids
