### 2020-10-09 86b311a
clarify figure
