### 2020-10-09 cfcefb1
adding support for constant cell-size (#73)
* adding support for constant cell-sizes

* enabling algorithm to use staggered grids in serial/parallel and 2d/3d
