+++ tests/test_README.py
@@ -12,7 +12,9 @@ this_dir = pathlib.Path(__file__).resolve().parent
 @pytest.mark.serial
 @pytest.mark.parametrize(
     "string,lineno",
-    exdown.extract(this_dir.parent / "README.md", syntax_filter="python"),
+    exdown.extract(
+        this_dir.parent / "README.md", syntax_filter="python", max_num_lines=100000
+    ),
 )
 def test_readme(string, lineno):
 
