### 2020-10-09 db787f4
try resolving path in ci
