+++ benchmarks/benchmark_BP2004.py
@@ -101,7 +101,7 @@ def run_SeismicMesh(ef, HMIN=75.0):
     t1 = time.time()
     points, cells = generate_mesh(
         domain=rectangle,
-        cell_size=ef,
+        edge_length=ef,
         h0=HMIN,
         max_iter=25,
         delta_t=0.3,
