### 2020-10-10 7dc4222
cell_size -> edge_length (#77)
* cell_size -> edge_length
