### 2020-10-10 0176ec0
updating disk benchmark (#79)
* updating disk benchmark

* bumping version
