+++ benchmarks/requirements.txt
@@ -1,5 +1,5 @@
 meshplex==0.13.3
-pygalmesh==0.7.2
+pygalmesh==0.8.2
 pygmsh==7.0.0
 termplotlib==0.3.2
 meshio==4.2.2
