+++ benchmarks/README.md
@@ -63,9 +63,9 @@ Legend:
 
 Details on experiments
 -----------------------
-* Mesh generation with `cgal` is accomplished via [pygalmesh](https://github.com/nschloe/pygalmesh)
+* Mesh generation with `cgal` is accomplished via [pygalmesh](https://github.com/nschloe/pygalmesh) version 0.8.2
 * For `cgal`'s 3D mesh generator, all default quality options are assumed (e.g., facet angle bound of 30 degrees and the radius edge bound 2--to their theoretical limit). A `cell_size` function is passed to create variable mesh resolution in a way that is approximately equivalent to the mesh size function in `SeismicMesh`. In 2D we do not use Lloyd smoothing as it can significantly increase mesh generation time (but produce higher quality cells).
-* Mesh generation with `gmsh` is accomplished via [pygmsh](https://github.com/nschloe/pygmsh) with all default options and, similar to `cgal`, an approximately equivalent cell-size function is passed.
-* For `SeismicMesh`, we perform all examples with 25 meshing iterations with a psuedo-timestep of 0.30 and then run the sliver removal implemention to bound the diheral angle to 10 degrees in 3D and, in 2D, we delete any lower quality elements on the boundary (with a cell quality less than 10 percent).
+* Mesh generation with `gmsh` is accomplished via [pygmsh](https://github.com/nschloe/pygmsh) 7.0.0 with all default options and, similar to `cgal`, an approximately equivalent cell size function is passed.
+* For `SeismicMesh` version 3.0.4, we perform all examples with 25 meshing iterations with a psuedo-timestep of 0.30 and then run the sliver removal implemention to bound the diheral angle to 10 degrees in 3D and, in 2D, we delete any lower quality elements on the boundary (with a cell quality less than 10 percent).
 * All programs are executed in a seqeuntial mode. It's important to note however that a significant speed-up can be achieved for moderate to large problems using the [parallel capabilities](https://seismicmesh.readthedocs.io/en/par3d/tutorial.html#basics) provided in `SeismicMesh`. Threading based parallelism can be used with `gmsh` and `cgal` but these benchmarks have not been explored.
 * The scripts with the prefix `run` iterate over a range of relevant problem sizes to produce the timining and quality scales at different scales.
