### 2020-10-11 77c5f12
improve level set for parallel
