### 2020-10-11 b72db67
renaming functions
