### 2020-10-11 76213fb
change default psuedo timestep to 0.3
