+++ SeismicMesh/generation/mesh_generator.py
@@ -29,7 +29,7 @@ from .cpp.delaunay_class3 import DelaunayTriangulation3 as DT3
 __all__ = ["sliver_removal", "generate_mesh"]
 
 opts = {
-    "verbose": False,
+    "verbose": 1,
     "max_iter": 50,
     "seed": 0,
     "perform_checks": False,
@@ -42,7 +42,33 @@ opts = {
 }
 
 
-def sliver_removal(points, domain, edge_length, comm=None, **kwargs):
+def silence(func):
+    def wrapper(*args, **kwargs):
+        None
+
+    return wrapper
+
+
+def talk(func):
+    def wrapper(*args, **kwargs):
+        func(*args, **kwargs)
+
+    return wrapper
+
+
+def _select_verbosity(opts):
+    if opts["verbose"] == 0:
+        return silence, silence
+    elif opts["verbose"] == 1:
+        return talk, silence
+    elif opts["verbose"] > 1:
+        return talk, talk
+
+    else:
+        raise ValueError("Unknown verbosity level")
+
+
+def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C901
     r"""Improve an existing 3D mesh by removing degenerate cells.
     commonly referred to as `slivers`.
 
@@ -65,8 +91,9 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):
     :Keyword Arguments:
         * *h0* (``float``) --
             The minimum  edge length in the domain. REQUIRED IF USING A VARIABLE RESOLUTION EDGE LENGTH.
-        * *verbose* (``logical``) --
-            Output to the screen `verbose`. (default==False)
+        * *verbose* (``int``) --
+            Output to the screen `verbose`. (default==1). If `verbose`==1 only start and end messages are
+            written, `verbose`==0, no messages are written. `verbose` > 1 all messages are written.
         * *max_iter* (``float``) --
             Maximum number of meshing iterations. (default==50)
         * *perform_checks* (`boolean`) --
@@ -89,6 +116,16 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):
 
     opts.update(kwargs)
     _parse_kwargs(kwargs)
+
+    verbosity1, verbosity2 = _select_verbosity(opts)
+
+    @verbosity1
+    def print_msg1(msg):
+        print(msg)
+
+    @verbosity2
+    def print_msg2(msg):
+        print(msg)
 
     dim = points.shape[1]
     if dim == 2:
@@ -126,7 +163,7 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):
         raise ValueError("`max_iter` must be > 0")
     max_iter = opts["max_iter"]
 
-    print(
+    print_msg1(
         "Will attempt " + str(max_iter) + " iterations to bound the dihedral angles..."
     )
 
@@ -135,12 +172,12 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):
     min_dh_bound = opts["min_dh_angle_bound"] * math.pi / 180
     max_dh_bound = opts["max_dh_angle_bound"] * math.pi / 180
 
-    print(
+    print_msg1(
         "Enforcing a min. dihedral bound of: "
         + str(min_dh_bound * 180 / math.pi)
         + " degrees..."
     )
-    print(
+    print_msg1(
         "Enforcing a max. dihedral bound of: "
         + str(max_dh_bound * 180 / math.pi)
         + " degrees..."
@@ -152,14 +189,12 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):
 
     N = len(p)
 
-    print(
[…]
