### 2020-10-11 3e937b5
more control over verbosity (#85)
* more control over verbosity
