+++ SeismicMesh/generation/mesh_generator.py
@@ -284,7 +284,7 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
     return p, t
 
 
-def generate_mesh(domain, edge_length, comm=None, **kwargs):
+def generate_mesh(domain, edge_length, comm=None, **kwargs):  # noqa: C901
     r"""Generate a 2D/3D mesh using callbacks to a sizing function `edge_length` and signed distance function `domain`
 
     :param domain:
