+++ benchmarks/run_disk.py
@@ -4,7 +4,7 @@ import numpy
 from benchmark_disk import run_gmsh, run_SeismicMesh, run_cgal
 
 
-plt.rcParams.update({'font.size': 18})
+plt.rcParams.update({"font.size": 18})
 
 
 colors1 = ["ko-", "ro-", "bo-"]
