+++ paper/paper.bib
@@ -21,7 +21,7 @@
 @misc{SeismicMeshDocs,
   title = {Tutorial: SeismicMesh},
   author={Roberts, Keith},
-  howpublished = {\url{https://seismicmesh.readthedocs.io/en/par3d/tutorial.html}},
+  howpublished = {\url{https://seismicmesh.readthedocs.io/en/master/tutorial.html}},
   publisher={ReadTheDocs},
   journal={ReadTheDocs website},
   year = {2020},
