### 2020-10-13 2d91c48
adding optimization benchmarks
