+++ benchmarks/benchmark_ball_optmi.py
@@ -1,0 +1,189 @@
+from math import sqrt
+import time
+import argparse
+import numpy
+
+import meshplex
+
+import meshio
+
+import pygalmesh
+import pygmsh
+import SeismicMesh
+import gmsh
+
+from helpers import print_stats_3d
+
+
+def optimize(points, cells):
+    meshio.write_points_cells(
+        "cube.msh", points, [("tetra", cells)], file_format="gmsh22", binary=False
+    )
+    # optimize quality
+    gmsh.initialize()
+    gmsh.model.add("Mesh from file")
+    gmsh.merge("cube.msh")
+    t2 = time.time()
+    gmsh.model.mesh.optimize("", force=True)
+    print("Time spent in optimization: " + str(time.time() - t2))
+    gmsh.write("cube_ng.msh")
+    gmsh.finalize()
+    # read it back in
+    mesh = meshio.read("cube_ng.msh")
+    points, cells = mesh.points, mesh.cells[0].data
+    return points, cells
+
+
+def test_seismic_mesh(benchmark):
+    angles, quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_SeismicMesh, iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(angles / numpy.pi * 180) > 10.0
+
+
+def test_gmsh(benchmark):
+    angles, quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_gmsh, iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(angles / numpy.pi * 180) > 10.0
+
+
+def test_cgal(benchmark):
+    angles, quality, elapsed, num_vertices, num_cells = benchmark.pedantic(
+        run_cgal, iterations=1, rounds=5, warmup_rounds=0
+    )
+    assert numpy.amin(angles / numpy.pi * 180) > 10.0
+
+
+def run_gmsh(HMIN=0.025):
+    with pygmsh.occ.Geometry() as geom:
+        geom.add_ball([0.0, 0.0, 0.0], 1.0)
+
+        geom.set_mesh_size_callback(
+            lambda dim, tag, x, y, z: (
+                abs(sqrt(x ** 2 + y ** 2 + z ** 2) - 0.5) / 5 + HMIN
+            )
+            / 1.1
+        )
+        t1 = time.time()
+        mesh = geom.generate_mesh()
+        elapsed = time.time() - t1
+
+    points = mesh.points
+    cells = mesh.cells[2].data
+
+    points, cells = optimize(points, cells)
+
+    num_cells = len(cells)
+    num_vertices = len(points)
+
+    plex = meshplex.MeshTetra(points, cells)
+    angles = plex.q_min_sin_dihedral_angles
+    quality = plex.q_radius_ratio
+
+    return angles, quality, elapsed, num_vertices, num_cells
+
+
+def run_cgal(HMIN=0.025):
+
+    t1 = time.time()
+    mesh = pygalmesh.generate_mesh(
+        pygalmesh.Ball([0.0, 0.0, 0.0], 1.0),
+        cell_size=lambda x: abs(numpy.sqrt(numpy.dot(x, x)) - 0.5) / 5 + HMIN,
+        facet_angle=30,
+        cell_radius_edge_ratio=2.0,
+    )
+    elapsed = time.time() - t1
+
+    points, cells = optimize(mesh.points, mesh.cells[1][1])
+
+    plex = meshplex.MeshTetra(mesh.points, mesh.cells[1][1])
+    angles = plex.q_min_sin_dihedral_angles
+    quality = plex.q_radius_ratio
+
[…]
