### 2020-10-14 9a94526
adding more options for 3d velocity model reading
