+++ SeismicMesh/sizing/mesh_size_function.py
@@ -50,6 +50,8 @@ opts = {
     "nx": None,
     "ny": None,
     "byte_order": "byte_order",
+    "axes_order": [0, 1, 2],
+    "axes_order_sort": "F",
 }
 
 
@@ -98,6 +100,10 @@ def get_sizing_function_from_segy(filename, bbox, comm=None, **kwargs):
              REQUIRED FOR BINARY VELOCITY MODEL. The number of grid points in the x-direction in the velocity model.
         * *byte_order* (``string``) --
              REQUIRED FOR BINARY VELOCITY MODEL. The order of bytes in a 3D sesimic velocity model (`little` or `big`).
+        * *axes_order* (``list``) --
+             The order of the axes (assumes z,x,y [0,1,2])
+        * *axes_order_sort* (``string``) --
+             The sort style of the data (either "F" for FORTRAN-style or "C" for C-style)
 
     :return: a :class:`SizeFunction` object with a `obj.bbox` field and an `obj.eval` method.
     :rtype: a :class:`SizeFunction` object
@@ -115,6 +121,8 @@ def get_sizing_function_from_segy(filename, bbox, comm=None, **kwargs):
             nx=opts["nx"],
             ny=opts["ny"],
             byte_order=opts["byte_order"],
+            axes_order=opts["axes_order"],
+            axes_order_sort=opts["axes_order_sort"],
         )
 
         if opts["units"] == "km-s":
@@ -490,20 +498,32 @@ def _pad_it(array, padding, style, extra):
     return array
 
 
-def _read_velocity_model(filename, nz=None, nx=None, ny=None, byte_order=None):
+def _read_velocity_model(
+    filename,
+    nz=None,
+    nx=None,
+    ny=None,
+    byte_order=None,
+    axes_order=None,
+    axes_order_sort=None,
+):
     """Read a velocity model"""
     if filename.endswith(".segy"):
         return _read_segy(filename)
     else:
-        return _read_bin(filename, nz, nx, ny, byte_order)
+        return _read_bin(filename, nz, nx, ny, byte_order, axes_order, axes_order_sort)
 
 
-def _read_bin(filename, nz, nx, ny, byte_order):
+def _read_bin(
+    filename, nz, nx, ny, byte_order, axes_order=[0, 1, 2], axes_order_sort="F"
+):
     """Read a velocity model from a binary"""
     if (nz is None) or (nx is None) or (ny is None):
         raise ValueError(
             "Please specify the number of grid points in each dimension (e.g., `nz`, `nx`, `ny`)..."
         )
+    axes = [nz, nx, ny]
+    axes = axes[axes_order]
     with open(filename, "r") as file:
         print("Reading binary file: " + filename)
         if byte_order == "big":
@@ -512,8 +532,8 @@ def _read_bin(filename, nz, nx, ny, byte_order):
             vp = np.fromfile(file, dtype=np.dtype("float32").newbyteorder("<"))
         else:
             raise ValueError("Please specify byte_order as either: little or big.")
-        vp = vp.reshape(nx, ny, nz, order="F")
-        return np.flipud(vp.transpose((2, 0, 1))), nz, nx, ny  # z, x and then y
+        vp = vp.reshape(*axes_order, order=axes_order_sort)
+        return np.flipud(vp.transpose((*axes_order))), nz, nx, ny  # z, x and then y
 
 
 def _read_segy(filename):
