### 2020-10-14 cedfbd8
updating write_velocity_model
