### 2020-10-14 ee34933
updating benchmark read
