+++ README.md
@@ -176,7 +176,7 @@ ef = get_sizing_function_from_segy(
     nx=nx,
     ny=ny,
     byte_order="big",
-    axes_order=[2, 0, 1],  # default order z, x, y -> order for EAGE x, y, z
+    axes_order=(2, 0, 1),  # order for EAGE (x, y, z) to default order (z,x,y)
     axes_order_sort="F", # binary is packed in a FORTRAN-style
 )
 
