### 2020-10-14 fe2834f
Improve 3d read (#90)
* adding more options and control for reading 3d velocity models in a binary format.
