+++ SeismicMesh/sizing/mesh_size_function.py
@@ -50,6 +50,8 @@ opts = {
     "nx": None,
     "ny": None,
     "byte_order": "byte_order",
+    "axes_order": (0, 1, 2),
+    "axes_order_sort": "F",
 }
 
 
@@ -98,6 +100,10 @@ def get_sizing_function_from_segy(filename, bbox, comm=None, **kwargs):
              REQUIRED FOR BINARY VELOCITY MODEL. The number of grid points in the x-direction in the velocity model.
         * *byte_order* (``string``) --
              REQUIRED FOR BINARY VELOCITY MODEL. The order of bytes in a 3D sesimic velocity model (`little` or `big`).
+        * *axes_order* (``tuple``) --
+             The order of the axes (assumes z,x,y is (0,1,2))
+        * *axes_order_sort* (``string``) --
+             The sort style of the data (either "F" for FORTRAN-style or "C" for C-style)
 
     :return: a :class:`SizeFunction` object with a `obj.bbox` field and an `obj.eval` method.
     :rtype: a :class:`SizeFunction` object
@@ -115,6 +121,8 @@ def get_sizing_function_from_segy(filename, bbox, comm=None, **kwargs):
             nx=opts["nx"],
             ny=opts["ny"],
             byte_order=opts["byte_order"],
+            axes_order=opts["axes_order"],
+            axes_order_sort=opts["axes_order_sort"],
         )
 
         if opts["units"] == "km-s":
@@ -148,6 +156,8 @@ def get_sizing_function_from_segy(filename, bbox, comm=None, **kwargs):
                 "nx",
                 "ny",
                 "byte_order",
+                "axes_order",
+                "axes_order_sort",
             }:
                 pass
             else:
@@ -217,7 +227,10 @@ def write_velocity_model(filename, ofname=None, comm=None, **kwargs):
              Width of the domain pad in meters.
         * *pad_style* (``string``) --
              Type of padding.
-
+        * *axes_order* (``tuple``) --
+             The order of the axes (assumes z,x,y is (0,1,2))
+        * *axes_order_sort* (``string``) --
+             The sort style of the data (either "F" for FORTRAN-style or "C" for C-style)
 
     """
     comm = comm or MPI.COMM_WORLD
@@ -234,6 +247,8 @@ def write_velocity_model(filename, ofname=None, comm=None, **kwargs):
             nx=opts["nx"],
             ny=opts["ny"],
             byte_order=opts["byte_order"],
+            axes_order=opts["axes_order"],
+            axes_order_sort=opts["axes_order_sort"],
         )
 
         if opts["domain_pad"] > 0.0:
@@ -490,20 +505,30 @@ def _pad_it(array, padding, style, extra):
     return array
 
 
-def _read_velocity_model(filename, nz=None, nx=None, ny=None, byte_order=None):
+def _read_velocity_model(
+    filename,
+    nz=None,
+    nx=None,
+    ny=None,
+    byte_order=None,
+    axes_order=None,
+    axes_order_sort=None,
+):
     """Read a velocity model"""
     if filename.endswith(".segy"):
         return _read_segy(filename)
     else:
-        return _read_bin(filename, nz, nx, ny, byte_order)
+        return _read_bin(filename, nz, nx, ny, byte_order, axes_order, axes_order_sort)
 
 
-def _read_bin(filename, nz, nx, ny, byte_order):
+def _read_bin(filename, nz, nx, ny, byte_order, axes_order, axes_order_sort):
     """Read a velocity model from a binary"""
     if (nz is None) or (nx is None) or (ny is None):
         raise ValueError(
[…]
