### 2020-10-14 23b28a4
Warning fix (#92)
* Switch to warning in unit detection.

* Docstrings
