### 2020-10-14 4a4dcfd
checking if bbox is all floats
