### 2020-10-14 182f3c5
ensure bbox is all floats (#93)
* Adding a check if bbox is all floats
