+++ tests/test_pfix.py
@@ -1,3 +1,4 @@
+import pytest
 import numpy as np
 import SeismicMesh
 
