### 2020-10-16 f7d9631
preserving pfix (#95)
* Ensuring fixed points `pfix` remain preserved in connectivity.
