+++ SeismicMesh/geometry/__init__.py
@@ -28,6 +28,7 @@ from .utils import (
     unique_rows,
     vertex_to_entities,
     vertex_in_entity3,
+    corners,
 )
 
 __all__ = [
@@ -36,6 +37,7 @@ __all__ = [
     "calc_circumsphere_grad",
     "calc_3x3determinant",
     "calc_4x4determinant",
+    "corners",
     "Rectangle",
     "Cube",
     "Disk",
