### 2020-10-16 49c55c7
improving code for sliver removal
