+++ SeismicMesh/generation/mesh_generator.py
@@ -196,15 +196,17 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
     count = 0
     pold = None
     push = 0.10
+
+    dt = DT()
+    dt.insert(p.flatten().tolist())
     while True:
 
         start = time.time()
 
+        num_move = 0
+
         # Using CGAL's incremental Delaunay triangulation capabilities.
-        if count == 0:
-            dt = DT()
-            dt.insert(p.flatten().tolist())
-        else:
+        if count > 0:
             to_move = np.where(_dist(p, pold) > 0)[0]
             dt.move(to_move.flatten().tolist(), p[to_move].flatten().tolist())
 
@@ -216,57 +218,6 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
         # Remove points outside the domain
         t = _remove_triangles_outside(p, t, fd, geps)
 
-        # Sliver removal
-        if count != (max_iter - 1):
-            num_move = 0
-            # calculate dihedral angles in mesh
-            dh_angles = geometry.calc_dihedral_angles(p, t)
-            out_of_bounds = np.argwhere(
-                (dh_angles[:, 0] < min_dh_bound) | (dh_angles[:, 0] > max_dh_bound)
-            )
-            ele_nums = np.floor(out_of_bounds / 6).astype("int")
-            ele_nums, ix = np.unique(ele_nums, return_index=True)
-
-            print_msg1(
-                "On rank: "
-                + str(comm.rank)
-                + " There are "
-                + str(len(ele_nums))
-                + " slivers...",
-            )
-
-            move = t[ele_nums, 0]
-            num_move = move.size
-            if num_move == 0:
-                print_msg1(
-                    "Termination reached in "
-                    + str(count)
-                    + " iterations...no slivers detected!",
-                )
-                p, t, _ = geometry.fix_mesh(p, t, dim=dim, delete_unused=True)
-                return p, t
-
-            p0, p1, p2, p3 = (
-                p[t[ele_nums, 0], :],
-                p[t[ele_nums, 1], :],
-                p[t[ele_nums, 2], :],
-                p[t[ele_nums, 3], :],
-            )
-
-            # perturb vector is based on INCREASING circumsphere's radius
-            perturb = geometry.calc_circumsphere_grad(p0, p1, p2, p3)
-            perturb[np.isinf(perturb)] = 1.0
-
-            # normalize perturbation vector
-            perturb_norm = np.sum(np.abs(perturb) ** 2, axis=-1) ** (1.0 / 2)
-            perturb /= perturb_norm[:, None]
-
-            # perturb % of local mesh size
-            p[move] += push * h0 * perturb
-
-        # Bring outside points back to the boundary
-        p = _project_points_back_newton(p, fd, deps)
-
         # Number of iterations reached, stop.
         if count == (max_iter - 1):
             print_msg1(
@@ -274,6 +225,56 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
             )
             p, t = _termination(p, t, opts, comm, sliver=True)
             break
+
+        # calculate the minimum dihedral angle in mesh
+        dh_angles = geometry.calc_dihedral_angles(p, t)
+        out_of_bounds = np.argwhere(
+            (dh_angles[:, 0] < min_dh_bound) | (dh_angles[:, 0] > max_dh_bound)
+        )
+        ele_nums = np.floor(out_of_bounds / 6).astype("int")
+        ele_nums, ix = np.unique(ele_nums, return_index=True)
+
+        print_msg1(
+            "On rank: "
+            + str(comm.rank)
[…]
