### 2020-10-16 57af9b7
some refactor for clarity
