### 2020-10-16 4788678
fix lgtm alerts
