### 2020-10-17 7b8ad18
adding user geps_mult parameter
