### 2020-10-17 #2687 @jorgensd
> @arfon I am not able to tick the checklist boxes above, and when trying to accept the invitation I get that it is expired.

@arfon Im still not able to tick the boxes above (even now that Im assigned as reveiwer), is it possible to resend an invitiation to me?
