### 2020-10-17 244d161
doc typos (#98)
* Typos in docs
