### 2020-10-17 #2687 @arfon
@jorgensd can you try accepting the invite again now? The invites should work for a week.
