### 2020-10-17 4ba1442
bad 3d sdf (#99)
* experimental 3d sdf does not work in parallel
