### 2020-10-18 ca93980
increasing size of buffer in parallel to 5*h0
