### 2020-10-18 89c1b3c
improving sliver removal robustness (#97)
* some `sliver_removal` refactor for clarity

* adding user geps_mult parameter

* bbox fix to benchmarks

* increasing size of buffer in parallel to 5*h0

* version bump
