+++ SeismicMesh/generation/mesh_generator.py
@@ -36,9 +36,10 @@ opts = {
     "pfix": None,
     "axis": 1,
     "min_dh_angle_bound": 10.0,
-    "max_dh_angle_bound": 170.0,
+    "max_dh_angle_bound": 180.0,
     "points": None,
     "delta_t": 0.30,
+    "geps_mult": 0.1,
 }
 
 
@@ -106,6 +107,8 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
             The minimum allowable dihedral angle bound. (default==10 degrees)
         * *max_dh_angle_bound* (`float`) --
             The maximum allowable dihedral angle bound. (default==170 degrees)
+        * *geps_mult* (`float`) --
+            The tolerance used to determine if a vertex is "in" the domain. (default==0.1)
 
     """
     comm = comm or MPI.COMM_WORLD
@@ -159,6 +162,8 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
     if h0 < 0:
         raise ValueError("`h0` must be > 0")
 
+    deps = np.sqrt(np.finfo(np.double).eps) * h0
+
     if opts["max_iter"] < 0:
         raise ValueError("`max_iter` must be > 0")
     max_iter = opts["max_iter"]
@@ -167,8 +172,7 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
         "Will attempt " + str(max_iter) + " iterations to bound the dihedral angles..."
     )
 
-    geps = 1e-1 * h0
-    deps = np.sqrt(np.finfo(np.double).eps) * h0
+    geps = opts["geps_mult"] * h0
     min_dh_bound = opts["min_dh_angle_bound"] * math.pi / 180
     max_dh_bound = opts["max_dh_angle_bound"] * math.pi / 180
 
@@ -196,15 +200,16 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
     count = 0
     pold = None
     push = 0.10
+
+    dt = DT()
+    dt.insert(p.flatten().tolist())
     while True:
 
         start = time.time()
 
+        num_move = 0
         # Using CGAL's incremental Delaunay triangulation capabilities.
-        if count == 0:
-            dt = DT()
-            dt.insert(p.flatten().tolist())
-        else:
+        if count > 0:
             to_move = np.where(_dist(p, pold) > 0)[0]
             dt.move(to_move.flatten().tolist(), p[to_move].flatten().tolist())
 
@@ -216,64 +221,57 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
         # Remove points outside the domain
         t = _remove_triangles_outside(p, t, fd, geps)
 
-        # Sliver removal
-        if count != (max_iter - 1):
-            num_move = 0
-            # calculate dihedral angles in mesh
-            dh_angles = geometry.calc_dihedral_angles(p, t)
-            out_of_bounds = np.argwhere(
-                (dh_angles[:, 0] < min_dh_bound) | (dh_angles[:, 0] > max_dh_bound)
-            )
-            ele_nums = np.floor(out_of_bounds / 6).astype("int")
-            ele_nums, ix = np.unique(ele_nums, return_index=True)
-
-            print_msg1(
-                "On rank: "
-                + str(comm.rank)
-                + " There are "
-                + str(len(ele_nums))
-                + " slivers...",
-            )
-
-            move = t[ele_nums, 0]
-            num_move = move.size
-            if num_move == 0:
-                print_msg1(
-                    "Termination reached in "
-                    + str(count)
-                    + " iterations...no slivers detected!",
-                )
[…]
