### 2020-10-19 d9fdb0a
some small fixes to README
