### 2020-10-19 04ffaa0
Improving images. (#100)
* adding torus example

* some small fixes to README

* prism
