### 2020-10-21 cf6ac61
run garbage collection upon exit
