+++ SeismicMesh/generation/mesh_generator.py
@@ -14,6 +14,7 @@
 #   along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import math
+import gc
 import time
 import warnings
 
@@ -28,12 +29,22 @@ from .cpp.delaunay_class3 import DelaunayTriangulation3 as DT3
 
 __all__ = ["sliver_removal", "generate_mesh"]
 
-opts = {
+gen_opts = {
     "verbose": 1,
     "max_iter": 50,
     "seed": 0,
     "perform_checks": False,
     "pfix": None,
+    "axis": 1,
+    "points": None,
+    "delta_t": 0.30,
+    "geps_mult": 0.1,
+}
+
+sliver_opts = {
+    "verbose": 1,
+    "max_iter": 50,
+    "perform_checks": False,
     "axis": 1,
     "min_dh_angle_bound": 10.0,
     "max_dh_angle_bound": 180.0,
@@ -117,10 +128,10 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
             warnings.warn("Sliver removal only works in serial for now")
         return True, True
 
-    opts.update(kwargs)
+    sliver_opts.update(kwargs)
     _parse_kwargs(kwargs)
 
-    verbosity1, verbosity2 = _select_verbosity(opts)
+    verbosity1, verbosity2 = _select_verbosity(sliver_opts)
 
     @verbosity1
     def print_msg1(msg):
@@ -156,7 +167,7 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
     if hmin is not None:
         h0 = hmin
     else:
-        h0 = opts["h0"]
+        h0 = sliver_opts["h0"]
 
     # check h0
     if h0 < 0:
@@ -164,17 +175,17 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
 
     deps = np.sqrt(np.finfo(np.double).eps) * h0
 
-    if opts["max_iter"] < 0:
+    if sliver_opts["max_iter"] < 0:
         raise ValueError("`max_iter` must be > 0")
-    max_iter = opts["max_iter"]
+    max_iter = sliver_opts["max_iter"]
 
     print_msg1(
         "Will attempt " + str(max_iter) + " iterations to bound the dihedral angles..."
     )
 
-    geps = opts["geps_mult"] * h0
-    min_dh_bound = opts["min_dh_angle_bound"] * math.pi / 180
-    max_dh_bound = opts["max_dh_angle_bound"] * math.pi / 180
+    geps = sliver_opts["geps_mult"] * h0
+    min_dh_bound = sliver_opts["min_dh_angle_bound"] * math.pi / 180
+    max_dh_bound = sliver_opts["max_dh_angle_bound"] * math.pi / 180
 
     print_msg1(
         "Enforcing a min. dihedral bound of: "
@@ -228,7 +239,7 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
             print_msg1(
                 "FAILURE: Termination...maximum number of iterations reached.",
             )
-            p, t = _termination(p, t, opts, comm, sliver=True)
+            p, t = _termination(p, t, sliver_opts, comm, sliver=True)
             break
 
         print_msg1(
@@ -279,6 +290,7 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
         if comm.rank == 0:
             print_msg2("     Elapsed wall-clock time %f : " % (end - start))
 
+    gc.collect()
     return p, t
 
 
[…]
