### 2020-10-21 44f773f
removing global variables
