+++ SeismicMesh/generation/mesh_generator.py
@@ -14,7 +14,6 @@
 #   along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import math
-import gc
 import time
 import warnings
 
@@ -28,30 +27,6 @@ from .cpp.delaunay_class import DelaunayTriangulation as DT2
 from .cpp.delaunay_class3 import DelaunayTriangulation3 as DT3
 
 __all__ = ["sliver_removal", "generate_mesh"]
-
-gen_opts = {
-    "verbose": 1,
-    "max_iter": 50,
-    "seed": 0,
-    "perform_checks": False,
-    "pfix": None,
-    "axis": 1,
-    "points": None,
-    "delta_t": 0.30,
-    "geps_mult": 0.1,
-}
-
-sliver_opts = {
-    "verbose": 1,
-    "max_iter": 50,
-    "perform_checks": False,
-    "axis": 1,
-    "min_dh_angle_bound": 10.0,
-    "max_dh_angle_bound": 180.0,
-    "points": None,
-    "delta_t": 0.30,
-    "geps_mult": 0.1,
-}
 
 
 def silence(func):
@@ -128,6 +103,18 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
             warnings.warn("Sliver removal only works in serial for now")
         return True, True
 
+    sliver_opts = {
+        "verbose": 1,
+        "max_iter": 50,
+        "perform_checks": False,
+        "axis": 1,
+        "min_dh_angle_bound": 10.0,
+        "max_dh_angle_bound": 180.0,
+        "points": None,
+        "delta_t": 0.30,
+        "geps_mult": 0.1,
+    }
+
     sliver_opts.update(kwargs)
     _parse_kwargs(kwargs)
 
@@ -146,7 +133,7 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
         if comm.rank == 0:
             raise Exception("Mesh improvement currently on works in 3D")
 
-    fd, bbox0 = _unpack_domain(domain)
+    fd, bbox0 = _unpack_domain(domain, sliver_opts)
 
     fh, bbox1, hmin = _unpack_sizing(edge_length)
 
@@ -290,7 +277,6 @@ def sliver_removal(points, domain, edge_length, comm=None, **kwargs):  # noqa: C
         if comm.rank == 0:
             print_msg2("     Elapsed wall-clock time %f : " % (end - start))
 
-    gc.collect()
     return p, t
 
 
@@ -341,7 +327,17 @@ def generate_mesh(domain, edge_length, comm=None, **kwargs):  # noqa: C901
 
     """
     comm = comm or MPI.COMM_WORLD
-
+    gen_opts = {
+        "verbose": 1,
+        "max_iter": 50,
+        "seed": 0,
+        "perform_checks": False,
+        "pfix": None,
+        "axis": 1,
+        "points": None,
+        "delta_t": 0.30,
+        "geps_mult": 0.1,
+    }
     # check call was correct
     gen_opts.update(kwargs)
     _parse_kwargs(kwargs)
@@ -358,7 +354,7 @@ def generate_mesh(domain, edge_length, comm=None, **kwargs):  # noqa: C901
         print(msg, flush=True)
 
     # unpack domain
-    fd, bbox0 = _unpack_domain(domain)
+    fd, bbox0 = _unpack_domain(domain, gen_opts)
 
     fh, bbox1, hmin = _unpack_sizing(edge_length)
 
@@ -453,7 +449,6 @@ def generate_mesh(domain, edge_length, comm=None, **kwargs):  # noqa: C901
                     "Termination reached...maximum number of iterations reached.",
                 )
             p, t = _termination(p, t, gen_opts, comm)
[…]
