### 2020-10-21 794f01a
fixing some tests
