### 2020-10-21 df2b7f4
global variables (#105)
* removing global dictionaries and placing them inside their functions to avoid mutable default arguments.
