+++ SeismicMesh/geometry/__init__.py
@@ -6,7 +6,15 @@ from .cpp.fast_geometry import (
     calc_volume_grad,
     unique_edges,
 )
-from .signed_distance_functions import Rectangle, Cube, Disk, drectangle, dblock
+from .signed_distance_functions import (
+    Rectangle,
+    Cube,
+    Disk,
+    drectangle,
+    dblock,
+    corners,
+    Union,
+)
 from .utils import (
     calc_re_ratios,
     delete_boundary_entities,
@@ -28,7 +36,6 @@ from .utils import (
     unique_rows,
     vertex_to_entities,
     vertex_in_entity3,
-    corners,
 )
 
 __all__ = [
@@ -64,4 +71,5 @@ __all__ = [
     "get_winded_boundary_edges",
     "vertex_in_entity3",
     "unique_edges",
+    "Union",
 ]
