+++ SeismicMesh/geometry/utils.py
@@ -1,4 +1,3 @@
-import itertools
 import numpy as np
 import scipy.sparse as spsparse
 
