### 2020-10-21 d732f0b
unpack in sliver_removal
