+++ tests/test_2dmesher_par.py
@@ -4,12 +4,8 @@ import numpy as np
 import pytest
 from mpi4py import MPI
 
-from SeismicMesh import (
-    get_sizing_function_from_segy,
-    generate_mesh,
-    Rectangle,
-    geometry,
-)
+from SeismicMesh import (Rectangle, generate_mesh, geometry,
+                         get_sizing_function_from_segy)
 
 comm = MPI.COMM_WORLD
 
