+++ tests/test_2dmesher_par_adapt.py
@@ -4,8 +4,12 @@ import numpy as np
 import pytest
 from mpi4py import MPI
 
-from SeismicMesh import (Rectangle, generate_mesh, geometry,
-                         get_sizing_function_from_segy)
+from SeismicMesh import (
+    Rectangle,
+    generate_mesh,
+    geometry,
+    get_sizing_function_from_segy,
+)
 
 comm = MPI.COMM_WORLD
 
