+++ SeismicMesh/geometry/__init__.py
@@ -10,6 +10,7 @@ from .signed_distance_functions import (
     Rectangle,
     Cube,
     Disk,
+    Ball,
     drectangle,
     dblock,
     corners,
@@ -48,6 +49,7 @@ __all__ = [
     "calc_4x4determinant",
     "corners",
     "Rectangle",
+    "Ball",
     "Cube",
     "Disk",
     "drectangle",
