### 2020-10-21 57943b1
do not do tuple expansion in sdf
