### 2020-10-21 7d6b724
improvements to geometry API (#106)
* adding Union/Intersection/Difference classes for geometric primitives.
* adding a ball SDF
* Automatic corner constraints in serial
