+++ tests/test_3dmesher.py
@@ -1,13 +1,13 @@
 import os
 
-from numpy import allclose
 import pytest
+from numpy import allclose
 
 from SeismicMesh import (
-    generate_mesh,
-    sliver_removal,
     Cube,
+    generate_mesh,
     get_sizing_function_from_segy,
+    sliver_removal,
     write_velocity_model,
 )
 
