+++ tests/test_3dmesher_par.py
@@ -5,11 +5,11 @@ import pytest
 from mpi4py import MPI
 
 from SeismicMesh import (
-    get_sizing_function_from_segy,
-    generate_mesh,
-    sliver_removal,
-    geometry,
     Cube,
+    generate_mesh,
+    geometry,
+    get_sizing_function_from_segy,
+    sliver_removal,
 )
 
 comm = MPI.COMM_WORLD
