+++ tests/test_2dmesher_par.py
@@ -5,10 +5,10 @@ import pytest
 from mpi4py import MPI
 
 from SeismicMesh import (
-    get_sizing_function_from_segy,
-    generate_mesh,
     Rectangle,
+    generate_mesh,
     geometry,
+    get_sizing_function_from_segy,
 )
 
 comm = MPI.COMM_WORLD
