+++ README.md
@@ -516,6 +516,7 @@ and this project (tries to) adhere to [Semantic Versioning](https://semver.org/s
 
 ## Unreleased
 
+## [3.0.6] - 2020-10-21
 ### Fixed
 - Silence messages about pfix when verbose=0
 ### Added
