### 2020-10-22 #2687 @krober10nd
@jorgensd Yes, you are correct. I made the technical contributions/implementations to this work and wrote the paper and I also maintain the software.  Both Rafael and William offered their scientific/professional opinions and we had numerous intellectual discussions which led to new ideas and which I later pursued and implemented in the code stack. Both helped review and edit offline earlier versions of the program despite the fact they have not committed to the repository.
