### 2020-10-22 d71ccee
readme phrasing (#108)
* phrasing of contributing
