+++ SeismicMesh/geometry/__init__.py
@@ -17,6 +17,7 @@ from .signed_distance_functions import (
     Union,
     Intersection,
     Difference,
+    Repeat,
 )
 from .utils import (
     calc_re_ratios,
@@ -78,4 +79,5 @@ __all__ = [
     "Union",
     "Intersection",
     "Difference",
+    "Repeat",
 ]
