+++ SeismicMesh/geometry/__init__.py
@@ -11,6 +11,7 @@ from .signed_distance_functions import (
     Cube,
     Disk,
     Ball,
+    Torus,
     drectangle,
     dblock,
     corners,
@@ -51,6 +52,7 @@ __all__ = [
     "Rectangle",
     "Ball",
     "Cube",
+    "Torus",
     "Disk",
     "drectangle",
     "dblock",
