+++ SeismicMesh/geometry/__init__.py
@@ -12,6 +12,7 @@ from .signed_distance_functions import (
     Disk,
     Ball,
     Torus,
+    Prism,
     drectangle,
     dblock,
     corners,
@@ -53,6 +54,7 @@ __all__ = [
     "Ball",
     "Cube",
     "Torus",
+    "Prism",
     "Disk",
     "drectangle",
     "dblock",
