+++ SeismicMesh/__init__.py
@@ -11,6 +11,7 @@ from .geometry import (
     Ball,
     Disk,
     Cube,
+    Cylinder,
     Rectangle,
     Union,
     Intersection,
@@ -32,6 +33,7 @@ __all__ = [
     "geometry",
     "Rectangle",
     "Cube",
+    "Cylinder",
     "Disk",
     "Union",
     "Torus",
