### 2020-10-24 3d3d60c
remove gigantic torus
