### 2020-10-25 74e04d9
Update signed_distance_functions.py
