+++ SeismicMesh/geometry/signed_distance_functions.py
@@ -13,12 +13,12 @@ def corners(bbox):
     return list(itertools.product(*zip(mins, maxs)))
 
 
- def _gather_corners(domains):
-     corners = [d.corners for d in domains if d.corners is not None]
-     if len(corners) == 0:
-         return None
-     else:
-         return np.concatenate(corners)
+def _gather_corners(domains):
+    corners = [d.corners for d in domains if d.corners is not None]
+    if len(corners) == 0:
+        return None
+    else:
+        return np.concatenate(corners)
  
 class Union:
     def __init__(self, domains):
