### 2020-10-25 7d4d6d6
only constrain points near level set
