### 2020-10-25 db5f7db
corners need to be np.array
