### 2020-10-25 63b0d7c
Add more primitives (#112)
*New Torus /Cylinder/Wedge geometric primitives 
* Corners are only constrained near the 0-level set.
