+++ SeismicMesh/geometry/__init__.py
@@ -9,8 +9,11 @@ from .cpp.fast_geometry import (
 from .signed_distance_functions import (
     Rectangle,
     Cube,
+    Cylinder,
     Disk,
     Ball,
+    Torus,
+    Prism,
     drectangle,
     dblock,
     corners,
@@ -51,6 +54,9 @@ __all__ = [
     "Rectangle",
     "Ball",
     "Cube",
+    "Cylinder",
+    "Torus",
+    "Prism",
     "Disk",
     "drectangle",
     "dblock",
