+++ SeismicMesh/__init__.py
@@ -7,7 +7,18 @@
 #  see <http://www.gnu.org/licenses/>.
 
 from . import decomp, geometry, migration
-from .geometry import Ball, Disk, Cube, Rectangle, Union, Intersection, Difference
+from .geometry import (
+    Ball,
+    Disk,
+    Cube,
+    Cylinder,
+    Rectangle,
+    Union,
+    Intersection,
+    Difference,
+    Torus,
+    Prism,
+)
 from .generation import generate_mesh, sliver_removal
 from .sizing import (
     get_sizing_function_from_segy,
@@ -22,8 +33,11 @@ __all__ = [
     "geometry",
     "Rectangle",
     "Cube",
+    "Cylinder",
     "Disk",
     "Union",
+    "Torus",
+    "Prism",
     "Ball",
     "Intersection",
     "Difference",
