+++ Makefile
@@ -25,9 +25,8 @@ clean:
 	@rm -rf dist/
 
 format:
-	isort -rc SeimicMesh/ examples/*.py tests/*.py
-	black SeismicMesh/ examples/*.py tests/*.py
-	blacken-docs README.md
+	isort -rc SeimicMesh/ tests/*.py
+	black SeismicMesh/ benchmarks/*.py tests/*.py
 	clang-format -i SeismicMesh/generation/cpp/*.cpp SeismicMesh/migration/cpp/*.cpp SeismicMesh/sizing/cpp/*.cpp SeismicMesh/geometry/cpp/*.cpp
 
 black:
