### 2020-10-27 #2687 @nschloe
Okay, I'm going over the text now. One thing I noticed is the following
- [ ] "The usage of 11 cores reduces the generation time of this example from 20 minutes to approximately 2 wall-clock minutes". According to the graph, it's more "from 6 minutes". I would also like to see a dashed line in the graph for perfect speed up (or equal scaling in the axes.)
- [ ] Also a question: Why don't we have perfect speed-up?
