### 2020-10-27 #2687 @krober10nd
Hey @nschloe

 For the future, I think we should discuss this in the repository with an issue but I will answer this here briefly. We can start an issue and discuss in the repo.

> "The usage of 11 cores reduces the generation time of this example from 20 minutes to approximately 2 wall-clock minutes". According to the graph, it's more "from 6 minutes". I would also like to see a dashed line in the graph for perfect speed up (or equal scaling in the axes.)

Sure. I can fix this.

> Also a question: Why don't we have perfect speed-up?

There are two reasons: 1) the domain decomposition strategy that I use does not consider load balancing (it simply slices one axis and does not consider the spatial distribution of points) 2) the communication cost becomes competitive to the parallelization speed-up of the Delaunay triangulation.
