### 2020-10-28 krober10nd/SeismicMesh#117 @nschloe
question about speed-up
From https://github.com/openjournals/joss-reviews/issues/2687#issuecomment-717389718:

> 2) the communication cost becomes competitive to the parallelization speed-up of the Delaunay triangulation.

What is the main bottleneck in the algorithm? Perhaps axpy or cell manipulations?
