### 2020-10-28 krober10nd/SeismicMesh#117 @krober10nd
I am re-running some timing results on our cluster, but while those run we can discuss here.

The primary bottleneck in the parallel DistMesh algorithm becomes 1) determining which points need to be migrated to neighboring domains (enqueue) and 2) then migrating points to neighboring ranks (which appears to be quicker but I'm working on those numbers). Both steps are performed each meshing iteration as points move around and I don't do any boundary point locking business.

https://github.com/krober10nd/SeismicMesh/blob/733d6b0a2f1ff99145dbf7630b57f377583b9b26/SeismicMesh/generation/mesh_generator.py#L656

Step 1) is the most costly and requires basically requires finding the intersection between two boxes (rank-1 and rank+1 for ranks > 0 and ranks < n-1 where n is the size of the communicator) and a  bunch of spheres; in particular a vertex is migrated to a neighboring domain if at least one of the circumcircles out of all the vertices' connected  triangles intersect with the extents of a neighboring rank. This ensures the triangulation is Delaunay because it satisfies the empty-circumcircle property and more details can be seen in the original parallel Delauanay algorithm paper, which I adapted for this...

https://mrzv.org/publications/distributed-delaunay/

Note: the original paper simply does parallel Delaunay and has no concerns about quality.

The relevant box intersects a sphere code is here (above that is the version in 2D which is way faster):

https://github.com/krober10nd/SeismicMesh/blob/733d6b0a2f1ff99145dbf7630b57f377583b9b26/SeismicMesh/migration/cpp/cpputils.cpp#L274

For this task I use CGAL's primitives however I haven't been able to time this yet perhaps there's some lingering bottlenecks.
