### 2020-10-28 krober10nd/SeismicMesh#117 @krober10nd
In serial, the primary bottleneck is actually getting the edges of the triangulation which I carefully wrote in CPP as a result.

https://github.com/krober10nd/SeismicMesh/blob/733d6b0a2f1ff99145dbf7630b57f377583b9b26/SeismicMesh/generation/mesh_generator.py#L621
