### 2020-10-28 krober10nd/SeismicMesh#117 @krober10nd
So here's a MWE to quantity things:
```
 from mpi4py import MPI
 import SeismicMesh
 import meshio

 comm = MPI.COMM_WORLD
 bbox = (-10.0, 10.0, -10.0, 10.0, -1.0, 1.0)
 hmin = 0.25
 cube = SeismicMesh.Cube(bbox)
 ball = SeismicMesh.Ball([0.0, 0.0, 0.0], 0.5)
 diff = SeismicMesh.Difference([cube, ball])
 points, cells = SeismicMesh.generate_mesh(
     domain=diff,
     edge_length=hmin,
     verbose=2,
     max_iter=25,
     axis=1,
 )
 points, cells = SeismicMesh.sliver_removal(points=points, domain=diff, edge_length=hmin)
 if comm.rank == 0:
     meshio.write_points_cells(
         "cube.vtk",
         points,
         [("tetra", cells)],
         file_format="vtk",
     )
```

Using `pyinstrument`  mpiexec -n 2 pyinstrument c.py

produces this

```
54.211 <module>  c.py:1
├─ 31.164 generate_mesh  SeismicMesh/generation/mesh_generator.py:283
│  ├─ 12.372 _add_ghost_vertices  SeismicMesh/generation/mesh_generator.py:666
│  │  ├─ 5.484 remove_external_entities  SeismicMesh/geometry/utils.py:52
│  │  │  ├─ 3.741 fix_mesh  SeismicMesh/geometry/utils.py:201
│  │  │  │  ├─ 2.189 unique_rows  SeismicMesh/geometry/utils.py:130
│  │  │  │  │  └─ 2.184 unique  <__array_function__ internals>:2
│  │  │  │  │     └─ 2.182 unique  numpy/lib/arraysetops.py:149
│  │  │  │  │           [4 frames hidden]  numpy
│  │  │  │  │              2.142 _unique1d  numpy/lib/arraysetops.py:309
│  │  │  │  │              └─ 1.633 ndarray.argsort  <built-in>:0
│  │  │  │  └─ 0.954 simp_vol  SeismicMesh/geometry/utils.py:172
│  │  │  │     └─ 0.792 [self]
│  │  │  ├─ 0.823 dblock  SeismicMesh/geometry/signed_distance_functions.py:263
│  │  │  └─ 0.805 [self]
│  │  ├─ 5.355 enqueue  SeismicMesh/migration/migration.py:114
│  │  │  ├─ 2.876 [self]
│  │  │  └─ 2.465 vertex_to_entities  SeismicMesh/geometry/utils.py:96
│  │  │     ├─ 1.299 ndarray.argsort  <built-in>:0
│  │  │     └─ 0.617 [self]
│  │  └─ 1.010 [self]
│  ├─ 7.918 _compute_forces  SeismicMesh/generation/mesh_generator.py:642
│  │  ├─ 3.181 _get_edges  SeismicMesh/generation/mesh_generator.py:631
│  │  │  └─ 3.020 [self]
│  │  ├─ 1.878 __call__  scipy/interpolate/interpolate.py:2450
│  │  │     [8 frames hidden]  scipy
│  │  ├─ 1.299 [self]
│  │  └─ 0.851 _sum  numpy/core/_methods.py:45
│  │     └─ 0.851 ufunc.reduce  <built-in>:0
│  ├─ 5.809 [self]
│  ├─ 2.176 _remove_triangles_outside  SeismicMesh/generation/mesh_generator.py:685
│  │  ├─ 1.067 _sum  numpy/core/_methods.py:45
│  │  │  └─ 1.067 ufunc.reduce  <built-in>:0
│  │  └─ 0.851 [self]
│  ├─ 1.685 _improve_level_set_newton  SeismicMesh/generation/mesh_generator.py:692
│  │  └─ 1.643 get_boundary_vertices  SeismicMesh/geometry/utils.py:347
│  │     └─ 1.635 get_boundary_facets  SeismicMesh/geometry/utils.py:404
│  │        └─ 1.276 unique  <__array_function__ internals>:2
│  │           └─ 1.276 unique  numpy/lib/arraysetops.py:149
│  │                 [4 frames hidden]  numpy
│  │                    1.273 _unique1d  numpy/lib/arraysetops.py:309
│  │                    └─ 1.248 ndarray.sort  <built-in>:0
│  └─ 0.612 _termination  SeismicMesh/generation/mesh_generator.py:617
├─ 21.872 sliver_removal  SeismicMesh/generation/mesh_generator.py:58
│  ├─ 17.695 [self]
│  ├─ 1.797 _calc_dihedral_angles  SeismicMesh/generation/mesh_generator.py:505
│  │  └─ 1.674 [self]
│  └─ 1.775 _remove_triangles_outside  SeismicMesh/generation/mesh_generator.py:685
│     ├─ 0.861 _sum  numpy/core/_methods.py:45
│     │  └─ 0.861 ufunc.reduce  <built-in>:0
│     └─ 0.699 [self]
└─ 0.891 <module>  SeismicMesh/__init__.py:9
   └─ 0.714 <module>  SeismicMesh/geometry/__init__.py:1
      └─ 0.710 <module>  SeismicMesh/geometry/utils.py:1
         └─ 0.650 <module>  SeismicMesh/generation/__init__.py:1
            └─ 0.644 <module>  SeismicMesh/generation/mesh_generator.py:16
               └─ 0.551 <module>  SeismicMesh/sizing/__init__.py:1
                  └─ 0.550 <module>  SeismicMesh/sizing/mesh_size_function.py:15

To view this report with different options, run:
    pyinstrument --load-prev 2020-10-28T09-40-55 [options]
```
