### 2020-10-28 5193b47
3d binary read bug fix
