### 2020-10-28 8bd4a5a
3d binary read bug fix (#118)
* 3d binary read bug fix
