### 2020-10-28 krober10nd/SeismicMesh#117 @krober10nd
I ran the following script three times for 1 to 11 core configurations and I produced the average timings displayed in this figure below:

```
import time
from mpi4py import MPI
import zipfile
import meshio

from SeismicMesh import (
    get_sizing_function_from_segy,
    generate_mesh,
    sliver_removal,
    Cube,
)

comm = MPI.COMM_WORLD

bbox = (-4200.0, 0.0, 0.0, 13520.0, 0.0, 13520.0)

hmin = 75.0

path = "Salt_Model_3D/3-D_Salt_Model/VEL_GRIDS/"
if comm.rank == 0:
    # Extract binary file Saltf@@ from SALTF.ZIP
    zipfile.ZipFile(path + "SALTF.ZIP", "r").extract("Saltf@@", path=path)

fname = path + "Saltf@@"

nx, ny, nz = 676, 676, 210

cube = Cube(bbox)

ef = get_sizing_function_from_segy(
    fname,
    bbox,
    hmin=hmin,
    dt=0.001,
    freq=2,
    wl=10,
    grade=0.15,
    hmax=5e3,
    nz=nz,
    nx=nx,
    ny=ny,
    byte_order="big",
    axes_order=(2, 0, 1),  # order for EAGE (x, y, z) to default order (z,x,y)
    axes_order_sort="F", # binary is packed in a FORTRAN-style
)

t1 = time.time()
points, cells = generate_mesh(domain=cube, h0=hmin, edge_length=ef, verbose=1)
print("Mesh generation time was : "+str(time.time()-t1))

points, cells = sliver_removal(
    points=points, bbox=bbox, h0=hmin, domain=cube, edge_length=ef
)
print(time.time()-t1)

if comm.rank == 0:

    meshio.write_points_cells(
        "EAGE_Salt.vtk",
        points[:, [1, 2, 0]] / 1000.0,
        [("tetra", cells)],
    )
```

![Performance](https://user-images.githubusercontent.com/18619644/97472600-c1752c00-1928-11eb-8448-372d57f3f547.png)
