### 2020-10-29 krober10nd/SeismicMesh#117 @krober10nd
I'm repeating the above for a significantly larger (more cells) version of this example (approximately 35 million cells).
