+++ paper/paper.md
@@ -62,7 +62,7 @@ The 2D/3D serial performance against `gmsh` and `cgal` [@cgal:rty-m3-20b, @cgal:
 
 A simplified version of the parallel Delaunay algorithm proposed by [ @peterka2014high] is implemented inside the DistMesh algorithm. This algorithm does not consider sophisticated domain decomposition or load balancing. \autoref{fig:speedup} shows a peak speed-up of approximately 4 times using 11 cores when performing 50 meshing iterations to generate both the light and heavy meshes of the EAGE P-wave velocity model. While the parallel performance is not ideal, this capability enables the generation of large meshes with 33 M cells that would otherwise take 92 minutes to generate down to approximately 21.7 minutes. The machine used for this experiment was an Intel Xeon Gold 6148 machine clocked at 2.4 GHz  with 192 GB of RAM connected together with a 100 Gb/s InfiniBand network.
 
-![The speedup (left-panel) as compared to the serial version of SeismicMesh V3.1.0 for a relatively light and heavy meshes adapted to P-wave data from the EAGE Salt seismic velocity model. The total mesh generation wall-clock time is annotated in decimal minutes next to each point. The panel on the right hand side shows the mesh generation rate normalized by the number of total cells in the mesh. \label{fig:speedup}](Performance.png)
+![The speedup (left-panel) as compared to the serial version of SeismicMesh V3.1.0 for a relatively light and heavy mesh each adapted to P-wave data from the EAGE Salt seismic velocity model. The total mesh generation wall-clock time is annotated in decimal minutes next to each point. The panel on the right hand side shows the mesh generation rate normalized by the number of total number of cells in the mesh. \label{fig:speedup}](Performance.png)
 
 
 # Ongoing and future applications
