### 2020-10-29 bd013e3
Parallel performance figure (#120)
* Updating JOSS paper with a hopefully better analysis of parallel performance at scale
