### 2020-10-30 5937a59
speeding up boundary vertices calc
