### 2020-10-31 4031a43
speeding up boundary vertices calc (#121)
* speeding up boundary vertices calc

* flake8 fix

* version bump
