+++ README.md
@@ -486,7 +486,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project (tries to) adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- None
+### Improved
+- Faster calculation of boundary vertices. 
+
 ## [3.1.0] - 2020-10-28
 ### Added
 - New geometric primitives--torus, wedge/prism, and cylinder.
