### 2020-11-02 #2687 @krober10nd
Just an updated: I remade the parallel figure, added it to the paper, and updated the text.

Python axpy in parallel indeed gets in the way. Something always to be improved.
