### 2020-11-03 4c86ade
removing boundary projection step in sliver_removal
