### 2020-11-03 50d1c02
getting rid of ad-hoc max_iter
