### 2020-11-03 eea25f9
Update setup.cfg
