### 2020-11-03 07c9aab
improving 3d robustness (#124)
* removing boundary projection step in sliver_removal

* getting rid of ad-hoc max_iter

* Update setup.cfg
