### 2020-11-04 c510fb5
increasing assert tolerance
