### 2020-11-04 a3af79a
critical bug fix to corners (#126)
* critical bug fix to corners

* increasing assert tolerance
