### 2020-11-04 0fa1e5e
removing plot command
