### 2020-11-04 73707bc
more control over grad option (#127)
* more control over grad/velocity transition option

* removing plot command

* testing new stencil_size
