+++ README.md
@@ -486,7 +486,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- None
+### Improved
+- More control over the `grad` option in the mesh sizing function.
 
 ## [3.1.2] - 2020-11-04
 ### Improved
