### 2020-11-05 #2687 @jorgensd
@krober10nd It seems to me that the cgal reference in the Performance Comparison chapter is broken. In the article proof above it shows up as:
The 2D/3D serial performance against gmsh and cgal (Alliez et al., 2020, pp. @cgal:r–ctm2–
20b)
