### 2020-11-05 df984ef
small paper edits
