### 2020-11-05 #2687 @krober10nd
Okay I put in a fix for this citation and removed the RTM acronym  @jorgensd .  I'll update the benchmark figure with hopefully bigger labels and the clarification about the 2D performance of SM.
