### 2020-11-05 9e3bd6d
updating paper and benchmarks
