+++ paper/paper.md
@@ -55,7 +55,7 @@ The mesh's domain geometry is defined as the 0-level set of a signed distance fu
 
 The 2D/3D serial performance against `gmsh` and `cgal` [@cgal:rty-m3-20b; @cgal:r-ctm2-20b] in terms of cell quality and creation time where cell quality is defined as dimension (2 or 3) multiplied by the circumcircle radius divided by the incircle radius. This cell quality is between 0 and 1, where 1 is a perfectly symmetrical simplex. For the analytical geometries (e.g. disk and ball) `gmsh` is the fastest to generate a mesh and then performance is approximately similar for both `SeismicMesh` and `cgal` with `cgal` outperforming `SeismicMesh` for the disk and vice versa for the ball. `gmsh` and `cgal` produce higher minimum cell qualities overall than `SeismicMesh` but importantly all generators are capable of producing sliver-free meshes. With that said, `SeismicMesh` produces higher mean cell qualities by about 5-10\% as compared to `gmsh` and `cgal`, which leads to reduced finite element matrix condition numbers. For the two seismic domains (e.g., BP2004 and EAGE), `SeismicMesh` is faster than `gmsh` for the 2D BP2004 benchmark but slightly slower for the 3D EAGE benchmark. Mesh quality results similarly follows the results observed in the two analytical cases. Interpolant-based mesh sizing functions significantly slow the mesh generation time of `gmsh` increasing its mesh generation time by a factor of 3x. The `DistMesh` algorithm used in `SeismicMesh` requires far less calls to the sizing function. For example in the EAGE benchmark, `gmsh` calls the sizing function 95,756 times whereas `DistMesh` calls it just 26 times.
 
-![The mesh creation time (left columns) and resulting cell quality for the four benchmark problems studied over a range of problem sizes. Two analytical problems (disk and a ball) and two non-analytical problems with sizing functions defined via regular gridded interpolants (BP2004 and EAGE). For the panels that show cell quality, solid lines indicate the mean and dashed lines indicate the minimum cell quality in the mesh. See the `SeismicMesh` github repository for information regarding the benchmarks and version of programs used. \label{fig:benchmark}](Benchmarks.png)
+![The mesh creation time (left columns) and resulting cell quality (right columns) for the four benchmark problems studied over a range of problem sizes. Two analytical problems (disk and a ball) and two non-analytical problems with sizing functions defined via regular gridded interpolants (BP2004 and EAGE). For the panels that show cell quality, solid lines indicate the mean and dashed lines indicate the minimum cell quality in the mesh. See the `SeismicMesh` github repository for information regarding the benchmarks and version of programs used. \label{fig:benchmark}](Benchmarks.png)
 
 
 # Parallelism
