### 2020-11-05 4f7eba5
updating paper and benchmarks (#130)
* updating paper text and benchmark figure, and figure creation scripts.
