+++ benchmarks/run_ball.py
@@ -40,10 +40,12 @@ plt.title("Number of cells vs. mesh generation time")
 plt.legend()
 plt.xlabel("Number of cells")
 plt.ylabel("Elapsed time (s)")
+plt.ticklabel_format(style="sci", axis="x", scilimits=(0, 0))
 plt.grid()
 
 plt.subplot(1, 2, 2)
 plt.title("Number of cells. vs cell quality")
+plt.ticklabel_format(style="sci", axis="x", scilimits=(0, 0))
 plt.xlabel("Number of cells")
 plt.ylabel("Cell quality")
 plt.grid()
