+++ README.md
@@ -483,6 +483,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Cylinder radius and height are now correct.
+
 ### Improved
 - More control over the `grad` option in the mesh sizing function.
 
