### 2020-11-06 543ed14
Update setup.cfg
