+++ README.md
@@ -482,9 +482,13 @@ Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## Unreleased 
+- None
+
+## [3.1.3] - 2020-11-06
 ### Fixed
 - Cylinder radius and height are now correct.
+- Torus, Prism, and Cylinder now have `dim` tag.
 
 ### Improved
 - More control over the `grad` option in the mesh sizing function.
