### 2020-11-06 #2687 @jorgensd
@krober10nd  As whedon is suggesting, you are missing the DOI of the Tournois paper in the references. The DOI is:  10.1007/978-3-642-04319-2_10
