+++ paper/paper.bib
@@ -238,6 +238,7 @@ eprint = {https://doi.org/10.1190/1.1441754}
   pages={157--173},
   year={2009},
   publisher={Springer}
+  doi={10.1007/978-3-642-04319-2_10}, 
 }
 
 @article{doi:10.1190/1.1437283,
