+++ paper/paper.bib
@@ -237,7 +237,7 @@ eprint = {https://doi.org/10.1190/1.1441754}
   booktitle={Proceedings of the 18th international meshing roundtable},
   pages={157--173},
   year={2009},
-  publisher={Springer}
+  publisher={Springer},
   doi={10.1007/978-3-642-04319-2_10}, 
 }
 
