### 2020-11-10 krober10nd/SeismicMesh#117 @krober10nd
@nschloe do you have any thoughts/ways to improve this or do you want to close this?
