### 2020-11-10 a922f2c
improve cell manipulation
