### 2020-11-11 264f89a
kicking triangles out
