### 2020-11-13 #2687 @krober10nd
@nschloe @jorgensd Can you please elaborate more on what ongoing work needs to be done for to finish this review?

Just to state some recent activity: I've fixed a broken reference in the paper and edited a statement regarding mesh quality on KSP performance. I've improved both the serial and parallel performance comparisons and at least the serial performance comparisons have been corroborated via Nico's meshgen comparison repository.

Other than that, I've also improved the README to follow the guidelines of the JOSS review by adding a contributing section and clarified the statement of need.

Thank you
