### 2020-11-13 krober10nd/SeismicMesh#117 @nschloe
Looks good now.
