### 2020-11-13 krober10nd/SeismicMesh#139 @nschloe
more hints for the article
Important:
* You speak of "modifications of distmesh". What modifications are these, exactly? Please go into more detail in the article. (If the mods aren't too small.) If the mods are small, don't mention them.
* I think you need to make clearer that SeismicMesh is a parellel distmesh implementation plus a mesh density function generator. I would like to hear _why_ distmesh_ and not some other algorithm? Easy to implement, known to produce high-quality cells, easily paralellizable, but can be slow.
* Why don't you simply use Gmsh, right? Aha, you find that Gmsh is slow for your use case! Why? Because your density function is so complicated and it heavily benefits from vectorization. Gmsh's algorithm doesn't allow for vectorization, distmesh does. ==> For complicated density functions, distmesh is better.

Less important:
* You're using Gmsh and CGAL (perhaps others) in texttt as if it was on a command line. The names are Gmsh and CGAL, really, no particular style needed.
