### 2020-11-13 krober10nd/SeismicMesh#139 @krober10nd
Okay lets break these comments down:

> You speak of "modifications of distmesh". What modifications are these, exactly? Please go into more detail in the article. (If the mods aren't too small.) If the mods are small, don't mention them.

Okay fair enough. I'll remove "modifications". Although I meant that the parallelism was never really an option in the `DistMesh` so it's a modification to the algorithm to support parallelism.

> I think you need to make clearer that SeismicMesh is a parellel distmesh implementation plus a mesh density function generator. I would like to hear why distmesh_ and not some other algorithm?

Proposed new summary:

# Summary

`SeismicMesh` is a Python package containing a parallel triangular mesh generator and mesh sizing function creator for  two and three dimensional domains such as those encountered in seismology.  C++ code is wrapped with Python for performance without losing ease-of-use. Mesh generation is performed in either serial or using distributed memory parallelism with an implementation of `DistMesh` [@doi:10.1137/S0036144503429121; @peterka2014high]. `DistMesh`  is an uncomplicated algorithm that is known to produce high-quality cells, can be easily parallelized, and does not require explicit domain geometry information, which can be useful for people unfamiliar with mesh generation. Our implementation uses the Computational Geometry Algorithms Library [@cgal:hs-chdt3-20a] to perform many geometric operations quickly.

I'm thinking about how to revise the `Statement of Need` to incorporate your other suggestions.
