### 2020-11-13 krober10nd/SeismicMesh#139 @nschloe
I tried to make it a bit more crisp (feel free to use or dismiss):

SeismicMesh is a Python package for simplex mesh generation in two or three dimensions. As an implementation of DistMesh [@doi:10.1137/S0036144503429121; @peterka2014high], it produces high-quality meshes at the expense of speed. For increased efficiency, the core package is written in C++, works in parallel, and uses the Computational Geometry Algorithms Library [@CGAL:hs-chdt3-20a]. SeismicMesh can also produce mesh-density functions from seismological data to be used in the mesh generator.
