### 2020-11-13 krober10nd/SeismicMesh#139 @krober10nd
That's excellent, thank you Nico.
