### 2020-11-13 c173bf2
improvements to summary from review
