### 2020-11-13 krober10nd/SeismicMesh#139 @krober10nd
This paragraph is a little more challenging for me, but I think it's better below. I think the point you make is quite interesting about querying sizing functions.

 Besides that it would be nice to perhaps highlight the implicit functions that I use make the geometry creation a lot easier than hand drawing an STL. I haven't yet determined where best to put that or if I will put that in.

> Why don't you simply use Gmsh, right? Aha, you find that Gmsh is slow for your use case! Why? Because your density function is so complicated and it heavily benefits from vectorization. Gmsh's algorithm doesn't allow for vectorization, distmesh does. ==> For complicated density functions, distmesh is better.

The success of many finite element methods in geophysics hinges on the creation of a quality mesh. Despite the fact that many mesh generation programs exist such as Gmsh and CGAL, it is rare to find the capabilities to incorporate geophysical data directly into the mesh generation process in order to size elements and thus automatic mesh generation remains an unsolved problem.

Recently, some packages have been created to this end to script mesh generation from geophysical datasets such as in coastal ocean modeling [e.g., @roberts2019oceanmesh2d; GMT and Terreno @gorman2008systematic] and reservoir modeling [e.g., MeshIT @cacace2015meshit]. Following in a similar direction to those programs, the aim of this package is to provide a simple-to-use Python package to script mesh generation from geophysical datasets by building complex mesh density function directly from seismic velocity models.

Our mesh density functions can be used with other mesh generators, however their usage can have significant implications on mesh generation performance. For example, Gmsh's advancing front and Delaunay refinement methods construct the mesh incrementally and do not permit vectorization leading to significant slow downs at scale in 2D/3D. In contrast, the DistMesh algorithm takes advantage of vectorization when querying the mesh density function and makes it a good choice for this kind of meshing problem.
