### 2020-11-14 1c8ff1e
Update setup.cfg
to incorporate extras for benchmarking
