+++ setup.cfg
@@ -34,4 +34,11 @@ install_requires =
         matplotlib
         mpi4py
         pybind11 >= 2.2
+extras_requires = 
+        meshplex==0.13.3
+        pygalmesh==0.8.2
+        pygmsh==7.0.0
+        termplotlib==0.3.2
+        meshio==4.2.2
+        termplotlib==0.3.2
 python_requires = >=3.0
