### 2020-11-14 0be1970
Update setup.cfg
syntax to setup.cfg
