+++ setup.cfg
@@ -34,7 +34,8 @@ install_requires =
         matplotlib
         mpi4py
         pybind11 >= 2.2
-extras_requires = 
+[options.extras_require]
+benchmarks = 
         meshplex==0.13.3
         pygalmesh==0.8.2
         pygmsh==7.0.0
