### 2020-11-14 #2687 @jorgensd
> @nschloe @jorgensd Can you please elaborate more on what ongoing work needs to be done for to finish this review?
>
> Just to state some recent activity: I've fixed a broken reference in the paper and edited a statement regarding mesh quality on KSP performance. I've improved both the serial and parallel performance comparisons and at least the serial performance comparisons have been corroborated via Nico's meshgen comparison repository.
>
> Other than that, I've also improved the README to follow the guidelines of the JOSS review by adding a contributing section and clarified the statement of need.
>
> Thank you

I think we are getting quite close to a finalized version. I've added some more notes to the various issues in SeismicMesh (and some of them you have already addressed). When we iron out the last few things, I am happy with the state of the software. The final thing would be to have a last read-through of the paper. Nico has given you some good input, and I believe that I will have very few corrections after the thread https://github.com/krober10nd/SeismicMesh/issues/139 is finalized.
