### 2020-11-14 60a29d8
Update setup.cfg
