+++ setup.py
@@ -16,6 +16,14 @@ if sys.version_info < (3, 0):
     print("Python 3.0 or higher required, please upgrade.")
     sys.exit(1)
 
+benchmarking = [
+     "meshplex == 0.13.3",
+     "pygalmesh == 0.8.2",
+     "pygmsh == 7.0.0",
+     "termplotlib == 0.3.2",
+     "meshio == 4.2.2",
+     "termplotlib == 0.3.2",
+ ]
 
 class CMakeExtension(Extension):
     def __init__(self, name, sourcedir=""):
@@ -88,5 +96,8 @@ setup(
         CMakeExtension("SeismicMesh/geometry/cpp/fast_geometry"),
     ],
     cmdclass=dict(build_ext=CMakeBuild),
+    extras_require={
+         "benchmarking": benchmarking,
+     },
     python_requires=">=3.0",
 )
