### 2020-11-16 krober10nd/SeismicMesh#139 @krober10nd
Okay, I did several iterations of the above and liked how it flowed. I also rewrote the performance section to rewrite the improved 2D minimum cell qualities.
