### 2020-11-18 e11f181
Update paper.bib
adding another CGAL reference #139
