### 2020-11-18 617ace5
Update paper.bib (#148)
adding another CGAL reference #139
