### 2020-11-18 #2687 @krober10nd
I fixed the CGAL reference above. Now there's a typsetting issue remaining.

  I made the benchmark figure quite long to deal with the fact that the JOSS article format has large margins but now the caption of that figure is all jumbled up.

I guess the solution is to make the figure wider now.
