### 2020-11-19 krober10nd/SeismicMesh#139 @krober10nd
Do you have any more comments on this article @nschloe ? The latest paper is on the main branch.
