### 2020-11-30 krober10nd/SeismicMesh#139 @krober10nd
@nschloe Can we close this issue?
