### 2020-11-30 krober10nd/SeismicMesh#139 @nschloe
I will check as soon as I have time.
