### 2020-12-08 krober10nd/SeismicMesh#139 @nschloe
Okay, closing.
