### 2020-12-16 #2687 @krober10nd
I've dealt with one more issue pertaining to the paper (https://github.com/krober10nd/SeismicMesh/pull/170) @nschloe brought up and I've independently improved parallel performance in 2d/3d by a significant margin in https://github.com/krober10nd/SeismicMesh/pull/171 and  https://github.com/krober10nd/SeismicMesh/pull/172 and updated the figure and text accordingly regarding 3d parallel performance.

I'd like to know please where this paper stands.

Thanks
