### 2020-12-18 #2687 @nschloe
This is good to go from my point of view.

Just for the record, let me add a few remarks.

 * @krober10nd was exceptionally responsive to the remarks, and I think the software and the article have greatly improved in the process.
 * I'm not a big fan of how the software targets seismologists only. SeismicMesh is a parallel implementation of distmesh, and as such is general-purpose mesh generator. There is a function that converts seismology data into a density function, but that part is actually separate and can be used with any other mesh generator. I also wouldn't like to see another mesh generator next year that only adapts that small part for medicine and calls it _CellMesh_.
 * For anyone interested, as part of the review process, I compiled a more extensive mesh generator comparison at https://github.com/nschloe/meshgen-comparison (including SeismicMesh).
 * I love the JOSS review process and I'll hopefully start writing up some articles for it myself next year.

Happy holidays everyone!
