### 2020-12-21 #2687 @krober10nd
just running Nico's `betterbib` package to add as much information to the papers bibliography as possible.
