### 2021-01-18 #2687 @meg-simula
@krober10nd Happy New Year! Thanks for your patience and nudging here.

In the paper, could you please address the following minor comment:

- [ ] Core functionality - item 2. This sentence is unclear (something about the "in either" bit), please improve.
