### 2021-01-18 #2687 @krober10nd
Hey @meg-simula ...Okay, I edited the text in the paper in the Core functionality- item 2 (it was a typo).

I then made a tagged release V3.3.1 https://github.com/krober10nd/SeismicMesh/releases/tag/V3.3.1 and archived it in Zenodo and edited the metadata on Zenodo to reflect and match the paper. https://zenodo.org/record/4447042#.YAWQvS1h1pQ

The doi is 10.5281/zenodo.4447042
