### 2021-01-18 #2687 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
