### 2021-01-19 #2687 @kthyng
@krober10nd Hi! I'll be taking over from here. I see your Zenodo archive is set up, and the version of your software is set.

Paper comments:
* Your #1 under Core Functionality — the last reference should not have ().
* your #3 — I'm outside this field but in ocean meshes, 2d meshes can have bad quality shape (e.g. really narrow triangles). Could you state briefly why 2d meshes don't have degenerate elements?
* Parallelism first sentence: no () around reference.
