### 2021-01-19 #2687 @krober10nd
@kthyng Thanks! I fixed up the two citations to be in-text rather than parenthetical. In regard to your comment about degenerate element quality, indeed I meant so-called sliver elements which occur only in 3D meshes. I edited the text minimally to be more specific what I meant by degenerate elements in this case.
